### 2022-05-13 #4392 @editorialbot
[PRE REVIEW]: ParMOO: A Python library for parallel multiobjective simulation optimization
**Submitting author:** @thchang (<a href="http://orcid.org/0000-0001-9541-7041">Tyler Chang</a>)
**Repository:** https://github.com/parmoo/parmoo
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.0
