### 2022-05-17 #4392 @kellyrowland
Yeah, the JOSS setup is pretty slick. 🙂

I would (or will) also be looking at the spreadsheet for potential reviewers, so if you're able to scroll through and look for folks with relevant expertise, I'd gladly take some suggestions.
