### 2022-06-10 #4468 @editorialbot
[REVIEW]: ParMOO: A Python library for parallel multiobjective simulation optimization
**Submitting author:** @thchang (<a href="http://orcid.org/0000-0001-9541-7041">Tyler Chang</a>)
**Repository:** https://github.com/parmoo/parmoo
**Branch with paper.md** (empty if default branch): joss-rev
**Version:** 0.2.0
