### 2022-06-23 67e3707
add viz folder + print functions
