### 2022-06-29 a0c8aed
docstrings + flake8 compliance
