+++ parmoo/viz/plots.py
@@ -1,6 +1,42 @@
-import plotly.express as px
-import plotly.graph_objects as go
-from parmoo import MOOP
+""" This module contains a library of common plotting functions.
+
+The functions are:
+
+ Print raw data to terminal:
+  * ``printDataTypes(moop)`` -- print data type
+  * ``printPF_raw(moop)`` -- print Pareto Front
+  * ``printObjectiveData_raw(moop)`` -- print objective database
+  * ``printSimulationData_raw(moop)`` -- print simulation database(s)
+  * ``printDesignVariables_raw(moop)`` -- print design variable data
+  * ``printConstraints_raw(moop)`` -- print constraint data
+  * ``printMOOP_raw(moop)`` -- print all MOOP data
+
+ Print data to terminal as table
+  * ``printPF(moop)`` -- print Pareto Front
+  * ``printObjectiveData(moop)`` -- print objective database
+  * ``printSimulationData(moop)`` -- print simulation database(s)
+  * ``printMOOP(moop)`` -- print all MOOP data
+
+ Display data in interactive browser plot
+  * ``scatter(moop)`` -- Plot MOOP results as matrix of 2D scatterplots
+  * ``scatter3d(moop)`` -- Plot MOOP results as matrix of 3D scatterplots
+  * ``radar(moop)`` -- Plot MOOP results as radar
+  * ``parallel_coordinates(moop)`` -- Plot MOOP results as parallel coordinates
+  * ``heatmap(moop)`` -- Plot MOOP results as heatmap
+  * ``petal(moop)`` -- Plot MOOP results as petal diagram
+  * ``radviz(moop)`` -- Plot MOOP results as RadViz
+  * ``star_coordinates(moop)`` -- Plot MOOP results as star coordinates
+
+ Development
+  * ``dummyFunction(moop)`` -- place functions here for testing
+
+"""
+
+import plotly.express as px         # 15.2 MB package
+# import plotly.graph_objects as go
+from tabulate import tabulate       # 29 kB package
+# from parmoo import MOOP
+# import numpy as np
 
 # des_type = moop.getDesignType()
 # obj_type = moop.getObjectiveType()
@@ -10,8 +46,17 @@ from parmoo import MOOP
 # obj_db = moop.getObjectiveData()
 # sim_db = moop.getSimulationData()
 
-# print Data Types to terminal
+
 def printDataTypes(moop):
+    """ Print data types to terminal.
+
+    Args:
+        moop (MOOP): A ParMOO MOOP containing the data types to print.
+
+    Returns:
+        None
+
+    """
     des_type = moop.getDesignType()
     obj_type = moop.getObjectiveType()
     sim_type = moop.getSimulationType()
@@ -22,14 +67,36 @@ def printDataTypes(moop):
     print("Objective type:         " + str(obj_type))
     print("Constraint type:        " + str(const_type))
 
-# print Pareto Front to terminal
-def printPF(moop):
+#
+# ! THESE FUNCTIONS PRINT DATA TO THE TERMINAL IN BLOCK FORM
+#
+
+
+def printPF_raw(moop):
+    """ Print Pareto Front to terminal as raw data.
+
+    Args:
+        moop (MOOP): A ParMOO MOOP containing the Pareto Front to print.
+
+    Returns:
+        None
+
+    """
     print("\nPARETO FRONT:\n")
     pf = moop.getPF()
     print(pf)
 
-# print Objective Data to terminal
-def printObjectiveData(moop):
+
+def printObjectiveData_raw(moop):
+    """ Print objective database to terminal as raw data.
+
+    Args:
+        moop (MOOP): A ParMOO MOOP containing the objective database to print.
+
[…]
