### 2022-06-29 #4468 @thchang
Hello all,

Glad to see some reviews are coming in!  Thanks all reviewers for your time!

Just wondering, will we get more detailed feedback than just this checklist?  Some of these items may be self-explanatory, but for many (e.g., the License entry, statement of need, installation, functionality, and references) it is not obvious to me why we are not satisfying these requirements, and I would appreciate more detailed feedback!

Thanks,
---Tyler

P.S., I have some revisions of my own, mostly other MOO software that was published in the last couple years that I was unaware of at the time I started this project, which should be included in the JOSS paper.  Also, I am working on some minor updates to the software.  I am assuming that I should I wait until the first round of reviews has completed before adding any such updates to the paper, like one would for a normal journal submission?
