### 2022-06-29 44ebbc4
adding support for integer and custom variables
