+++ parmoo/moop.py
@@ -89,16 +89,18 @@ class MOOP:
 
     # Slots for the MOOP class
     __slots__ = ['n', 'm', 'm_total', 'o', 'p', 's', 'n_dat', 'lb', 'ub',
-                 'n_cat_d', 'cat_lb', 'cat_scale', 'RSVT', 'mean', 'n_cat',
-                 'n_cont', 'n_lvls', 'des_order', 'cat_names',
-                 'sim_names', 'des_names', 'obj_names', 'const_names',
-                 'lam', 'objectives', 'data', 'sim_funcs', 'sim_db',
-                 'des_tols', 'searches', 'surrogates', 'optimizer',
+                 'n_cat_d', 'n_custom_d', 'cat_lb', 'cat_scale', 'RSVT',
+                 'mean', 'custom_embedders', 'custom_extracters',
+                 'n_cat', 'n_cont', 'n_int', 'n_custom', 'n_raw', 'n_lvls',
+                 'des_order', 'int_inds', 'cat_names', 'sim_names',
+                 'des_names', 'obj_names', 'const_names',
+                 'lam', 'epsilon', 'objectives', 'data', 'sim_funcs',
+                 'sim_db', 'des_tols', 'searches', 'surrogates', 'optimizer',
                  'constraints', 'hyperparams', 'acquisitions', 'history',
                  'scale', 'scaled_lb', 'scaled_ub', 'scaled_des_tols',
-                 'cat_des_tols', 'use_names', 'iteration', 'checkpoint',
-                 'checkpointfile', 'checkpoint_data', 'new_checkpoint',
-                 'new_data']
+                 'cat_des_tols', 'custom_des_tols', 'use_names', 'iteration',
+                 'checkpoint', 'checkpointfile', 'checkpoint_data',
+                 'new_checkpoint', 'new_data']
 
     def __embed__(self, x):
         """ Embed a design input as n-dimensional vector for ParMOO.
@@ -117,15 +119,19 @@ class MOOP:
         """
 
         # Unpack x into an ordered unstructured array
-        x_tmp = np.zeros(self.n_cont + self.n_cat)
+        x_tmp = np.zeros(self.n_cont + self.n_cat + self.n_int + self.n_custom
+                         + self.n_raw)
         if self.use_names:
             x_labels = []
             for d_name in self.des_names:
                 x_labels.append(x[d_name[0]])
             for i, j in enumerate(self.des_order):
-                if (j >= self.n_cont) and (len(self.cat_names[j - self.n_cont])
-                                           > 0):
-                    x_tmp[i] = float(self.cat_names[j - self.n_cont].index(
+                if ((i in range(self.n_cont+self.n_int,
+                                self.n_cont+self.n_int+self.n_cat))
+                    and (len(self.cat_names[i - self.n_cont - self.n_int]))
+                         > 0):
+                    x_tmp[i] = float(self.cat_names[j - self.n_cont -
+                                                    self.n_int].index(
                                                                 x_labels[j]))
                 else:
                     x_tmp[i] = x_labels[j]
@@ -141,6 +147,14 @@ class MOOP:
         # Pull inside bounding box, in case perturbed outside
         xx[start:end] = np.maximum(xx[start:end], self.scaled_lb[start:end])
         xx[start:end] = np.minimum(xx[start:end], self.scaled_ub[start:end])
+        # Rescale the integer variables
+        start = end
+        end = start + self.n_int
+        xx[start:end] = ((x_tmp[start:end] - self.lb[start:end]) /
+                         self.scale[start:end] + self.scaled_lb[start:end])
+        # Pull inside bounding box, in case perturbed outside
[…]
