+++ parmoo/tests/unit_tests/test_moop.py
@@ -157,6 +157,43 @@ def test_MOOP_addDesign_bad_cat():
                         'levels': ["hi"]})
 
 
+def test_MOOP_addDesign_bad_int():
+    """ Check that the MOOP class handles adding bad integer variables.
+
+    Initialize a MOOP objects, and add several bad integer design
+    variables.
+
+    """
+
+    from parmoo import MOOP
+    from parmoo.optimizers import LocalGPS
+    import pytest
+
+    # Initialize a MOOP with no hyperparameters
+    moop = MOOP(LocalGPS)
+    # Add some bad integer variables
+    with pytest.raises(AttributeError):
+        moop.addDesign({'des_type': "integer"})
+    with pytest.raises(TypeError):
+        moop.addDesign({'des_type': "integer",
+                        'des_tol': "hello world",
+                        'lb': 0.0,
+                        'ub': 1.0})
+    with pytest.raises(TypeError):
+        moop.addDesign({'des_type': "integer",
+                        'lb': "hello",
+                        'ub': "world"})
+    with pytest.raises(ValueError):
+        moop.addDesign({'des_type': "integer",
+                        'lb': 0,
+                        'ub': 0})
+    with pytest.raises(TypeError):
+        moop.addDesign({'name': 5,
+                        'des_type': "integer",
+                        'lb': 0,
+                        'ub': 1})
+
+
 def test_MOOP_addDesign():
     """ Check that the MOOP class handles adding design variables properly.
 
@@ -169,17 +206,17 @@ def test_MOOP_addDesign():
 
     # Initialize a MOOP with no hyperparameters
     moop = MOOP(LocalGPS)
-    # Now add some continuous design variables
+    # Now add some continuous and integer design variables
     assert(moop.n == 0)
     moop.addDesign({'des_type': "continuous",
                     'lb': 0.0,
                     'ub': 1.0})
     assert(moop.n == 1)
-    moop.addDesign({'des_type': "continuous",
-                    'des_tol': 0.01,
-                    'lb': 0.0,
-                    'ub': 1.0})
+    moop.addDesign({'des_type': "integer",
+                    'lb': 0,
+                    'ub': 4})
     assert(moop.n == 2)
+    assert(moop.n_int == 1)
     moop.addDesign({'name': "x_3",
                     'des_type': "continuous",
                     'des_tol': 0.01,
@@ -198,7 +235,7 @@ def test_MOOP_addDesign():
                     'lb': 0.0,
                     'ub': 1.0})
     assert(moop.n == 6)
-    # Now add some good categorical design variables
+    # Now add some categorical design variables
     assert(moop.n_cat == 0)
     moop.addDesign({'des_type': "categorical",
                     'levels': 2})
@@ -218,10 +255,10 @@ def test_MOOP_addDesign():
     moop.addDesign({'des_type': "continuous",
                     'lb': 0.0,
                     'ub': 1.0})
-    assert(moop.n_cont == 7)
+    assert(moop.n_cont == 6)
     moop.addDesign({'lb': 0.0,
                     'ub': 1.0})
-    assert(moop.n_cont == 8)
+    assert(moop.n_cont == 7)
 
 
 def test_MOOP_embed_extract_unnamed1():
@@ -243,8 +280,9 @@ def test_MOOP_embed_extract_unnamed1():
     # Initialize a MOOP with no hyperparameters
     moop = MOOP(LocalGPS)
     # Add two continuous variables and check that they are embedded correctly
[…]
