+++ parmoo/viz/plots.py
@@ -27,16 +27,18 @@ The functions are:
   * ``radviz(moop)`` -- Plot MOOP results as RadViz
   * ``star_coordinates(moop)`` -- Plot MOOP results as star coordinates
 
- Development
+ Utilities
+
   * ``dummyFunction(moop)`` -- place functions here for testing
 
 """
 
+from ast import Return
 import plotly.express as px         # 15.2 MB package
-# import plotly.graph_objects as go
+import plotly.graph_objects as go
 from tabulate import tabulate       # 29 kB package
 # from parmoo import MOOP
-# import numpy as np
+import numpy as np
 
 # des_type = moop.getDesignType()
 # obj_type = moop.getObjectiveType()
@@ -302,17 +304,6 @@ def scatter(moop):
     for obj_key in obj_type.names:
         axes.append(obj_key)
 
-    # sim_axes = [] # axes from the simulation database
-    # for
-
-    # plot pareto front
-    pf_fig = px.scatter_matrix(pf,
-                               dimensions=axes,
-                               title="Pareto Front",
-                               )
-    pf_fig.update_traces(diagonal_visible=False)
-    pf_fig.show()
-
     # plot objectives
     obj_fig = px.scatter_matrix(obj_db,
                                 dimensions=axes,
@@ -321,14 +312,13 @@ def scatter(moop):
     obj_fig.update_traces(diagonal_visible=False)
     obj_fig.show()
 
-    # plot simulation data
-    # sim_fig = px.scatter_matrix(sim_db,
-    #                             dimensions=axes,
-    #                             title="Simulation Data",
-    #                             )
-    # sim_fig.update_traces(diagonal_visible=False)
-    # sim_fig.show()
-
+    # plot pareto front
+    pf_fig = px.scatter_matrix(pf,
+                               dimensions=axes,
+                               title="Pareto Front",
+                               )
+    pf_fig.update_traces(diagonal_visible=False)
+    pf_fig.show()
 
 def scatter3d(moop):
     """ Display MOOP results as matrix of 3D scatterplots.
@@ -360,43 +350,82 @@ def radar(moop):
 
     """
 
-    # algo
-    # each index in db gets a plot
-    # for each objective db
-        # objective[index] is the r for that plot
-        
-
-    # des_type = moop.getDesignType()
     obj_type = moop.getObjectiveType()
-    # sim_type = moop.getSimulationType()
-    # const_type = moop.getConstraintType()
+    obj_db = moop.getObjectiveData()
     pf = moop.getPF()
-    # obj_db = moop.getObjectiveData()
-    # sim_db = moop.getSimulationData()
 
-    # choose axes
-    axes = []  # each axis relates to an objective
+    # create axes
+    axes = [] 
     for obj_key in obj_type.names:
         axes.append(obj_key)
 
-    # plot pareto front
-    pf_fig = px.line_polar(pf,
-                           theta=axes,
-                           r=pf[axes]
-                           # title="Pareto Front",
-                           )
-    # pf_fig.update_traces(diagonal_visible=False)
+    # if there are less than three objectives, prompt alternate options
+    if recommendPlot(moop=moop, objective_count=len(axes), min_count=3, plot_name='radar') == False:
+        return
+
+    # create figure
+    obj_fig = go.Figure()
+    pf_fig = go.Figure()
+
+    # plotting code here
+    for i in range(len(obj_db)):
+        values = []
+        for obj_key in obj_type.names:
+            values.append(obj_db[obj_key][i])
+            # print(tabulate(values))
+        traceName = ("design " + str(i))
+        obj_fig.add_trace(go.Scatterpolar(
+            r=values,
+            theta=axes,
+            # fill='toself',
[…]
