### 2022-06-30 parmoo/parmoo#32 @Viech
Misleading exception when running a libE_MOOP script without mpirun/arguments
When I first tried to run the `libe_basic_ex.py` example, I thought that there was a bug:
```
% PYTHONPATH=. python examples/libe_basic_ex.py
Traceback (most recent call last):
  File "…/parmoo/examples/libe_basic_ex.py", line 60, in <module>
    my_moop.solve(sim_max=30)
  File "…/parmoo/parmoo/extras/libe.py", line 715, in solve
    raise ValueError("Cannot run ParMOO + libE with less than 2 " +
ValueError: Cannot run ParMOO + libE with less than 2 workers -- aborting...
```
I then learned that the following works:
```
% PYTHONPATH=. python examples/libe_basic_ex.py --comms local --nworkers 32
```

The `ValueError` does not really explain what went wrong. I suggest to either raise a more descriptive exception (e.g. `RuntimeError: You need to specify the number of workers on the command line.`), or to default to #CPU many workers. If possible, I would personally prefer the latter.
