### 2022-06-30 #4468 @Viech
It was a pleasure to review this! I will certainly keep this library in mind for when I need to solve a multi-objective problem with expensive evaluation.

To put my comments into perspective, I was previously aware of *design of experiments* type problems, various types of metaheuristics, and multi-objective optimization, but have not heard the term *simulation optimization* before.

I marked as *nitpicks* small suggestions that you are invited to ignore, and as *optional suggestions* larger changes that you could consider for the future but that should also not get in the way of publication. (I will tick the associated boxes when you respond to these suggestions or when all other issues are resolved.)

**General checks/License**

- Nitpick: Your copyright line reads `Copyright (c) 2022, parmoo`. I'm not sure if mentioning the program name here is appropriate. Maybe `Copyright (c) 2022, parmoo developers`, or using names or institutions, would be better?

**Functionality/Installation**

- Nitpick: Writing `pip install [--user] parmoo[extras]` might confuse inexperienced users as the `[--user]` is optional (and the brackets must be removed) while for `[extras]` the brackets are part of the input. Also, for some shells (e.g. zsh) one would need to write `"parmoo[extras]"` instead of `parmoo[extras]` (the former should work for all shells).
- Optional suggestion: To support Windows and also Mac users, I've found Anaconda to be a useful distribution system in addition to PyPI/pip.

**Documentation**

- In the *Quickstart* section, the concepts of design variables, simulations, objectives and constraints are well explained but the surrogate model, its scalarazation, and the acquisition function remain somewhat enigmatic. In particular, I did not initially understand why "we must add one or more acquisition functions". I suggest to either explain these concepts already in *Quickstart* or to refer the reader to explanations within other sections where appropriate.
- The acquisition function is missing from the *Glossary* subsection of *Learn about MOOPs*. Non-specialist intuition could also be given in the *Adding Acquisitions* subsection of *Write a ParMOO Script*.
- In the *Minimum Working Example*, it would be helpful if the output could be explained and/or plotted.
- JOSS requirement: You should mention the target audience in the documentation.
- JOSS requirement: You should mention how to report issues and seek support.

**Software paper**

- In the summary, which I otherwise find very accessible, it is not clear to me what is meant by a "simulation-based structure". Briefly explaining this concept could make also the examples more accessible to non-specialists (here I do not understand the parts that explain where said structure comes from).
- "Scientific software packages for solving multiobjective simulation optimization problems include": Here I would prefer to read the names of the software packages (with authors and years in parentheses). I would also be nice to have a partition into Python packages and other software (unless all of them are for Python; then this should be mentioned).
- In the EDF example, I do not quite understand what is meant by ParMOO being used to "minimiz[e] the error between expensive simulation outputs and experimental data".
- If available, references to the mentioned research using ParMOO would be very interesting.
- Nitpick: The HPC abbreviation is not introduced.
- Nitpick: For your *ACM Transactions on Mathematical Software* paper, in the references, maybe add a note/comment that this paper was just accepted. Otherwise, it looks like volume and issue numbers are missing.
- Nitpick/unsure: The string "In arXiv preprint arXiv:2105.03000." in the references looks odd (but I do not know the prefered way to cite an arXiv paper in the JOSS).
- JOSS requirement: The *statement of need* section should mention the target audience.
- (In my experience, the editor will later ask you to investigate the missing DOIs that were mentioned by @editorialbot.)

**Code**

JOSS reviews do not seem to include code review, so all of these points are just suggestions.

- You do not need to have a `return` statement at the end of a function. (Python automatically does `return None` at the end of a non-generator function.)
- When using the `ABC` base class, you should use the `abc.abstractmethod` decorator instead of raising `NotImplementedError`(e.g. in `sim_func` add `@abstractmethod` above `__call__`, then inside `__call__`, just `pass`). Otherwise I do not think that adding the `ABC` base class has any effect.
- You could make your `xerror` function future-proof by giving the arguments default values and callling it with keyword arguments, like you do with `__check_optionals__` (i.e. instead of `xerror(1, lb, ub, {})` you would write `xerror(lb=lb, ub=ub)`). This would allow you to add new checks later on.
- Is there a reason you do `from parmoo.util import xerror` within `__init__`? It does not seem to produce a circular import if you import it at the start of the source file, which is the typical reason to import within a function. Importing at the start of the file might profit software that analyzes your code, including your IDE.
- `if 'name' in arg.keys()` -> `if 'name' in arg`
- You have some weird `try: assert(_) except BaseException: raise _` checks in `sum_sim_bound`. Can you not just condition the raised exception on what you check within `assert`? Normally, `assert` is reserved for things that, when they go wrong, imply an error in the software that makes the assertion (as opposed to a user error).
- When checking whether something is a NumPy array, the proper exception would by `TypeError` instead of `ValueError`. Although, I cannot think of a scenario where this would make any difference, so it doesn't get any more nitpicky than this!
- In `setup.py`, `exec(open("parmoo/version.py").read())` looks a bit dangerous. Have you considered `from parmoo.version import __version__`?
- Thank you for making me aware of `__slots__`, I've literally never seen this before!

**Issues** (not impeding acceptance; only for reference)

- https://github.com/parmoo/parmoo/issues/32
