### 2022-06-30 bcd6539
upgrade radar(), nuke recommendPlot()
added hover info, changed plot recommendations to use warnings
instead of an interactive dialogue, and other minor improvements
