+++ parmoo/viz/plots.py
@@ -31,6 +31,10 @@ The functions are:
 
   * ``dummyFunction(moop)`` -- place functions here for testing
 
+For all interactve browser plot functions, there is a known issue where Plotly images
+may not export from Safari correctly. To resolve this issue,
+change your default browser to Chrome, Firefox, or Edge.
+
 """
 
 from ast import Return
@@ -40,6 +44,7 @@ import plotly.io as pio
 from tabulate import tabulate       # 29 kB package
 # from parmoo import MOOP
 import numpy as np
+import warnings
 
 # des_type = moop.getDesignType()
 # obj_type = moop.getObjectiveType()
@@ -292,13 +297,9 @@ def scatter(moop):
         None
 
     """
-    # des_type = moop.getDesignType()
     obj_type = moop.getObjectiveType()
-    # sim_type = moop.getSimulationType()
-    # const_type = moop.getConstraintType()
     pf = moop.getPF()
     obj_db = moop.getObjectiveData()
-    # sim_db = moop.getSimulationData()
 
     # choose axes
     axes = []  # each axis relates to an objective
@@ -338,7 +339,7 @@ def scatter3d(moop):
     pass
 
 
-def radar(moop, db):
+def radar(moop, db='pf'):
     """ Display MOOP results as radar plot.
 
     Create an interactive plot that displays in the browser.
@@ -359,14 +360,16 @@ def radar(moop, db):
     obj_db = moop.getObjectiveData()
     pf = moop.getPF()
 
-    if (db == 'pf') or (db == None):
+    if (db == 'pf'):
         database = pf
         plotTitle = "Pareto Front"
     elif db == 'obj':
         database = obj_db
         plotTitle = "Objective Data"
     else:
-        raise ValueError("'" + str(db) + "' is not an acceptible value for 'db'")
+        message = "'" + str(db) + "' is not an acceptible value for 'db'\n"
+        message += "Consider using 'pf' or 'obj' instead."
+        raise ValueError(message)
 
     # create axes
     axes = []
@@ -374,56 +377,64 @@ def radar(moop, db):
         axes.append(obj_key)
 
     # if there are less than three objectives, prompt alternate options
-    if recommendPlot(moop, objective_count=len(axes), min_count=3, plot_name='radar') == False:
-        return
+    if len(axes) < 3:
+        message = """
+        Radar plots are best suited for MOOPs with at least 3 objectives.
+        Consider using a scatterplot or parallel coordinates plot instead.
+        """
+        warnings.warn(message)
 
-    # create figure
+    # plot figure
     fig = go.Figure()
-
-    # plotting code here
     for i in range(len(database)):
+        if plotTitle == "Pareto Front":
+            hoverInfo = ""
+        else:
+            hoverInfo = "Design #" + str(i) + "\n"
+        for key in database.dtype.names:
+            hoverInfo += str(key)
+            hoverInfo += ": "
+            hoverInfo += str(database[key][i])
+            hoverInfo += "<br>"
+            # since plotly is JavaScript-based
+            # it uses HTML string formatting
         values = []
         for obj_key in obj_type.names:
             values.append(database[obj_key][i])
-            traceName = ("design #" + str(i))
+            traceName = ("design #" + str(i + 1))
         fig.add_trace(go.Scatterpolar(
             r=values,
             theta=axes,
-            name=traceName
+            name=traceName,
+            hovertext=hoverInfo
         ))
 
-    # aesthetics code here
+    # improve aesthetics
+    fig.update_traces(
+        hoverinfo='text',
+        selector=dict(
+            type='scatterpolar'
+        )
+    )
     fig.update_layout(
         polar=dict(
             radialaxis=dict(
                 visible=True
-            )),
+            )
[…]
