### 2022-06-30 d16f8f4
update for Better Comments
Update to play nicely with Aaron Bond's Better Comments VS Code
extension.
