### 2022-06-30 d37d3e8
embeddings passing some tests, more needed
