+++ parmoo/moop.py
@@ -88,7 +88,8 @@ class MOOP:
     """
 
     # Slots for the MOOP class
-    __slots__ = ['n', 'm', 'm_total', 'o', 'p', 's', 'n_dat', 'lb', 'ub',
+    __slots__ = ['n', 'm', 'm_total', 'o', 'p', 's', 'n_dat',
+                 'cont_lb', 'cont_ub', 'int_lb', 'int_ub',
                  'n_cat_d', 'n_custom_d', 'cat_lb', 'cat_scale', 'RSVT',
                  'mean', 'custom_embedders', 'custom_extracters',
                  'n_cat', 'n_cont', 'n_int', 'n_custom', 'n_raw', 'n_lvls',
@@ -128,8 +129,8 @@ class MOOP:
             for i, j in enumerate(self.des_order):
                 if ((i in range(self.n_cont+self.n_int,
                                 self.n_cont+self.n_int+self.n_cat))
-                    and (len(self.cat_names[i - self.n_cont - self.n_int]))
-                         > 0):
+                    and (len(self.cat_names[i - self.n_cont - self.n_int])
+                         > 0)):
                     x_tmp[i] = float(self.cat_names[j - self.n_cont -
                                                     self.n_int].index(
                                                                 x_labels[j]))
@@ -139,32 +140,32 @@ class MOOP:
             x_tmp = x[self.des_order]
         # Create the output array
         xx = np.zeros(self.n)
-        # Rescale the continuous variables
+        # Rescale the continuous and integer variables
         start = 0
         end = self.n_cont
-        xx[start:end] = ((x_tmp[start:end] - self.lb[start:end]) /
+        xx[start:end] = ((x_tmp[start:end] - self.cont_lb[:]) /
                          self.scale[start:end] + self.scaled_lb[start:end])
         # Pull inside bounding box, in case perturbed outside
         xx[start:end] = np.maximum(xx[start:end], self.scaled_lb[start:end])
         xx[start:end] = np.minimum(xx[start:end], self.scaled_ub[start:end])
-        # Rescale the integer variables
+        # Rescale the continuous and integer variables
         start = end
         end = start + self.n_int
-        xx[start:end] = ((x_tmp[start:end] - self.lb[start:end]) /
+        xx[start:end] = ((x_tmp[start:end] - self.int_lb[:]) /
                          self.scale[start:end] + self.scaled_lb[start:end])
         # Pull inside bounding box, in case perturbed outside
         xx[start:end] = np.maximum(xx[start:end], self.scaled_lb[start:end])
         xx[start:end] = np.minimum(xx[start:end], self.scaled_ub[start:end])
         # Embed the categorical variables
         if self.n_cat_d > 0:
+            start = end
+            end = start + self.n_cat_d
             bvec = np.zeros(sum(self.n_lvls))
             count = 0
             for i, n_lvl in enumerate(self.n_lvls):
-                bvec[count + int(x_tmp[self.n_cont + i])] = 1.0
+                bvec[count + int(x_tmp[start + i])] = 1.0
                 count += n_lvl
             bvec -= self.mean
-            start = end
-            end = start + self.n_cat_d
             xx[start:end] = ((np.matmul(self.RSVT, bvec) - self.cat_lb[:])
                              / self.scale[start:end]
                              + self.scaled_lb[start:end])
@@ -203,12 +204,20 @@ class MOOP:
                       + self.n_raw)
         # Descale the continuous variables
         start = 0
-        end = self.n_cont + self.n_int
+        end = self.n_cont
         xx[start:end] = ((x[start:end] - self.scaled_lb[start:end])
-                         * self.scale[start:end] + self.lb[start:end])
+                         * self.scale[start:end] + self.cont_lb[:])
         # Pull inside bounding box, in case perturbed outside
[…]
