### 2022-07-01 8d2b026
restructure viz
separated plotting functions into a plot module and a print module
renamed plots.py to plot.py
