+++ parmoo/viz/plot.py
@@ -2,21 +2,6 @@
 
 The functions are:
 
- Print raw data to terminal:
-  * ``printDataTypes(moop)`` -- print data type
-  * ``printPF_raw(moop)`` -- print Pareto Front
-  * ``printObjectiveData_raw(moop)`` -- print objective database
-  * ``printSimulationData_raw(moop)`` -- print simulation database(s)
-  * ``printDesignVariables_raw(moop)`` -- print design variable data
-  * ``printConstraints_raw(moop)`` -- print constraint data
-  * ``printMOOP_raw(moop)`` -- print all MOOP data
-
- Print data to terminal as table
-  * ``printPF(moop)`` -- print Pareto Front
-  * ``printObjectiveData(moop)`` -- print objective database
-  * ``printSimulationData(moop)`` -- print simulation database(s)
-  * ``printMOOP(moop)`` -- print all MOOP data
-
  Display data in interactive browser plot
   * ``scatter(moop)`` -- Plot MOOP results as matrix of 2D scatterplots
   * ``scatter3d(moop)`` -- Plot MOOP results as matrix of 3D scatterplots
@@ -41,10 +26,9 @@ or Edge.
 import plotly.express as px         # 15.2 MB package
 import plotly.graph_objects as go
 # import plotly.io as pio
-from tabulate import tabulate       # 29 kB package
 # from parmoo import MOOP
 # import numpy as np
-import warnings
+import warnings                     # native python package
 
 # des_type = moop.getDesignType()
 # obj_type = moop.getObjectiveType()
@@ -55,229 +39,8 @@ import warnings
 # sim_db = moop.getSimulationData()
 
 
-def printDataTypes(moop):
-    """ Print data types to terminal.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing the data types to print.
-
-    Returns:
-        None
-
-    """
-    des_type = moop.getDesignType()
-    obj_type = moop.getObjectiveType()
-    sim_type = moop.getSimulationType()
-    const_type = moop.getConstraintType()
-    print("\nDATA TYPES FOR YOUR MOOP:\n")
-    print("Design variable type:   " + str(des_type))
-    print("Simulation output type: " + str(sim_type))
-    print("Objective type:         " + str(obj_type))
-    print("Constraint type:        " + str(const_type))
-
 #
-# ! THESE FUNCTIONS PRINT DATA TO THE TERMINAL IN BLOCK FORM
-#
-
-
-def printPF_raw(moop):
-    """ Print Pareto Front to terminal as raw data.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing the Pareto Front to print.
-
-    Returns:
-        None
-
-    """
-    print("\nPARETO FRONT:\n")
-    pf = moop.getPF()
-    print(pf)
-
-
-def printObjectiveData_raw(moop):
-    """ Print objective database to terminal as raw data.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing the objective database to print.
-
-    Returns:
-        None
-
-    """
-    print("\nOBJECTIVE DATA:\n")
-    obj_type = moop.getObjectiveType()
-    obj_db = moop.getObjectiveData()
-    for obj_key in obj_type.names:
-        print(f"Objective: {obj_key}:")
-        print(obj_db[obj_key])
-
-
-def printSimulationData_raw(moop):
-    """ Print simulation database(s) to terminal as raw data.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing simulation database(s) to print.
-
-    Returns:
-        None
-
-    """
-    print("\nSIMULATION DATA:\n")
-    sim_db = moop.getSimulationData()
[…]
