+++ parmoo/viz/__init__.py
@@ -1,0 +1,2 @@
+from .plot import *
+from .print import *
