+++ parmoo/moop.py
@@ -93,7 +93,7 @@ class MOOP:
                  'n_cat_d', 'n_custom_d', 'cat_lb', 'cat_scale', 'RSVT',
                  'mean', 'custom_embedders', 'custom_extracters',
                  'n_cat', 'n_cont', 'n_int', 'n_custom', 'n_raw', 'n_lvls',
-                 'des_order', 'int_inds', 'cat_names', 'sim_names',
+                 'des_order', 'cat_names', 'sim_names',
                  'des_names', 'obj_names', 'const_names',
                  'lam', 'epsilon', 'objectives', 'data', 'sim_funcs',
                  'sim_db', 'des_tols', 'searches', 'surrogates', 'optimizer',
@@ -137,7 +137,7 @@ class MOOP:
                 else:
                     x_tmp[i] = x_labels[j]
         else:
-            x_tmp = x[self.des_order]
+            x_tmp[self.des_order] = x[:]
         # Create the output array
         xx = np.zeros(self.n)
         # Rescale the continuous and integer variables
@@ -180,6 +180,11 @@ class MOOP:
             end = start + self.n_custom_d[i]
             xx[start:end] = embed_i(x_tmp[self.n_cont + self.n_cat +
                                           self.n_int + i])
+        # Embed the raw variables
+        start = end
+        end = start + self.n_raw
+        xx[start:end] = x_tmp[self.n_cont + self.n_cat + self.n_int +
+                              self.n_custom:]
         return xx
 
     def __extract__(self, x):
@@ -200,7 +205,7 @@ class MOOP:
         """
 
         # Create the output array
-        xx = np.zeros(self.n_cont + self.n_cat + self.n_cont + self.n_custom
+        xx = np.zeros(self.n_cont + self.n_cat + self.n_int + self.n_custom
                       + self.n_raw)
         # Descale the continuous variables
         start = 0
@@ -239,11 +244,20 @@ class MOOP:
         for i, ex_i in enumerate(self.custom_extracters):
             start = end
             end = start + self.n_custom_d[i]
-            xx[self.n_cont + self.n_cat + self.n_int + i] = ex_i(x[start:end])
+            if not self.use_names:
+                xx[self.n_cont + self.n_cat + self.n_int + i] = \
+                    ex_i(x[start:end])
+        # Extract the raw variables
+        start = end
+        end = start + self.n_raw
+        xx[self.n_cont + self.n_cat + self.n_int + self.n_custom:] = \
+            x[start:end]
         # Unshuffle xx and pack into a numpy structured array
         if self.use_names:
             out = np.zeros(1, dtype=np.dtype(self.des_names))
+            n_customs = 0
             for i, j in enumerate(self.des_order):
+                # Unpack categorical variables when cat_names given
                 if ((i in range(self.n_cont+self.n_int,
                                 self.n_cont+self.n_int+self.n_cat))
                     and (len(self.cat_names[i - self.n_cont - self.n_int])
@@ -252,6 +266,16 @@ class MOOP:
                                                                 self.n_cont -
                                                                 self.n_int]
                                                                [int(xx[j])])
+                # Unpack custom variables
+                elif (i in range(self.n_cont + self.n_int + self.n_cat,
+                                 self.n_cont + self.n_int + self.n_cat +
+                                 self.n_custom)):
+                    start = (self.n_cont + self.n_cat_d + self.n_int +
+                             sum(self.n_custom_d[:n_customs]))
+                    end = start + self.n_custom_d[n_customs]
+                    exi = self.custom_extracters[n_customs]
+                    out[self.des_names[i][0]] = exi(x[start:end])
+                    n_customs += 1  # increment counter
                 else:
[…]
