+++ parmoo/tests/unit_tests/test_moop.py
@@ -32,489 +32,6 @@ def test_MOOP_init():
     assert(moop.hyperparams['test'] == 0)
 
 
-def test_MOOP_addDesign_bad_cont():
-    """ Check that the MOOP class handles adding bad continuous variables.
-
-    Initialize a MOOP objects, and add several bad continuous design
-    variables.
-
-    """
-
-    from parmoo import MOOP
-    from parmoo.optimizers import LocalGPS
-    import pytest
-
-    # Initialize a MOOP with no hyperparameters
-    moop = MOOP(LocalGPS)
-    # Try to add some bad design variable types
-    with pytest.raises(TypeError):
-        moop.addDesign([])
-    with pytest.raises(TypeError):
-        moop.addDesign({'des_type': 1.0})
-    with pytest.raises(ValueError):
-        moop.addDesign({'des_type': "hello world"})
-    # Add some bad continuous variables
-    with pytest.raises(AttributeError):
-        moop.addDesign({'des_type': "continuous"})
-    with pytest.raises(TypeError):
-        moop.addDesign({'des_type': "continuous",
-                        'des_tol': "hello world",
-                        'lb': 0.0,
-                        'ub': 1.0})
-    with pytest.raises(ValueError):
-        moop.addDesign({'des_type': "continuous",
-                        'des_tol': -1.0,
-                        'lb': 0.0,
-                        'ub': 1.0})
-    with pytest.raises(TypeError):
-        moop.addDesign({'des_type': "continuous",
-                        'lb': "hello",
-                        'ub': "world"})
-    with pytest.raises(ValueError):
-        moop.addDesign({'des_type': "continuous",
-                        'lb': 0.0,
-                        'ub': 0.0})
-    with pytest.raises(TypeError):
-        moop.addDesign({'name': 5,
-                        'des_type': "continuous",
-                        'lb': 0.0,
-                        'ub': 1.0})
-    # Add some bad continuous variables, using default option
-    with pytest.raises(AttributeError):
-        moop.addDesign({})
-    with pytest.raises(TypeError):
-        moop.addDesign({'des_tol': "hello world",
-                        'lb': 0.0,
-                        'ub': 1.0})
-    with pytest.raises(ValueError):
-        moop.addDesign({'des_tol': -1.0,
-                        'lb': 0.0,
-                        'ub': 1.0})
-    with pytest.raises(TypeError):
-        moop.addDesign({'lb': "hello",
-                        'ub': "world"})
-    with pytest.raises(ValueError):
-        moop.addDesign({'lb': 0.0,
-                        'ub': 0.0})
-    with pytest.raises(TypeError):
-        moop.addDesign({'name': 5,
-                        'lb': 0.0,
-                        'ub': 1.0})
-    # Try to use a repeated name to test error handling
-    with pytest.raises(ValueError):
-        moop.addDesign({'name': "x_1", 'lb': 0.0, 'ub': 1.0})
-        moop.addDesign({'name': "x_1",
-                        'des_type': "continuous",
-                        'lb': 0.0,
-                        'ub': 1.0})
-    # Add variables out of order
-    with pytest.raises(RuntimeError):
-        moop1 = MOOP(LocalGPS)
-        moop1.acquisitions.append(0)
-        moop1.addDesign({'des_type': "continuous",
-                         'lb': 0.0,
-                         'ub': 1.0})
[…]

[…]
