### 2022-07-01 eed5fd0
printing makeover
rebuild print module
add style keyword to choose between table or raw printed form
reduce number of printing functions by combining them and using keywords
to differentiate, and eliminate unnecessary functions
