+++ parmoo/viz/print.py
@@ -2,23 +2,14 @@
 
 The functions are:
 
- Print raw data to terminal:
+ Print data to terminal:
   * ``printDataTypes(moop)`` -- print data type
-  * ``printPF_raw(moop)`` -- print Pareto Front
-  * ``printObjectiveData_raw(moop)`` -- print objective database
-  * ``printSimulationData_raw(moop)`` -- print simulation database(s)
-  * ``printDesignVariables_raw(moop)`` -- print design variable data
-  * ``printConstraints_raw(moop)`` -- print constraint data
-  * ``printMOOP_raw(moop)`` -- print all MOOP data
-
- Print data to terminal as table
-  * ``printPF(moop)`` -- print Pareto Front
-  * ``printObjectiveData(moop)`` -- print objective database
-  * ``printSimulationData(moop)`` -- print simulation database(s)
-  * ``printMOOP(moop)`` -- print all MOOP data
+  * ``printPF(moop, style)`` -- print Pareto Front
+  * ``printObjectiveData(moop, style)`` -- print objective database
+  * ``printSimulationData(moop, style)`` -- print simulation database(s)
+  * ``printMOOP(moop, style)`` -- print all MOOP data
 
  Utilities
-
   * ``dummyPrintFunction(moop)`` -- place functions here for testing
 
 """
@@ -35,6 +26,11 @@ from tabulate import tabulate       # 29 kB package
 # pf = moop.getPF()
 # obj_db = moop.getObjectiveData()
 # sim_db = moop.getSimulationData()
+
+
+#
+# ! THESE FUNCTIONS PRINT DATA TO THE TERMINAL
+#
 
 
 def printDataTypes(moop):
@@ -57,206 +53,157 @@ def printDataTypes(moop):
     print("Objective type:         " + str(obj_type))
     print("Constraint type:        " + str(const_type))
 
-#
-# ! THESE FUNCTIONS PRINT DATA TO THE TERMINAL IN BLOCK FORM
-#
 
-
-def printPF_raw(moop):
-    """ Print Pareto Front to terminal as raw data.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing the Pareto Front to print.
-
-    Returns:
-        None
-
-    """
-    print("\nPARETO FRONT:\n")
-    pf = moop.getPF()
-    print(pf)
-
-
-def printObjectiveData_raw(moop):
-    """ Print objective database to terminal as raw data.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing the objective database to print.
-
-    Returns:
-        None
-
-    """
-    print("\nOBJECTIVE DATA:\n")
-    obj_type = moop.getObjectiveType()
-    obj_db = moop.getObjectiveData()
-    for obj_key in obj_type.names:
-        print(f"Objective: {obj_key}:")
-        print(obj_db[obj_key])
-
-
-def printSimulationData_raw(moop):
-    """ Print simulation database(s) to terminal as raw data.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing simulation database(s) to print.
-
-    Returns:
-        None
-
-    """
-    print("\nSIMULATION DATA:\n")
-    sim_db = moop.getSimulationData()
-    sim_type = moop.getSimulationType()
-    for sim_key in sim_type.names:
-        print(f"Simulation: {sim_key}:")
-        print(sim_db[sim_key])
-
-
-def printDesignVariables_raw(moop):
-    """ Print design variable data to terminal as raw data.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing design variable data to print.
-
-    Returns:
-        None
-
-    """
-    print("\nDESIGN VARIABLES:\n")
[…]
