### 2022-07-01 0bcf898
upgrade scatter() + parallel_coordinates()
added choice between plotting either the pareto front or objective data
