### 2022-07-01 c9353f5
improve hover info for scatterplot
hover info is not supported for parallel coordinates plot.
see https://github.com/plotly/plotly.js/issues/3012
I'm working on a viz workaround to let users see input info
when only output info is displayed on the graph,
but that looks to be a harder issue to solve
