### 2022-07-04 8fd5224
separated dummy() from committed files
removed dummyFunction from committed files because changes in testing
don't need to be in the git repo. Created separate dummy.py file for
testing that doesn't need to be committed
