+++ parmoo/viz/__init__.py
@@ -1,2 +1,3 @@
 from .plot import *
 from .print import *
+from .dummy import *
