### 2022-07-05 47f1d95
pushing incomplete work on named variable testing
