### 2022-07-06 3bef91a
updated moop class to handle named custom vars, one more fix before it is ready
