### 2022-07-06 a61acec
add static image export options
static export adds Kaleido dependency to Parmoo
(85.8 MB, pip installable).
Static image export types:
pdf
svg
webp
jpeg
png
These options export original figures (post-interaction export not yet
supported)
