+++ parmoo/viz/plot.py
@@ -44,7 +44,7 @@ import warnings                     # native python package
 #
 
 
-def scatter(moop, db='pf'):
+def scatter(moop, db='pf', export='none', browser=True):
     """ Display MOOP results as matrix of 2D scatterplots.
 
     Create an interactive plot that displays in the browser.
@@ -55,10 +55,26 @@ def scatter(moop, db='pf'):
 
     Args:
         moop (MOOP): A ParMOO MOOP containing the MOOP results to plot.
-        db (String): Indicates which database to plot.
-                     Defaults to "pf".
-                     Other options: "obj".
-                     A keyword other than "pf" or "obj" throws an error.
+        db (String): Indicate which database to plot.
+                     'pf' (default) plot Pareto Front
+                     'obj' plot objective data
+        export (String): Indicate export type
+                     'none' (default) don't export image file
+                     'pdf' export plot as pdf
+                     'svg' export plot as svg
+                     'webp' export plot as webp
+                     'jpeg' export plot as jpeg
+                     'png' export plot as png
+        browser (boolean): Choose to open interactive plot in browser window.
+                    True: (default) display interactive plot in browser window
+                    False: don't display interactive plot in browser window
+                    It is recommended that this setting be left on True
+                    The 'browser' and 'export' keywords will not
+                    interfere with each other. If you choose to export an image
+                    of the plot by using the 'export' keyword, and leave
+                    'browser' to True, you will BOTH export an image file to
+                    the current working directory AND open an interactive
+                    figure in the browser.
 
     Returns:
         None
@@ -78,10 +94,10 @@ def scatter(moop, db='pf'):
     # * choose database
     if (db == 'pf'):
         database = pf
-        plotTitle = "Pareto Front"
+        plotName = "Pareto Front"
     elif db == 'obj':
         database = obj_db
-        plotTitle = "Objective Data"
+        plotName = "Objective Data"
     else:
         message = "'" + str(db) + "' is not an acceptible value for 'db'\n"
         message += "Consider using 'pf' or 'obj' instead."
@@ -100,15 +116,18 @@ def scatter(moop, db='pf'):
     fig = px.scatter_matrix(database,
                             dimensions=axes,
                             # color=,
-                            title=plotTitle,
+                            title=plotName,
                             hover_data=hoverInfo
                             )
     fig.update_traces(diagonal_visible=False)
 
-
-
     # * display plot
-    fig.show()
+    if browser is True:
+        fig.show()
+
+    # * export plot
+    if export != 'none':
+        exportFile(fig=fig, plotName=plotName, fileType=export)
 
 
 def scatter3d(moop):
@@ -128,17 +147,33 @@ def scatter3d(moop):
     pass
 
 
-def radar(moop, db='pf'):
+def radar(moop, db='pf', export='none', browser=True):
     """ Display MOOP results as radar plot.
 
     Create an interactive plot that displays in the browser.
 
     Args:
         moop (MOOP): A ParMOO MOOP containing the results to plot.
-        db (String): Indicates which database to plot.
-                     Defaults to "pf".
-                     Other options: "obj".
-                     A keyword other than "pf" or "obj" throws an error.
+        db (String): Indicate which database to plot.
+                     'pf' (default) plot Pareto Front
+                     'obj' plot objective data
+        export (String): Indicate export type
+                     'none' (default) don't export image file
+                     'pdf' export plot as pdf
+                     'svg' export plot as svg
+                     'webp' export plot as webp
+                     'jpeg' export plot as jpeg
+                     'png' export plot as png
[…]
