### 2022-07-06 52db503
add post-interaction download options
increased download quality 4x
new options:
svg
jpeg
webp
