### 2022-07-08 086e518
scatter() doesn't plot matrix for two objectives
also removed dev test import from __init__
