### 2022-07-08 54c03d0
added a regression test for mixed variables
