### 2022-07-11 b2e2164
add download dataset as CSV button
TODO next: add download selection as CSV button
