### 2022-07-11 e66447f
viz use pandas instead of numpy arrays
Although numpy arrays are slightly faster, pandas are easier to work
with. Since viz is an inexpensive part of ParMOO, it makes sense to
convert to pandas from numpy arrays when processing vizualizations.
With this update, viz functions will export MOOP databases to pandas
format before use. This will make several dev issues easier to solve,
especially exports through Dash.
