### 2022-07-11 ff7af8a
added support for custom variable dtypes, and updated docs
