### 2022-07-12 4c1e71d
add button to download selection to csv
NOTE: this works for scatterplot and radar plot, but not for parallel
coordinates plot
