### 2022-07-12 78fa866
fix issue porting config from plotly to dash
fix issue where configuration settings weren't passed with the plotly
graph to the dash app. Fixing this issue means things like having
screenshot downloads be in svg format instead of png will work inside of
the dash app.
