### 2022-07-12 55443b2
control aspect ratio of screenshot
With this update, default behavior makes screenshot size and shape based
on window size and shape.
No-default behavior allows you to pre-set screenshot pixel dimensions
via keywords when creating a plot.
