### 2022-07-13 09287c0
updated READM.rst to reflect changes to docs
