### 2022-07-14 85720d2
goodbye print functions
