+++ parmoo/viz/__init__.py
@@ -1,4 +1,2 @@
 from .plot import *
-from .print import *
-# from .dummy import *
 from .dashboard import *
