### 2022-07-14 f46f93c
apps auto load (no terminal printing necessary)
