### 2022-07-14 6c67fc1
make hot reload off by default + option to turn on
Hot reloading is convenient - if you change something in the code
and save the file, the browser automatically updates with
the new changes. When using hot reloading, there is no need to re-run
the file for changes to take effect. However, this may cause issues for
users - if they are editing their script and save it before it's ready
to run, their plot may crash because of a syntax error. This could be
annoying for users who have autosave turned on.
