### 2022-07-14 26dbaaa
updated libE tests, examples, and docs to include calls to libE_MOOP.solve() inside if __name__ == '__main__': blocks, as required for using the Python MP spawn method (default on MacOS)
