+++ parmoo/tests/libe_tests/test_libe_gen.py
@@ -29,7 +29,9 @@ from parmoo.acquisitions import RandomConstraint
 n = 5
 o = 3
 
-def dtlz2_sim(x):
+# Define functions for unnamed runs
+
+def dtlz2_sim_unnamed(x):
     """ Evaluates the sim function for a collection of points given in
     ``H['x']``.
 
@@ -50,42 +52,12 @@ def dtlz2_sim(x):
         for y in x[:o-1-i]:
             f[i] *= math.cos(math.pi * y / 2.0)
     return f
+    
+def obj1_unnamed(x, s): return s[0]
+def obj2_unnamed(x, s): return s[1]
+def obj3_unnamed(x, s): return s[2]
 
-# Create a libE_MOOP
-moop = libE_MOOP(LocalGPS, hyperparams={})
-# Add n design vars
-for i in range(n):
-    moop.addDesign({'lb': 0.0, 'ub': 1.0})
-# Add simulation
-moop.addSimulation({'m': o,
-                    'sim_func': dtlz2_sim,
-                    'hyperparams': {'search_budget': 100},
-                    'search': LatinHypercube,
-                    'surrogate': GaussRBF,
-                    'sim_db': {},
-                    'des_tol': 0.00000001})
-# Add o objectives
-def obj1(x, s): return s[0]
-def obj2(x, s): return s[1]
-def obj3(x, s): return s[2]
-moop.addObjective({'obj_func': obj1})
-moop.addObjective({'obj_func': obj2})
-moop.addObjective({'obj_func': obj3})
-# Add 4 acquisition functions
-for i in range(4):
-    moop.addAcquisition({'acquisition': RandomConstraint})
-
-# Solve
-moop.solve()
-assert(moop.getObjectiveData()['x_vals'].shape[0] == 200)
-
-#print(moop.getPF())
-
-# Create a libE_MOOP with named variables
-moop = libE_MOOP(LocalGPS, hyperparams={})
-# Add n design vars
-for i in range(n):
-    moop.addDesign({'name': "x" + str(i + 1), 'lb': 0.0, 'ub': 1.0})
+# Define functions for named runs
 
 def dtlz2_sim_named(x):
     """ Evaluates the sim function for a collection of points given in
@@ -98,31 +70,65 @@ def dtlz2_sim_named(x):
     for i, name in enumerate(moop.moop.des_names):
         xx[i] = x[name[0]]
     # Use dtlz2_sim to evaluate
-    return dtlz2_sim(xx)
+    return dtlz2_sim_unnamed(xx)
 
-# Add simulation
-moop.addSimulation({'name': "DTLZ2",
-                    'm': o,
-                    'sim_func': dtlz2_sim_named,
-                    'hyperparams': {'search_budget': 100},
-                    'search': LatinHypercube,
-                    'surrogate': GaussRBF,
-                    'sim_db': {},
-                    'des_tol': 0.00000001})
-# Add o objectives
-def obj1(x, s): return s['DTLZ2'][0]
-def obj2(x, s): return s['DTLZ2'][1]
-def obj3(x, s): return s['DTLZ2'][2]
-moop.addObjective({'name': "obj1", 'obj_func': obj1})
-moop.addObjective({'name': "obj2", 'obj_func': obj2})
-moop.addObjective({'name': "obj3", 'obj_func': obj3})
-# Add 4 acquisition functions
-for i in range(4):
-    moop.addAcquisition({'acquisition': RandomConstraint})
+def obj1_named(x, s): return s['DTLZ2'][0]
+def obj2_named(x, s): return s['DTLZ2'][1]
+def obj3_named(x, s): return s['DTLZ2'][2]
 
-# Solve
-moop.solve()
[…]
