### 2022-07-14 c55dcd2
fixed a small issue in libE test cases, which causes a value to go undefined when using Python MP's spawn method
