### 2022-07-15 00ddf68
moved libe test funcs into a separate module for compatibility with coverage.py
