### 2022-07-18 8c562d4
fix issue where selection exports duplicate rows
