+++ parmoo/viz_test/plot.py
@@ -1,0 +1,615 @@
+""" This module contains a library of common plotting functions.
+
+The functions are:
+
+ Display data in interactive browser plot
+  * ``scatter(moop)`` -- Plot MOOP results as matrix of 2D scatterplots
+  * ``scatter3d(moop)`` -- Plot MOOP results as matrix of 3D scatterplots
+  * ``radar(moop)`` -- Plot MOOP results as radar
+  * ``parallel_coordinates(moop)`` -- Plot MOOP results as parallel coordinates
+  * ``heatmap(moop)`` -- Plot MOOP results as heatmap
+  * ``petal(moop)`` -- Plot MOOP results as petal diagram
+  * ``radviz(moop)`` -- Plot MOOP results as RadViz
+  * ``star_coordinates(moop)`` -- Plot MOOP results as star coordinates
+
+For all interactve browser plot functions, there is a known issue
+where Plotly images may not export from Safari correctly. If you
+encounter this issue, change your default browser to Chrome, Firefox,
+or Edge.
+
+"""
+
+import plotly.express as px         # 15.2 MB package
+import plotly.graph_objects as go
+# import plotly.io as pio
+# from parmoo import MOOP
+# import numpy as np
+import warnings                     # native python package
+from .dashboard import buildDashApp
+import pandas as pd
+
+# des_type = moop.getDesignType()
+# obj_type = moop.getObjectiveType()
+# sim_type = moop.getSimulationType()
+# const_type = moop.getConstraintType()
+# pf = moop.getPF()
+# obj_db = moop.getObjectiveData()
+# sim_db = moop.getSimulationData()
+
+
+#
+# ! THESE FUNCTIONS DISPLAY DATA IN AN INTERACTIVE BROWSER PLOT
+#
+
+# def printDataTypes(moop):
+#     """ Print data types to terminal.
+
+#     Args:
+#         moop (MOOP): A ParMOO MOOP containing the data types to print.
+
+#     Returns:
+#         None
+
+#     """
+#     des_type = moop.getDesignType()
+#     obj_type = moop.getObjectiveType()
+#     sim_type = moop.getSimulationType()
+#     const_type = moop.getConstraintType()
+#     print("\nDATA TYPES FOR YOUR MOOP:\n")
+#     print("Design variable type:   " + str(des_type))
+#     print("Simulation output type: " + str(sim_type))
+#     print("Objective type:         " + str(obj_type))
+#     print("Constraint type:        " + str(const_type))
+
+
+def scatter(moop,
+            db='pf',
+            export='none',
+            display=True,
+            height='auto',
+            width='auto',
+            verbose=True,
+            hot_reload=False,):
+    """ Display MOOP results as matrix of 2D scatterplots.
+
+    Create an interactive plot that displays in the browser.
+
+    For ``n`` objectives, generate an ``n x n`` matrix of 2D scatterplots.
+
+    Users can hover above an output point to see input information pop up.
+
+    Args:
+        moop (MOOP): A ParMOO MOOP containing the MOOP results to plot.
+        db (String): Choose database to plot
+                     'pf' (default) plot Pareto Front
+                     'obj' plot objective data
+        export (String): Export plot to working directory.
+                     'none' (default) don't export image file
+                     'html' export plot as html
+                     'pdf' export plot as pdf
+                     'svg' export plot as svg
+                     'webp' export plot as webp
+                     'jpeg' export plot as jpeg
+                     'png' export plot as png
+        browser (boolean): Display interactive plot in browser window.
+                    True: (default) display interactive plot in browser window
+                    False: don't display interactive plot in browser window
+                    It is recommended that this setting be left on True
+                    The 'browser' and 'export' keywords will not
+                    interfere with each other. If you choose to export an image
[…]
