### 2022-07-18 af7d91d
add option to have plot pop up (default is True)
