+++ parmoo/viz/plot.py
@@ -64,12 +64,12 @@ import pandas as pd
 
 def scatter(moop,
             db='pf',
-            export='none',
-            display=True,
+            output='dash',
             height='auto',
             width='auto',
             verbose=True,
-            hot_reload=True,): # TODO
+            hot_reload=True,
+            pop_up=True,):  # TODO
     """ Display MOOP results as matrix of 2D scatterplots.
 
     Create an interactive plot that displays in the browser.
@@ -83,14 +83,15 @@ def scatter(moop,
         db (String): Choose database to plot
                      'pf' (default) plot Pareto Front
                      'obj' plot objective data
-        export (String): Export plot to working directory.
-                     'none' (default) don't export image file
-                     'html' export plot as html
-                     'pdf' export plot as pdf
-                     'svg' export plot as svg
-                     'webp' export plot as webp
-                     'jpeg' export plot as jpeg
-                     'png' export plot as png
+        output (String):
+                     'dash' (default) display plot in dash app
+                     'no_dash' display plot in browser without dash
+                     'html' export plot as html to working directory
+                     'pdf' export plot as pdf to working directory
+                     'svg' export plot as svg to working directory
+                     'webp' export plot as webp to working directory
+                     'jpeg' export plot as jpeg to working directory
+                     'png' export plot as png to working directory
         browser (boolean): Display interactive plot in browser window.
                     True: (default) display interactive plot in browser window
                     False: don't display interactive plot in browser window
@@ -100,8 +101,7 @@ def scatter(moop,
                     of the plot by using the 'export' keyword, and leave
                     'browser' to True, you will BOTH export an image file to
                     the current working directory AND open an interactive
-                    figure in the browser.
-        x (String)
+                    figure in the browser
 
     Returns:
         None
@@ -157,26 +157,26 @@ def scatter(moop,
         fig.update_traces(diagonal_visible=False)
 
     # * configure plot
-    config = configure(export=export,
+    config = configure(export=output,
                        height=height,
                        width=width,
                        plotName=plotName,)
 
-    # * export plot
-    if export != 'none':
-        exportFile(fig=fig,
-                   plotName=plotName,
-                   fileType=export)
-
-        # * display plot
-    if display is True:
-        # fig.show(config=config)
+    # * output
+    if output == 'dash':
         buildDashApp(moop=moop,
                      db=db,
                      fig=fig,
                      config=config,
                      verbose=verbose,
-                     hot_reload=hot_reload,)
+                     hot_reload=hot_reload,
+                     pop_up=pop_up,)
+    elif output == 'no_dash':
+        fig.show(config=config)
+    else:
+        exportFile(fig=fig,
+                   plotName=plotName,
+                   fileType=output)
 
 
 def scatter3d(moop):
@@ -198,12 +198,12 @@ def scatter3d(moop):
 
 def radar(moop,
           db='pf',
-          export='none',
-          display=True,
+          output='dash',
           height='auto',
           width='auto',
           verbose=True,
-          hot_reload=True,): # TODO
+          hot_reload=True,
+          pop_up=True,):  # TODO
     """ Display MOOP results as radar plot.
 
     Create an interactive plot that displays in the browser.
@@ -261,6 +261,12 @@ def radar(moop,
         message += "Consider using 'pf' or 'obj' instead."
         raise ValueError(message)
 
+    # * create scaled database
+    j = pd.DataFrame(database)
+    for i in j.columns:
+        j[i] = (j[i] - j[i].min()) / (j[i].max() - j[i].min())
[…]
