### 2022-07-18 d00c8d8
restructure dashboard-plot interaction
with this change, dashboard buttons can regenerate plots
previously, the plot controlled the dashboard, leaving limited options
for the dashboard to change the plot. Now, the dashboard controls the
plot, making it possible to put more functionality in dashboard buttons
see Google Drive file below for graphical explanation
drive.google.com/file/d/1mAf0UVzfgo-PaDGlpt27glvIXcDv_uEq/view?usp=sharing
