+++ examples/named_var_ex.py
@@ -1,3 +1,5 @@
+import os
+os.system("rm parmoo.moop")
 
 import numpy as np
 from parmoo import MOOP
@@ -5,8 +7,9 @@ from parmoo.searches import LatinHypercube
 from parmoo.surrogates import GaussRBF
 from parmoo.acquisitions import UniformWeights
 from parmoo.optimizers import LocalGPS
+from parmoo.viz import *
 
-my_moop = MOOP(LocalGPS)
+moop = MOOP(LocalGPS)
 
 # Define a simulation to use below
 def sim_func(x):
@@ -17,32 +20,47 @@ def sim_func(x):
 
 # Add a design variable, simulation, objective, and constraint.
 # Note the 'name' keys for each
-my_moop.addDesign({'name': "MyDes",
+moop.addDesign({'name': "MyDes",
                    'des_type': "continuous",
                    'lb': 0.0, 'ub': 1.0})
-my_moop.addDesign({'name': "MyCat",
+moop.addDesign({'name': "MyCat",
                    'des_type': "categorical",
                    'levels': 2})
 
-my_moop.addSimulation({'name': "MySim",
+moop.addSimulation({'name': "MySim",
                        'm': 2,
                        'sim_func': sim_func,
                        'search': LatinHypercube,
                        'surrogate': GaussRBF,
                        'hyperparams': {'search_budget': 20}})
 
-my_moop.addObjective({'name': "MyObj",
+moop.addObjective({'name': "MyObj",
                       'obj_func': lambda x, s: sum(s["MySim"])})
 
-my_moop.addConstraint({'name': "MyCon",
+moop.addConstraint({'name': "MyCon",
                        'constraint': lambda x, s: 0.1 - x["MyDes"]})
 
-# Extract numpy dtypes for all of this MOOP's inputs/outputs
-des_dtype = my_moop.getDesignType()
-obj_dtype = my_moop.getObjectiveType()
-sim_dtype = my_moop.getSimulationType()
+# # Extract numpy dtypes for all of this MOOP's inputs/outputs
+# des_dtype = moop.getDesignType()
+# obj_dtype = moop.getObjectiveType()
+# sim_dtype = moop.getSimulationType()
+#
+# # Display the dtypes as strings
+# print("Design variable type:   " + str(des_dtype))
+# print("Simulation output type: " + str(sim_dtype))
+# print("Objective type:         " + str(obj_dtype))
+#
+# # display Objectives
+# printObjectives(moop)
+#
+# # display PF
+# printPF(moop)
 
-# Display the dtypes as strings
-print("Design variable type:   " + str(des_dtype))
-print("Simulation output type: " + str(sim_dtype))
-print("Objective type:         " + str(obj_dtype))
+# # Get and print full simulation database
+# sim_db = moop.getSimulationData()
+# print("Simulation data:")
+# print(sim_db)
+
+# Display solution
+vizTest(moop)
+
