+++ examples/quickstart.py
@@ -1,43 +1,46 @@
 
+import os
+os.system("rm parmoo.moop")
+
 import numpy as np
 from parmoo import MOOP
 from parmoo.searches import LatinHypercube
 from parmoo.surrogates import GaussRBF
 from parmoo.acquisitions import UniformWeights
 from parmoo.optimizers import LocalGPS
+from parmoo.viz import *
 
-my_moop = MOOP(LocalGPS)
+moop = MOOP(LocalGPS)
 
-my_moop.addDesign({'name': "x1",
+moop.addDesign({'name': "A",
                    'des_type': "continuous",
                    'lb': 0.0, 'ub': 1.0})
-my_moop.addDesign({'name': "x2", 'des_type': "categorical",
+moop.addDesign({'name': "B", 'des_type': "categorical",
                    'levels': 3})
 
 def sim_func(x):
-   if x["x2"] == 0:
-      return np.array([(x["x1"] - 0.2) ** 2, (x["x1"] - 0.8) ** 2])
+   if x["B"] == 0:
+      return np.array([(x["A"] - 0.2) ** 2, (x["A"] - 0.8) ** 2])
    else:
       return np.array([99.9, 99.9])
 
-my_moop.addSimulation({'name': "MySim",
+moop.addSimulation({'name': "MySim",
                        'm': 2,
                        'sim_func': sim_func,
                        'search': LatinHypercube,
                        'surrogate': GaussRBF,
                        'hyperparams': {'search_budget': 20}})
 
-my_moop.addObjective({'name': "f1", 'obj_func': lambda x, s: s["MySim"][0]})
-my_moop.addObjective({'name': "f2", 'obj_func': lambda x, s: s["MySim"][1]})
+moop.addObjective({'name': "C", 'obj_func': lambda x, s: s["MySim"][0]})
+moop.addObjective({'name': "D", 'obj_func': lambda x, s: s["MySim"][1]})
 
-my_moop.addConstraint({'name': "c1", 'constraint': lambda x, s: 0.1 - x["x1"]})
+moop.addConstraint({'name': "E", 'constraint': lambda x, s: 0.1 - x["A"]})
 
 for i in range(3):
-   my_moop.addAcquisition({'acquisition': UniformWeights,
+   moop.addAcquisition({'acquisition': UniformWeights,
                            'hyperparams': {}})
 
-my_moop.solve(5)
-results = my_moop.getPF()
+moop.solve(5)
 
 # Display solution
-print(results, "\n dtype=" + str(results.dtype))
+vizTest(moop)
