### 2022-07-18 2b32240
restructure dashboard-plot interaction with this change, dashboard buttons can regenerate plots previously, the plot controlled the dashboard, leaving limited options for the dashboard to change the plot. Now, the dashboard controls the plot, making it possible to regenerate graphs through dashboard buttons. for a graphical explanation of the change, see the Google Drive file drive.google.com/file/d/1mAf0UVzfgo-PaDGlpt27glvIXcDv_uEq/
