### 2022-07-19 1f16795
Revert "Merge branch 'feature/viz' of https://github.com/parmoo/parmoo into feature/viz"
This reverts commit fc57a6ce7b6d0278d77b4d611701a38998e9e4f7, reversing
changes made to 2b322404ec28ab3f84c857d57e0e89d2896523fc.
