+++ examples/advanced_ex.py
@@ -1,3 +1,5 @@
+import os
+os.system("rm parmoo.moop")
 
 import numpy as np
 from parmoo import MOOP
@@ -5,19 +7,20 @@ from parmoo.acquisitions import UniformWeights, FixedWeights
 from parmoo.searches import LatinHypercube
 from parmoo.surrogates import GaussRBF
 from parmoo.optimizers import LBFGSB
+from parmoo.viz import *
 
 # Create a new MOOP with a derivative-based solver
-my_moop = MOOP(LBFGSB, hyperparams={})
+moop = MOOP(LBFGSB, hyperparams={})
 
 # Add 3 continuous variables named x1, x2, x3
 for i in range(3):
-    my_moop.addDesign({'name': "x" + str(i+1),
+    moop.addDesign({'name': "x" + str(i+1),
                        'des_type': "continuous",
                        'lb': 0.0,
                        'ub': 1.0,
                        'des_tol': 1.0e-8})
 # Add one categorical variable named x4
-my_moop.addDesign({'name': "x4",
+moop.addDesign({'name': "x4",
                    'des_type': "categorical",
                    'levels': 3})
 
@@ -37,7 +40,7 @@ def quad_sim(x):
 
 # Add the quadratic simulation to the problem
 # Use a 10 point LH search for ex design and a Gaussian RBF surrogate model
-my_moop.addSimulation({'name': "f_conv",
+moop.addSimulation({'name': "f_conv",
                        'm': 2,
                        'sim_func': quad_sim,
                        'search': LatinHypercube,
@@ -69,9 +72,9 @@ def obj_f2(x, sim, der=0):
         return result
 
 # Minimize each of the 2 outputs from the quadratic simulation
-my_moop.addObjective({'name': "f1",
+moop.addObjective({'name': "f1",
                       'obj_func': obj_f1})
-my_moop.addObjective({'name': "f2",
+moop.addObjective({'name': "f2",
                       'obj_func': obj_f2})
 
 def const_x4(x, sim, der=0):
@@ -86,17 +89,17 @@ def const_x4(x, sim, der=0):
         return np.zeros(1, dtype=sim.dtype)[0]
 
 # Add the single constraint to the problem
-my_moop.addConstraint({'name': "c_x4",
+moop.addConstraint({'name': "c_x4",
                        'constraint': const_x4})
 
 # Add 2 different acquisition functions to the problem
-my_moop.addAcquisition({'acquisition': UniformWeights})
-my_moop.addAcquisition({'acquisition': FixedWeights,
+moop.addAcquisition({'acquisition': UniformWeights})
+moop.addAcquisition({'acquisition': FixedWeights,
                         # Fixed weight with equal weight on both objectives
                         'hyperparams': {'weights': np.array([0.5, 0.5])}})
 
 # Turn on checkpointing -- creates the files parmoo.moop and parmoo.surrogate.1
-my_moop.setCheckpoint(True, checkpoint_data=False, filename="parmoo")
+moop.setCheckpoint(True, checkpoint_data=False, filename="parmoo")
 
 # Turn on logging
 import logging
@@ -105,15 +108,18 @@ logging.basicConfig(level=logging.INFO,
                     datefmt='%Y-%m-%d %H:%M:%S')
 
 # Solve the problem
-my_moop.solve(5)
+moop.solve(5)
 
-# Get and print full simulation database
-sim_db = my_moop.getSimulationData()
-print("Simulation data:")
-print(sim_db)
+# # Get and print full simulation database
+# sim_db = moop.getSimulationData()
+# print("Simulation data:")
+# print(sim_db)
 
-# Get and print results
-soln = my_moop.getPF()
-print("\n\n")
[…]
