+++ examples/parmoo_cfr_reload_viz_3obj.py
@@ -1,0 +1,308 @@
+""" Reload ParMOO experiment data from CFR/MDML log-files. """
+
+INIT_FILE  = 'sample-config.json'
+RELD_FILE  = 'reformatted_data_edit.json'
+INIT_TOPIC  = 'mdml-CFR-experiment-01-definition'
+SEND_TOPIC  = 'mdml-CFR-experiment-01-EXP'
+RECV_TOPIC  = 'mdml-CFR-experiment-01-DATA'
+STOP_SIGNAL = 'mdml-CFR-experiment-01-stop'
+GROUP_ID    = 'mdml-CFR-experiment-01-tyler-g1'
+
+RELOAD = True
+RESTART = False
+
+def parmoo_funcx(initializer=None, reloader=None):
+
+    import time
+    import numpy as np
+    from parmoo import MOOP
+    from parmoo.optimizers import LBFGSB
+    from parmoo.surrogates import GaussRBF
+    from parmoo.searches import LatinHypercube
+    from parmoo.acquisitions import RandomConstraint, FixedWeights
+    from parmoo import viz
+
+    # Check whether we wil be reloading, restarting, or both
+    RELOAD = True
+    RESTART = True
+    if initializer is None and reloader is None:
+        raise ValueError("Must specify either initializer, reloader, or both")
+    else:
+        if initializer is None:
+            RESTART = False
+        if reloader is None:
+            RELOAD = False
+
+    # Helper function makes objectives
+    def make_objective(goal, i):
+
+        if goal.lower() == 'max':
+            return lambda x, s, der=0: -s["cfr_out"][i] if der == 0 \
+                                       else (np.zeros(1, x.dtype)[0]
+                                             if der == 1 else
+                                             np.array([(-np.eye(2)[i], )],
+                                                       dtype=[('cfr_out',
+                                                               float, 2)]))
+        elif goal.lower() == 'min':
+            return lambda x, s, der=0: s['cfr_out'][i] if der == 0 \
+                                       else (np.zeros(1, x.dtype)[0]
+                                             if der == 1 else
+                                             np.array([(np.eye(2)[i], )],
+                                                       dtype=[('cfr_out',
+                                                               float, 2)]))
+        else:
+            raise(ValueError("Error in msg['value']['Dependent_Par']: ",
+                             "Goal must be `min` or `max`, not `",
+                             str(goal), "`..."))
+
+    def min_RT(x, sx, der=0):
+        if der==0:
+            return x["RT"]
+        elif der==1:
+            ret_val = np.zeros(1, dtype=x.dtype)[0]
+            ret_val["RT"] = 1.0
+            return ret_val
+        else:
+            return np.zeros(1, dtype=sx.dtype)[0]
+
+    # Init function
+    def read_config(initializer, database=None):
+        lb = []
+        ub = []
+        lvls = []
+        n = 0
+        o = 0
+        des_names = []
+        cont_names = []
+        cat_names = []
+        obj_names = []
+        obj_funcs = []
+        for msg in initializer:
+            print(f"Received config data: {msg['value']}")
+            for name in msg['value']['Continuous_Par'].keys():
+                n += 1
+                des_names.append(name)
+                cont_names.append(name)
+                lb.append(float(msg['value']['Continuous_Par'][name]['Low']))
+                ub.append(float(msg['value']['Continuous_Par'][name]['High']))
+            for name in msg['value']['Categorical_Par'].keys():
+                n += 1
+                des_names.append(name)
+                cat_names.append(name)
+                lvls.append(msg['value']['Categorical_Par'][name]['Levels'])
+            for i, name in enumerate(msg['value']['Dependent_Par'].keys()):
+                o += 1
+                obj_names.append(name)
[…]

[…]
