### 2022-07-19 5700364
Revert "Revert "restructure dashboard-plot interaction""
This reverts commit 4e028046cfca43d7847c4a1d137887cdd31a4947.
