### 2022-07-22 82b913d
fix issue where radar hover labels show bad values
