### 2022-07-22 5f1849e
remove radar tick labels
since radar plots axes are normalized, tick label values are meaningless
