### 2022-07-22 70d66ad
fixed small broken things, reformatting
renamed "settings.py" to "utilities.py"
