+++ parmoo/viz/__init__.py
@@ -1,4 +1,10 @@
-from .plot import *
-from .dashboard import *
-from .graph import *
-from .settings import *
+from .plot import (scatter,
+                   parallel_coordinates,
+                   radar,)
+from .dashboard import buildDashApp
+from .graph import (generate_scatter,
+                    generate_parallel,
+                    generate_radar,)
+from .utilities import (exportFile,
+                        configure,
+                        setPlotName,)
