### 2022-07-22 8cce881
fix static export bug
