### 2022-07-25 b913143
clean code, add helper functions
add helper functions:
 - setDatabase
 - setHoverInfo
 - setPlotName

extensively clean up graph.py code

port documentation to user-facing functions at plot.py
