+++ parmoo/viz/graph.py
@@ -1,65 +1,23 @@
-""" This module contains a library of common plotting functions.
+""" This module contains tools to generate interactive Plotly-based graphs.
 
 The functions are:
 
  Display data in interactive browser plot
-  * ``scatter(moop)`` -- Plot MOOP results as matrix of 2D scatterplots
-  * ``scatter3d(moop)`` -- Plot MOOP results as matrix of 3D scatterplots
-  * ``radar(moop)`` -- Plot MOOP results as radar
-  * ``parallel_coordinates(moop)`` -- Plot MOOP results as parallel coordinates
-  * ``heatmap(moop)`` -- Plot MOOP results as heatmap
-  * ``petal(moop)`` -- Plot MOOP results as petal diagram
-  * ``radviz(moop)`` -- Plot MOOP results as RadViz
-  * ``star_coordinates(moop)`` -- Plot MOOP results as star coordinates
-
-For all interactve browser plot functions, there is a known issue
-where Plotly images may not export from Safari correctly. If you
-encounter this issue, change your default browser to Chrome, Firefox,
-or Edge.
+  * ``generate_scatter(moop)`` -- Generate interactive scatterplot matrix
+  * ``generate_parallel(moop)`` -- Generate interactive parallel plot
+  * ``generate_radar(moop)`` -- Generate interactive radar plot
 
 """
 
-import plotly.express as px         # 15.2 MB package
+import plotly.express as px
 import plotly.graph_objects as go
-# import plotly.io as pio
-# from parmoo import MOOP
-# import numpy as np
-from warnings import warn                    # native python package
-# from .settings import configure, exportFile
-import pandas as pd
-
-# des_type = moop.getDesignType()
-# obj_type = moop.getObjectiveType()
-# sim_type = moop.getSimulationType()
-# const_type = moop.getConstraintType()
-# pf = moop.getPF()
-# obj_db = moop.getObjectiveData()
-# sim_db = moop.getSimulationData()
+from warnings import warn
+from .utilities import setPlotName, setDatabase, setHoverInfo
 
 
 #
 # ! THESE FUNCTIONS DISPLAY DATA IN AN INTERACTIVE BROWSER PLOT
 #
-
-# def printDataTypes(moop):
-#     """ Print data types to terminal.
-
-#     Args:
-#         moop (MOOP): A ParMOO MOOP containing the data types to print.
-
-#     Returns:
-#         None
-
-#     """
-#     des_type = moop.getDesignType()
-#     obj_type = moop.getObjectiveType()
-#     sim_type = moop.getSimulationType()
-#     const_type = moop.getConstraintType()
-#     print("\nDATA TYPES FOR YOUR MOOP:\n")
-#     print("Design variable type:   " + str(des_type))
-#     print("Simulation output type: " + str(sim_type))
-#     print("Objective type:         " + str(obj_type))
-#     print("Constraint type:        " + str(const_type))
 
 
 def generate_scatter(moop,
@@ -67,92 +25,27 @@ def generate_scatter(moop,
                      height,
                      width,
                      verbose,):
-    """ Display MOOP results as matrix of 2D scatterplots.
-
-    Create an interactive plot that displays in the browser.
-
-    For ``n`` objectives, generate an ``n x n`` matrix of 2D scatterplots.
-
-    Users can hover above an output point to see input information pop up.
-
-    Args:
-        moop (MOOP): A ParMOO MOOP containing the MOOP results to plot.
-        db (String): Choose database to plot
-                     'pf' (default) plot Pareto Front
-                     'obj' plot objective data
-        output (String):
-                     'dash' (default) display plot in dash app
-                     'no_dash' display plot in browser without dash
-                     'html' export plot as html to working directory
-                     'pdf' export plot as pdf to working directory
[…]
