+++ parmoo/viz/__init__.py
@@ -1,10 +1,10 @@
 from .plot import (scatter,
                    parallel_coordinates,
                    radar,)
-from .dashboard import buildDashApp
+from .dashboard import build_dash_app
 from .graph import (generate_scatter,
                     generate_parallel,
                     generate_radar,)
-from .utilities import (exportFile,
+from .utilities import (export_file,
                         configure,
-                        setPlotName,)
+                        set_plot_name,)
