+++ parmoo/viz/utilities.py
@@ -1,14 +1,14 @@
 import pandas as pd
 
 
-def exportFile(fig, plotName, fileType):
+def export_file(fig, plot_name, file_type):
     """ Display MOOP plot.
 
     Export plot.
 
     Args:
         fig (Plotly figure): figure to export.
-        plotName (String): Used for naming export files
+        plot_name (String): Used for naming export files
         export (String): Indicate export type
                      'none' (default) don't export image file
                      'html' export plot as html
@@ -22,20 +22,20 @@ def exportFile(fig, plotName, fileType):
         None
 
     """
-    if fileType == 'html':
-        fig.write_html(plotName + ".html")
-    elif fileType == 'pdf':
-        fig.write_image(plotName + ".pdf")
-    elif fileType == 'svg':
-        fig.write_image(plotName + ".svg")
-    elif fileType == 'webp':
-        fig.write_image(plotName + ".webp")
-    elif fileType == 'jpeg':
-        fig.write_image(plotName + ".jpeg")
-    elif fileType == 'png':
-        fig.write_image(plotName + ".png")
+    if file_type == 'html':
+        fig.write_html(plot_name + ".html")
+    elif file_type == 'pdf':
+        fig.write_image(plot_name + ".pdf")
+    elif file_type == 'svg':
+        fig.write_image(plot_name + ".svg")
+    elif file_type == 'webp':
+        fig.write_image(plot_name + ".webp")
+    elif file_type == 'jpeg':
+        fig.write_image(plot_name + ".jpeg")
+    elif file_type == 'png':
+        fig.write_image(plot_name + ".png")
     else:
-        message = "ParMOO does not support outputting to '" + fileType + "'.\n"
+        message = "ParMOO does not support '" + file_type + "'.\n"
         message += "Supported outputs:\n"
         message += "'dash'\n"
         message += "'no_dash'\n"
@@ -48,7 +48,7 @@ def exportFile(fig, plotName, fileType):
         raise ValueError(message)
 
 
-def configure(height, width, plotName):
+def configure(height, width, plot_name):
 
     # # * set screenshot type based on export type
     # if export == 'png':
@@ -68,7 +68,7 @@ def configure(height, width, plotName):
             'displayModeBar': True,
             'toImageButtonOptions': {
                 'format': screenshot,  # one of png, svg, jpeg, webp
-                'filename': str(plotName),
+                'filename': str(plot_name),
                 'height': int(height),
                 'width': int(width),
                 'scale': 1  # Multiply title/legend/axis/canvas sizes by factor
@@ -80,40 +80,32 @@ def configure(height, width, plotName):
             'displayModeBar': True,
             'toImageButtonOptions': {
                 'format': screenshot,  # one of png, svg, jpeg, webp
-                'filename': str(plotName),
+                'filename': str(plot_name),
             }
         }
     return config
 
-# def output()
-#     if output == 'dash':
-#         buildDashApp(moop=moop,
-#                      db=db,
-#                      fig=fig,
-#                      config=config,
-#                      verbose=verbose,
-#                      hot_reload=hot_reload,
-#                      pop_up=pop_up,
-#                      port=port,)
-#     elif output == 'no_dash':
-#         fig.show(config=config)
-#     else:
-#         exportFile(fig=fig,
-#                    plotName=plotName,
-#                    fileType=output)
+
+def customize(fig,
+              font,):
+    if font != 'auto':
+        fig.update_layout(
+            font_family=font
+        )
+    return fig
 
 
-def setPlotName(db):
+def set_plot_name(db):
     if db == 'pf':
-        plotName = "Pareto Front"
[…]

[…]
