### 2022-07-25 1feb7ad
make font size customizable
add helper function
 - generate_graph
 - customize

supporting multiple inputs that update the same output
(i.e. font dropdown and fontsize input box)
requires the callback_context.triggered multiplexer.
future developers should consult
https://community.plotly.com/t/duplicate-callback-outputs-solution-api-discussion/55909
before adding new GUI features that update the graph.

also, rename lots of functions to keep naming style consistent

redesign how some function inputs look
