+++ parmoo/viz/__init__.py
@@ -1,10 +1,22 @@
-from .plot import (scatter,
-                   parallel_coordinates,
-                   radar,)
-from .dashboard import build_dash_app
-from .graph import (generate_scatter,
-                    generate_parallel,
-                    generate_radar,)
-from .utilities import (export_file,
-                        configure,
-                        set_plot_name,)
+from .plot import (
+    scatter,
+    parallel_coordinates,
+    radar,
+)
+from .dashboard import (
+    build_dash_app,
+    generate_graph,
+)
+from .graph import (
+    generate_scatter,
+    generate_parallel,
+    generate_radar,
+)
+from .utilities import (
+    export_file,
+    customize,
+    configure,
+    set_plot_name,
+    set_database,
+    set_hover_info,
+)

[…]
