### 2022-07-26 9c8ec17
switched to using class for dash app
this is beneficial because
- dashboard GUI updates graph elements. it makes sense to store state in
dash app so that as buttons are pressed, the output of previous button
presses remains stored.
it is possible to acheive this using functions alone, but it is a hassle
- using classes removes a lot of boilerplate code, passing items from
function to another. Since this viz module has a lot of wrapper
functions with many inputs, this kind of boilerplate code is getting
unweildy

One immediate benefit of this change is that when you change the font
size and the
[…]
