+++ parmoo/viz/dashboard.py
@@ -1,3 +1,7 @@
+import pandas as pd
+from os import environ
+from webbrowser import open_new
+from warnings import warn
 from dash import (
     Dash,
     callback_context,
@@ -7,10 +11,6 @@ from dash import (
     dcc,
     exceptions,
 )
-import pandas as pd
-from os import environ
-from webbrowser import open_new
-from warnings import warn
 from .graph import (
     generate_scatter,
     generate_parallel,
@@ -36,304 +36,307 @@ from .utilities import (
 # make various kinds of plots and have a consistent functionality
 # wrapper around them
 
+class Dash_App:
 
-def build_dash_app(
-    plot_type,
-    moop,
-    db,
-    height,
-    width,
-    verbose,
-    font,
-    fontsize,
-    background_color,
-    margins,
-    hot_reload,
-    pop_up,
-    port,
-    objectives_only,
-):
-
-    # * define database
-    # (initially, all graph data is selected)
-    database = set_database(moop, db=db)
-    plot_name = set_plot_name(db=db)
-
-    # * create app
-    app = Dash(__name__)
-    selection_indexes = []
-
-    # * create plot
-    graph = generate_graph(
-        plot_type=plot_type,
-        moop=moop,
-        db=db,
-        height=height,
-        width=width,
-        verbose=verbose,
-        font=font,
-        fontsize=fontsize,
-        background_color=background_color,
-        margins=margins,
-        objectives_only=objectives_only,
-    )
-
-    config = configure(
-        height=height,
-        width=width,
-        plot_name=plot_name,
-    )
-
-    # * lay out app
-    app.layout = html.Div(children=[
-        # * parmoo logo
-        # html.Img(
-        #     src='data:image/png;base64,{}'.format(parmoo_logo_encoded)
-        # ),
-        # * main plot
-        dcc.Graph(
-            id='parmoo_graph',
-            figure=graph,
-            config=config,
-        ),
-        dcc.Store(
-            id='selection',
-        ),
-        # * download dataset button
-        html.Button(
-            children='Download dataset as CSV',
-            id='download_dataset_button',
-        ),
-        dcc.Download(
-            id='dataset_download_csv',
-        ),
-        html.Br(),
-        html.Br(),
-        # * download selection button
-        html.Button(
-            children='Download selection as CSV',
-            id='download_selection_button',
-        ),
-        dcc.Download(
-            id='selection_download_csv',
-        ),
-        html.Br(),
-        html.Br(),
-        dcc.Dropdown(
-            ['Open Sans',
-             'Times New Roman',
-             'Verdana',
-             'Arial',
-             'Calibri'],
-            placeholder='Select font',
-            id='font_selection_downdown',
-        ),
-        html.Br(),
-        html.Br(),
-        dcc.Input(
-            placeholder='Select font size',
-            type='number',
-            value='',
-            id='font_size_dropdown',
-        )
-    ])
-
-    # * functionality of dataset download button
-    @app.callback(
-        Output(
-            component_id='dataset_download_csv',
-            component_property='data'),
-        Input(
-            component_id='download_dataset_button',
-            component_property='n_clicks'),
-    )
-    def download_dataset(n_clicks):
-        if n_clicks is None:
-            raise exceptions.PreventUpdate
-        else:
-            database.index.name = 'index'
-            return dict(
-                filename=str(plot_name) + ".csv",
-                content=database.to_csv(),
-            )
-
-    # * create object holding selected data
-    @app.callback(
-        Output(
-            component_id='selection',
-            component_property='data'),
-        Input(
-            component_id='parmoo_graph',
-            component_property='selectedData'),
-    )
-    def store_selection(selectedData):
-        if selectedData is None:
[…]
