+++ parmoo/viz/plot.py
@@ -1,4 +1,6 @@
-from .dashboard import build_dash_app
+from .dashboard import (
+    Dash_App
+)
 from .utilities import (
     export_file,
     set_plot_name,
@@ -59,31 +61,45 @@ def scatter(
     moop,
     db='pf',
     output='dash',
+    axes='objectives',
+    specificaxes='auto',
     height='auto',
     width='auto',
-    verbose=True,
     font='auto',
     fontsize='auto',
     background_color='auto',
     margins='auto',
-    objectives_only=True,
+    dummy1='auto',
+    dummy2='auto',
+    dummy3='auto',
+    dummy4='auto',
+    dummy5='auto',
+    dummy6='auto',
+    verbose=True,
     hot_reload=False,
     pop_up=True,
     port='http://127.0.0.1:8050/',
 ):
     if output == 'dash':
-        build_dash_app(
+        Dash_App(
             plot_type='scatter',
             moop=moop,
             db=db,
-            height='auto',
-            width='auto',
-            verbose=verbose,
+            axes=axes,
+            specificaxes=specificaxes,
+            height=height,
+            width=width,
             font=font,
             fontsize=fontsize,
             background_color=background_color,
             margins=margins,
-            objectives_only=objectives_only,
+            dummy1=dummy1,
+            dummy2=dummy2,
+            dummy3=dummy3,
+            dummy4=dummy4,
+            dummy5=dummy5,
+            dummy6=dummy6,
+            verbose=verbose,
             hot_reload=hot_reload,
             pop_up=pop_up,
             port=port,
@@ -94,14 +110,21 @@ def scatter(
             fig=generate_scatter(
                 moop,
                 db=db,
+                axes=axes,
+                specificaxes=specificaxes,
                 height=height,
                 width=width,
-                verbose=verbose,
                 font=font,
                 fontsize=fontsize,
                 background_color=background_color,
                 margins=margins,
-                objectives_only=objectives_only,
+                dummy1=dummy1,
+                dummy2=dummy2,
+                dummy3=dummy3,
+                dummy4=dummy4,
+                dummy5=dummy5,
+                dummy6=dummy6,
+                verbose=verbose,
             ),
             plot_name=plot_name,
             file_type=output,
@@ -152,32 +175,46 @@ def parallel_coordinates(
     moop,
     db='pf',
     output='dash',
+    axes='objectives',
+    specificaxes='auto',
     height='auto',
     width='auto',
-    verbose=True,
     font='auto',
     fontsize='auto',
     background_color='auto',
     margins='auto',
-    objectives_only=True,
+    dummy1='auto',
+    dummy2='auto',
+    dummy3='auto',
+    dummy4='auto',
+    dummy5='auto',
+    dummy6='auto',
+    verbose=True,
     hot_reload=False,
     pop_up=True,
     port='http://127.0.0.1:8050/',
 ):
 
     if output == 'dash':
-        build_dash_app(
+        Dash_App(
             plot_type='parallel',
             moop=moop,
             db=db,
+            axes=axes,
+            specificaxes=specificaxes,
             height=height,
             width=width,
-            verbose=verbose,
             font=font,
             fontsize=fontsize,
             background_color=background_color,
             margins=margins,
-            objectives_only=objectives_only,
+            dummy1=dummy1,
+            dummy2=dummy2,
+            dummy3=dummy3,
+            dummy4=dummy4,
+            dummy5=dummy5,
+            dummy6=dummy6,
+            verbose=verbose,
             hot_reload=hot_reload,
             pop_up=pop_up,
             port=port,
@@ -188,14 +225,21 @@ def parallel_coordinates(
             fig=generate_parallel(
                 moop=moop,
                 db=db,
+                axes=axes,
+                specificaxes=specificaxes,
[…]
