### 2022-07-26 3e50813
move some update functionality to dashboard from utilities
