+++ parmoo/viz/graph.py
@@ -31,7 +31,7 @@ def generate_scatter(
     fontsize,
     background_color,
     margins,
-    dummy1,
+    screenshot,
     dummy2,
     dummy3,
     dummy4,
@@ -89,7 +89,7 @@ def generate_parallel(
     fontsize,
     background_color,
     margins,
-    dummy1,
+    screenshot,
     dummy2,
     dummy3,
     dummy4,
@@ -147,7 +147,7 @@ def generate_radar(
     fontsize,
     background_color,
     margins,
-    dummy1,
+    screenshot,
     dummy2,
     dummy3,
     dummy4,
