+++ parmoo/viz/plot.py
@@ -69,7 +69,7 @@ def scatter(
     fontsize='auto',
     background_color='auto',
     margins='auto',
-    dummy1='auto',
+    screenshot='auto',
     dummy2='auto',
     dummy3='auto',
     dummy4='auto',
@@ -93,7 +93,7 @@ def scatter(
             fontsize=fontsize,
             background_color=background_color,
             margins=margins,
-            dummy1=dummy1,
+            screenshot=screenshot,
             dummy2=dummy2,
             dummy3=dummy3,
             dummy4=dummy4,
@@ -118,7 +118,7 @@ def scatter(
                 fontsize=fontsize,
                 background_color=background_color,
                 margins=margins,
-                dummy1=dummy1,
+                screenshot=screenshot,
                 dummy2=dummy2,
                 dummy3=dummy3,
                 dummy4=dummy4,
@@ -183,7 +183,7 @@ def parallel_coordinates(
     fontsize='auto',
     background_color='auto',
     margins='auto',
-    dummy1='auto',
+    screenshot='auto',
     dummy2='auto',
     dummy3='auto',
     dummy4='auto',
@@ -208,7 +208,7 @@ def parallel_coordinates(
             fontsize=fontsize,
             background_color=background_color,
             margins=margins,
-            dummy1=dummy1,
+            screenshot=screenshot,
             dummy2=dummy2,
             dummy3=dummy3,
             dummy4=dummy4,
@@ -233,7 +233,7 @@ def parallel_coordinates(
                 fontsize=fontsize,
                 background_color=background_color,
                 margins=margins,
-                dummy1=dummy1,
+                screenshot=screenshot,
                 dummy2=dummy2,
                 dummy3=dummy3,
                 dummy4=dummy4,
@@ -292,7 +292,7 @@ def radar(
     fontsize='auto',
     background_color='auto',
     margins='auto',
-    dummy1='auto',
+    screenshot='auto',
     dummy2='auto',
     dummy3='auto',
     dummy4='auto',
@@ -317,7 +317,7 @@ def radar(
             fontsize=fontsize,
             background_color=background_color,
             margins=margins,
-            dummy1=dummy1,
+            screenshot=screenshot,
             dummy2=dummy2,
             dummy3=dummy3,
             dummy4=dummy4,
@@ -342,7 +342,7 @@ def radar(
                 fontsize=fontsize,
                 background_color=background_color,
                 margins=margins,
-                dummy1=dummy1,
+                screenshot=screenshot,
                 dummy2=dummy2,
                 dummy3=dummy3,
                 dummy4=dummy4,
