### 2022-07-26 ac71e62
add working GUI dropdown to choose plot type
put lots of buttons on the screen (most not organized or working yet)
