+++ parmoo/viz/plot.py
@@ -69,7 +69,7 @@ def scatter(
     fontsize='auto',
     background_color='auto',
     margins='auto',
-    screenshot='auto',
+    screenshot='svg',
     dummy2='auto',
     dummy3='auto',
     dummy4='auto',
@@ -183,7 +183,7 @@ def parallel_coordinates(
     fontsize='auto',
     background_color='auto',
     margins='auto',
-    screenshot='auto',
+    screenshot='svg',
     dummy2='auto',
     dummy3='auto',
     dummy4='auto',
@@ -292,7 +292,7 @@ def radar(
     fontsize='auto',
     background_color='auto',
     margins='auto',
-    screenshot='auto',
+    screenshot='svg',
     dummy2='auto',
     dummy3='auto',
     dummy4='auto',
