+++ parmoo/viz/dashboard.py
@@ -17,7 +17,6 @@ from .graph import (
     generate_radar,
 )
 from .utilities import (
-    configure,
     set_plot_name,
     set_database
 )
@@ -90,11 +89,7 @@ class Dash_App:
         self.plot_name = set_plot_name(db=self.db)
         self.database = set_database(moop, db=self.db)
         self.graph = self.generate_graph()
-        self.config = configure(
-            height=self.height,
-            width=self.width,
-            plot_name=self.plot_name,
-        )
+        self.config = self.configure()
 
         # * initialize app
         app = Dash(__name__)
@@ -131,24 +126,75 @@ class Dash_App:
             dcc.Download(
                 id='selection_download_csv',
             ),
-            html.Br(),
-            html.Br(),
-            dcc.Dropdown(
-                ['Open Sans',
-                 'Times New Roman',
-                 'Verdana',
-                 'Arial',
-                 'Calibri'],
+            dcc.Input(
                 placeholder='Select font',
-                id='font_selection_downdown',
+                type='text',
+                value='',
+                id='font_selection_input',
             ),
-            html.Br(),
-            html.Br(),
             dcc.Input(
                 placeholder='Select font size',
                 type='number',
                 value='',
-                id='font_size_dropdown',
+                id='font_size_input',
+            ),
+            dcc.Input(
+                placeholder='Select graph width',
+                type='number',
+                value='',
+                id='graph_width_input',
+            ),
+            dcc.Input(
+                placeholder='Select graph height',
+                type='number',
+                value='',
+                id='graph_height_input',
+            ),
+            dcc.Input(
+                placeholder='Select margin size',
+                type='number',
+                value='',
+                id='graph_margins_input',
+            ),
+            dcc.Input(
+                placeholder='Select graph title',
+                type='text',
+                value='',
+                id='plot_name_input',
+            ),
+            dcc.Dropdown(
+                options=['svg',
+                         'png',
+                         'jpeg',
+                         'webp'],
+                placeholder='Select screenshot format',
+                id='screenshot_dropdown',
+            ),
+            dcc.Dropdown(
+                options=['Scatterplot',
+                         'Parallel Coordinates plot',
+                         'Radar plot'],
+                placeholder='Select plot type',
+                id='plot_type_dropdown',
+            ),
+            dcc.Dropdown(
+                options=['Pareto Front',
+                         'Objective data'],
+                placeholder='Select database',
+                id='database_dropdown',
+            ),
+            dcc.Dropdown(
+                options=['White',
+                         'Grey',
+                         'Black',
+                         'Red',
+                         'Yellow',
+                         'Blue',
+                         'Green',
+                         'Orange',
+                         'Purple'],
+                placeholder='Select background color',
+                id='background_color_dropdown',
             )
         ])
 
@@ -217,32 +263,117 @@ class Dash_App:
                     content=selection_db.to_csv(),
                 )
 
-        # * customize graph
+        # * configure graph
+        # TODO whenever possible, configure instead of updating or regenerating
+        @app.callback(
+            Output(
+                component_id='parmoo_graph',
+                component_property='config',),
+            # height
+            Input(
+                component_id='graph_height_input',
+                component_property='value',),
+            # width
+            Input(
+                component_id='graph_width_input',
+                component_property='value',),
+            # screenshot
+            Input(
+                component_id='screenshot_dropdown',
+                component_property='value',),
+            # title
+            Input(
+                component_id='plot_name_input',
+                component_property='value',),
+            prevent_initial_call=True
+        )
+        def update_config(
+            height_value,
+            width_value,
+            screenshot_value,
+            plot_name_value,
+        ):
[…]
