### 2022-07-26 cfb3e83
make font & fontsize input-safe
fixed various bugs, as best I can tell, now, users can't break the
dashboard through the font and fontsize inputs
