### 2022-07-26 a0aaeec
GUI button to change db now working
tested changing the plot type and database against each other; no issues
