### 2022-07-26 4b8a604
add functionality to change plot name via GUI
tested - nothing breaks when combined with interactions from the
other working buttons

Currently working buttons:
- choose font
- choose font size
- choose plot title
- choose database plotted
- choose plot type
