### 2022-07-27 cc7b46e
add GUI option to customize background color
