### 2022-07-27 c9e3f1a
add paper and graph background color options
known bugs:
 - title doesn't update when changing database
 - parallel plot may crash if MOOP has one and only one constraint

 recommended solutions:
 - code title to update when and only when the database changes
 AND the current title is one of the default titles
 - code parallel plot to be single-constraint safe via copypasta of
 generate_radar()'s `setup axes` code

known feature/bug:
- parallel plot only updates when updating paper color variable
this is because parallel plot doesn't have axis grid lines with a
special graph area between them.
[…]
