+++ parmoo/viz/plot.py
@@ -67,10 +67,10 @@ def scatter(
     width='auto',
     font='auto',
     fontsize='auto',
-    background_color='auto',
+    paper_background_color='auto',
     margins='auto',
     screenshot='svg',
-    dummy2='auto',
+    graph_background_color='auto',
     dummy3='auto',
     dummy4='auto',
     dummy5='auto',
@@ -91,10 +91,10 @@ def scatter(
             width=width,
             font=font,
             fontsize=fontsize,
-            background_color=background_color,
+            paper_background_color=paper_background_color,
             margins=margins,
             screenshot=screenshot,
-            dummy2=dummy2,
+            graph_background_color=graph_background_color,
             dummy3=dummy3,
             dummy4=dummy4,
             dummy5=dummy5,
@@ -116,10 +116,10 @@ def scatter(
                 width=width,
                 font=font,
                 fontsize=fontsize,
-                background_color=background_color,
+                paper_background_color=paper_background_color,
                 margins=margins,
                 screenshot=screenshot,
-                dummy2=dummy2,
+                graph_background_color=graph_background_color,
                 dummy3=dummy3,
                 dummy4=dummy4,
                 dummy5=dummy5,
@@ -181,10 +181,10 @@ def parallel_coordinates(
     width='auto',
     font='auto',
     fontsize='auto',
-    background_color='auto',
+    paper_background_color='auto',
     margins='auto',
     screenshot='svg',
-    dummy2='auto',
+    graph_background_color='auto',
     dummy3='auto',
     dummy4='auto',
     dummy5='auto',
@@ -206,10 +206,10 @@ def parallel_coordinates(
             width=width,
             font=font,
             fontsize=fontsize,
-            background_color=background_color,
+            paper_background_color=paper_background_color,
             margins=margins,
             screenshot=screenshot,
-            dummy2=dummy2,
+            graph_background_color=graph_background_color,
             dummy3=dummy3,
             dummy4=dummy4,
             dummy5=dummy5,
@@ -231,10 +231,10 @@ def parallel_coordinates(
                 width=width,
                 font=font,
                 fontsize=fontsize,
-                background_color=background_color,
+                paper_background_color=paper_background_color,
                 margins=margins,
                 screenshot=screenshot,
-                dummy2=dummy2,
+                graph_background_color=graph_background_color,
                 dummy3=dummy3,
                 dummy4=dummy4,
                 dummy5=dummy5,
@@ -290,10 +290,10 @@ def radar(
     width='auto',
     font='auto',
     fontsize='auto',
-    background_color='auto',
+    paper_background_color='auto',
     margins='auto',
     screenshot='svg',
-    dummy2='auto',
+    graph_background_color='auto',
     dummy3='auto',
     dummy4='auto',
     dummy5='auto',
@@ -315,10 +315,10 @@ def radar(
             width=width,
             font=font,
             fontsize=fontsize,
-            background_color=background_color,
+            paper_background_color=paper_background_color,
             margins=margins,
             screenshot=screenshot,
-            dummy2=dummy2,
+            graph_background_color=graph_background_color,
             dummy3=dummy3,
             dummy4=dummy4,
             dummy5=dummy5,
@@ -340,10 +340,10 @@ def radar(
                 width=width,
                 font=font,
                 fontsize=fontsize,
-                background_color=background_color,
+                paper_background_color=paper_background_color,
                 margins=margins,
                 screenshot=screenshot,
-                dummy2=dummy2,
+                graph_background_color=graph_background_color,
                 dummy3=dummy3,
                 dummy4=dummy4,
                 dummy5=dummy5,
