+++ parmoo/viz/plot.py
@@ -71,7 +71,7 @@ def scatter(
     margins='auto',
     screenshot='svg',
     graph_background_color='auto',
-    dummy3='auto',
+    theme='auto',
     dummy4='auto',
     dummy5='auto',
     dummy6='auto',
@@ -95,7 +95,7 @@ def scatter(
             margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            dummy3=dummy3,
+            theme=theme,
             dummy4=dummy4,
             dummy5=dummy5,
             dummy6=dummy6,
@@ -120,7 +120,7 @@ def scatter(
                 margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                dummy3=dummy3,
+                theme=theme,
                 dummy4=dummy4,
                 dummy5=dummy5,
                 dummy6=dummy6,
@@ -185,7 +185,7 @@ def parallel_coordinates(
     margins='auto',
     screenshot='svg',
     graph_background_color='auto',
-    dummy3='auto',
+    theme='auto',
     dummy4='auto',
     dummy5='auto',
     dummy6='auto',
@@ -210,7 +210,7 @@ def parallel_coordinates(
             margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            dummy3=dummy3,
+            theme=theme,
             dummy4=dummy4,
             dummy5=dummy5,
             dummy6=dummy6,
@@ -235,7 +235,7 @@ def parallel_coordinates(
                 margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                dummy3=dummy3,
+                theme=theme,
                 dummy4=dummy4,
                 dummy5=dummy5,
                 dummy6=dummy6,
@@ -294,7 +294,7 @@ def radar(
     margins='auto',
     screenshot='svg',
     graph_background_color='auto',
-    dummy3='auto',
+    theme='auto',
     dummy4='auto',
     dummy5='auto',
     dummy6='auto',
@@ -319,7 +319,7 @@ def radar(
             margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            dummy3=dummy3,
+            theme=theme,
             dummy4=dummy4,
             dummy5=dummy5,
             dummy6=dummy6,
@@ -344,7 +344,7 @@ def radar(
                 margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                dummy3=dummy3,
+                theme=theme,
                 dummy4=dummy4,
                 dummy5=dummy5,
                 dummy6=dummy6,
