+++ parmoo/viz/graph.py
@@ -33,7 +33,7 @@ def generate_scatter(
     margins,
     screenshot,
     graph_background_color,
-    dummy3,
+    theme,
     dummy4,
     dummy5,
     dummy6,
@@ -91,7 +91,7 @@ def generate_parallel(
     margins,
     screenshot,
     graph_background_color,
-    dummy3,
+    theme,
     dummy4,
     dummy5,
     dummy6,
@@ -149,7 +149,7 @@ def generate_radar(
     margins,
     screenshot,
     graph_background_color,
-    dummy3,
+    theme,
     dummy4,
     dummy5,
     dummy6,
