### 2022-07-27 29e7c07
fix bug where plot names don't update for new db
