### 2022-07-27 3d67333
add options to customize graph height and width
