### 2022-07-27 b0e3623
clean dashbaord code
Make callback code cleaner and easier to understand by porting
functionality to helper functions. Organize dashboard code into sections
