+++ parmoo/viz/dashboard.py
@@ -53,14 +53,17 @@ class Dash_App:
         screenshot,
         graph_background_color,
         theme,
-        dummy4,
-        dummy5,
+        export_image_format,
+        export_data_format,
         dummy6,
         verbose,
         hot_reload,
         pop_up,
         port,
     ):
+
+    # ! STATE
+
         # * define independent state
         self.moop = moop
         self.plot_type = plot_type
@@ -76,8 +79,8 @@ class Dash_App:
         self.screenshot = screenshot
         self.graph_background_color = graph_background_color
         self.theme = theme
-        self.dummy4 = dummy4
-        self.dummy5 = dummy5
+        self.export_image_format = export_image_format
+        self.export_data_format = export_data_format
         self.dummy6 = dummy6
         self.verbose = verbose
         self.hot_reload = hot_reload
@@ -91,6 +94,8 @@ class Dash_App:
         self.graph = self.generate_graph()
         self.config = self.configure()
 
+    # ! LAYOUT
+
         # * initialize app
         app = Dash(__name__)
         # * lay out app
@@ -101,14 +106,13 @@ class Dash_App:
             # ),
             # * main plot
             dcc.Graph(
+                id='parmoo_graph',
                 figure=self.graph,
                 config=self.config,
-                id='parmoo_graph',
             ),
             dcc.Store(
                 id='selection',
             ),
-            # * download dataset button
             html.Button(
                 children='Download dataset as CSV',
                 id='download_dataset_button',
@@ -116,7 +120,6 @@ class Dash_App:
             dcc.Download(
                 id='dataset_download_csv',
             ),
-            # * download selection button
             html.Button(
                 children='Download selection as CSV',
                 id='download_selection_button',
@@ -125,52 +128,53 @@ class Dash_App:
                 id='selection_download_csv',
             ),
             dcc.Input(
+                id='font_selection_input',
                 placeholder='Select font',
                 type='text',
                 value='',
                 debounce=True,
-                id='font_selection_input',
             ),
             dcc.Input(
+                id='font_size_input',
                 placeholder='Select font size',
                 type='number',
                 value='',
                 min=1,
                 max=100,
                 debounce=True,
-                id='font_size_input',
             ),
             dcc.Input(
+                id='graph_width_input',
                 placeholder='Select graph width',
                 type='number',
                 value='',
                 min=10,
                 step=1,
-                id='graph_width_input',
             ),
             dcc.Input(
+                id='graph_height_input',
                 placeholder='Select graph height',
                 type='number',
                 value='',
                 min=10,
                 step=1,
-                id='graph_height_input',
             ),
             dcc.Input(
+                id='graph_margins_input',
                 placeholder='Select margin size',
                 type='number',
                 value='',
                 disabled=True,
-                id='graph_margins_input',
             ),
             dcc.Input(
+                id='plot_name_input',
                 placeholder='Select plot name',
                 type='text',
                 value='',
                 debounce=True,
-                id='plot_name_input',
             ),
             dcc.Dropdown(
+                id='screenshot_dropdown',
                 options=[
                     'svg',
                     'png',
@@ -179,24 +183,24 @@ class Dash_App:
                 ],
                 placeholder='Select screenshot format',
                 disabled=True,
-                id='screenshot_dropdown',
             ),
             dcc.Dropdown(
+                id='plot_type_dropdown',
                 options=['Scatterplot',
                          'Parallel Coordinates plot',
                          'Radar plot'],
                 placeholder='Select plot type',
-                id='plot_type_dropdown',
             ),
             dcc.Dropdown(
+                id='database_dropdown',
                 options=[
                     'Pareto Front',
                     'Objective Data'
                 ],
                 placeholder='Select database',
-                id='database_dropdown',
             ),
[…]
