+++ parmoo/viz/graph.py
@@ -33,9 +33,8 @@ def generate_scatter(
     margins,
     screenshot,
     graph_background_color,
-    theme,
-    dummy4,
-    dummy5,
+    export_image_format,
+    export_data_format,
     dummy6,
     verbose,
 ):
@@ -91,9 +90,8 @@ def generate_parallel(
     margins,
     screenshot,
     graph_background_color,
-    theme,
-    dummy4,
-    dummy5,
+    export_image_format,
+    export_data_format,
     dummy6,
     verbose,
 ):
@@ -149,9 +147,8 @@ def generate_radar(
     margins,
     screenshot,
     graph_background_color,
-    theme,
-    dummy4,
-    dummy5,
+    export_image_format,
+    export_data_format,
     dummy6,
     verbose,
 ):
