+++ parmoo/viz/plot.py
@@ -71,9 +71,8 @@ def scatter(
     margins='auto',
     screenshot='svg',
     graph_background_color='auto',
-    theme='auto',
-    dummy4='auto',
-    dummy5='auto',
+    export_image_format='auto',
+    export_data_format='auto',
     dummy6='auto',
     verbose=True,
     hot_reload=False,
@@ -95,9 +94,8 @@ def scatter(
             margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            theme=theme,
-            dummy4=dummy4,
-            dummy5=dummy5,
+            export_image_format=export_image_format,
+            export_data_format=export_data_format,
             dummy6=dummy6,
             verbose=verbose,
             hot_reload=hot_reload,
@@ -120,9 +118,8 @@ def scatter(
                 margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                theme=theme,
-                dummy4=dummy4,
-                dummy5=dummy5,
+                export_image_format=export_image_format,
+                export_data_format=export_data_format,
                 dummy6=dummy6,
                 verbose=verbose,
             ),
@@ -185,9 +182,8 @@ def parallel_coordinates(
     margins='auto',
     screenshot='svg',
     graph_background_color='auto',
-    theme='auto',
-    dummy4='auto',
-    dummy5='auto',
+    export_image_format='auto',
+    export_data_format='auto',
     dummy6='auto',
     verbose=True,
     hot_reload=False,
@@ -210,9 +206,8 @@ def parallel_coordinates(
             margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            theme=theme,
-            dummy4=dummy4,
-            dummy5=dummy5,
+            export_image_format=export_image_format,
+            export_data_format=export_data_format,
             dummy6=dummy6,
             verbose=verbose,
             hot_reload=hot_reload,
@@ -235,9 +230,8 @@ def parallel_coordinates(
                 margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                theme=theme,
-                dummy4=dummy4,
-                dummy5=dummy5,
+                export_image_format=export_image_format,
+                export_data_format=export_data_format,
                 dummy6=dummy6,
                 verbose=verbose,
             ),
@@ -294,9 +288,8 @@ def radar(
     margins='auto',
     screenshot='svg',
     graph_background_color='auto',
-    theme='auto',
-    dummy4='auto',
-    dummy5='auto',
+    export_image_format='auto',
+    export_data_format='auto',
     dummy6='auto',
     verbose=True,
     hot_reload=False,
@@ -319,9 +312,8 @@ def radar(
             margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            theme=theme,
-            dummy4=dummy4,
-            dummy5=dummy5,
+            export_image_format=export_image_format,
+            export_data_format=export_data_format,
             dummy6=dummy6,
             verbose=verbose,
             hot_reload=hot_reload,
@@ -344,9 +336,8 @@ def radar(
                 margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                theme=theme,
-                dummy4=dummy4,
-                dummy5=dummy5,
+                export_image_format=export_image_format,
+                export_data_format=export_data_format,
                 dummy6=dummy6,
                 verbose=verbose,
             ),
