+++ parmoo/viz/dashboard.py
@@ -52,7 +52,6 @@ class Dash_App:
         margins,
         screenshot,
         graph_background_color,
-        theme,
         export_image_format,
         export_data_format,
         dummy6,
@@ -78,7 +77,6 @@ class Dash_App:
         self.margins = margins
         self.screenshot = screenshot
         self.graph_background_color = graph_background_color
-        self.theme = theme
         self.export_image_format = export_image_format
         self.export_data_format = export_data_format
         self.dummy6 = dummy6
@@ -114,18 +112,25 @@ class Dash_App:
                 id='selection',
             ),
             html.Button(
-                children='Download dataset as CSV',
+                children='Download dataset',
                 id='download_dataset_button',
             ),
             dcc.Download(
-                id='dataset_download_csv',
+                id='download_dataset_dcc',
             ),
             html.Button(
-                children='Download selection as CSV',
+                children='Download selection',
                 id='download_selection_button',
             ),
             dcc.Download(
-                id='selection_download_csv',
+                id='download_selection_dcc',
+            ),
+            html.Button(
+                children='Download image',
+                id='download_image_button',
+            ),
+            dcc.Download(
+                id='image_download_button',
             ),
             dcc.Input(
                 id='font_selection_input',
@@ -232,23 +237,6 @@ class Dash_App:
                 placeholder='Select graph background color',
             ),
             dcc.Dropdown(
-                id='theme_dropdown',
-                options=[
-                    'ggplot2',
-                    'seaborn',
-                    'simple_white',
-                    'plotly',
-                    'plotly_white',
-                    'plotly_dark',
-                    'presentation',
-                    'xgridoff',
-                    'ygridoff',
-                    'gridon',
-                    'none'
-                ],
-                placeholder='Select theme',
-            ),
-            dcc.Dropdown(
                 id='image_export_format_dropdown',
                 options=[
                     'PNG',
@@ -315,10 +303,6 @@ class Dash_App:
             Input(
                 component_id='graph_background_color_dropdown',
                 component_property='value',),
-            # theme - update
-            Input(
-                component_id='theme_dropdown',
-                component_property='value',),
             # plot name - update
             Input(
                 component_id='plot_name_input',
@@ -341,7 +325,6 @@ class Dash_App:
             margins_value,
             paper_background_color_value,
             graph_background_color_value,
-            theme_value,
             plot_name_value,
             plot_type_value,
             database_value,
@@ -361,8 +344,6 @@ class Dash_App:
                 return self.evaluate_paper_color(paper_background_color_value)
             elif 'graph_background_color_dropdown.value' == triggered_id:
                 return self.evaluate_graph_color(graph_background_color_value)
-            elif 'theme_dropdown.value' == triggered_id:
-                return self.evaluate_theme(theme_value)
             elif 'plot_name_input.value' == triggered_id:
                 return self.evaluate_plot_name(plot_name_value)
             elif 'plot_type_dropdown.value' == triggered_id:
@@ -373,7 +354,7 @@ class Dash_App:
         # * download dataset
         @app.callback(
             Output(
-                component_id='dataset_download_csv',
+                component_id='download_dataset_dcc',
                 component_property='data'),
             Input(
                 component_id='download_dataset_button',
@@ -397,7 +378,7 @@ class Dash_App:
         # * download selection
         @app.callback(
             Output(
-                component_id='selection_download_csv',
+                component_id='download_selection_dcc',
                 component_property='data'),
             Input(
                 component_id='download_selection_button',
@@ -406,6 +387,7 @@ class Dash_App:
         def download_selection(n_clicks):
[…]
