### 2022-08-01 ae090d2
nuke margins
wasn't able to get customizability to work within internship time frame
