+++ parmoo/viz/plot.py
@@ -68,7 +68,6 @@ def scatter(
     font='auto',
     fontsize='auto',
     paper_background_color='auto',
-    margins='auto',
     screenshot='svg',
     graph_background_color='auto',
     export_image_format='auto',
@@ -91,7 +90,6 @@ def scatter(
             font=font,
             fontsize=fontsize,
             paper_background_color=paper_background_color,
-            margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
             export_image_format=export_image_format,
@@ -115,7 +113,6 @@ def scatter(
                 font=font,
                 fontsize=fontsize,
                 paper_background_color=paper_background_color,
-                margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
                 export_image_format=export_image_format,
@@ -179,7 +176,6 @@ def parallel_coordinates(
     font='auto',
     fontsize='auto',
     paper_background_color='auto',
-    margins='auto',
     screenshot='svg',
     graph_background_color='auto',
     export_image_format='auto',
@@ -203,7 +199,6 @@ def parallel_coordinates(
             font=font,
             fontsize=fontsize,
             paper_background_color=paper_background_color,
-            margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
             export_image_format=export_image_format,
@@ -227,7 +222,6 @@ def parallel_coordinates(
                 font=font,
                 fontsize=fontsize,
                 paper_background_color=paper_background_color,
-                margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
                 export_image_format=export_image_format,
@@ -285,7 +279,6 @@ def radar(
     font='auto',
     fontsize='auto',
     paper_background_color='auto',
-    margins='auto',
     screenshot='svg',
     graph_background_color='auto',
     export_image_format='auto',
@@ -309,7 +302,6 @@ def radar(
             font=font,
             fontsize=fontsize,
             paper_background_color=paper_background_color,
-            margins=margins,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
             export_image_format=export_image_format,
@@ -333,7 +325,6 @@ def radar(
                 font=font,
                 fontsize=fontsize,
                 paper_background_color=paper_background_color,
-                margins=margins,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
                 export_image_format=export_image_format,
