### 2022-08-01 263202a
select image export filetype via GUI
exports file to working directory, not browser download
