+++ parmoo/viz/plot.py
@@ -70,8 +70,8 @@ def scatter(
     paper_background_color='auto',
     screenshot='svg',
     graph_background_color='auto',
-    export_image_format='auto',
-    export_data_format='auto',
+    image_export_format='svg',
+    data_export_format='csv',
     dummy6='auto',
     verbose=True,
     hot_reload=False,
@@ -92,8 +92,8 @@ def scatter(
             paper_background_color=paper_background_color,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            export_image_format=export_image_format,
-            export_data_format=export_data_format,
+            image_export_format=image_export_format,
+            data_export_format=data_export_format,
             dummy6=dummy6,
             verbose=verbose,
             hot_reload=hot_reload,
@@ -115,8 +115,8 @@ def scatter(
                 paper_background_color=paper_background_color,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                export_image_format=export_image_format,
-                export_data_format=export_data_format,
+                image_export_format=image_export_format,
+                data_export_format=data_export_format,
                 dummy6=dummy6,
                 verbose=verbose,
             ),
@@ -178,8 +178,8 @@ def parallel_coordinates(
     paper_background_color='auto',
     screenshot='svg',
     graph_background_color='auto',
-    export_image_format='auto',
-    export_data_format='auto',
+    image_export_format='svg',
+    data_export_format='csv',
     dummy6='auto',
     verbose=True,
     hot_reload=False,
@@ -201,8 +201,8 @@ def parallel_coordinates(
             paper_background_color=paper_background_color,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            export_image_format=export_image_format,
-            export_data_format=export_data_format,
+            image_export_format=image_export_format,
+            data_export_format=data_export_format,
             dummy6=dummy6,
             verbose=verbose,
             hot_reload=hot_reload,
@@ -224,8 +224,8 @@ def parallel_coordinates(
                 paper_background_color=paper_background_color,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                export_image_format=export_image_format,
-                export_data_format=export_data_format,
+                image_export_format=image_export_format,
+                data_export_format=data_export_format,
                 dummy6=dummy6,
                 verbose=verbose,
             ),
@@ -281,8 +281,8 @@ def radar(
     paper_background_color='auto',
     screenshot='svg',
     graph_background_color='auto',
-    export_image_format='auto',
-    export_data_format='auto',
+    image_export_format='svg',
+    data_export_format='csv',
     dummy6='auto',
     verbose=True,
     hot_reload=False,
@@ -304,8 +304,8 @@ def radar(
             paper_background_color=paper_background_color,
             screenshot=screenshot,
             graph_background_color=graph_background_color,
-            export_image_format=export_image_format,
-            export_data_format=export_data_format,
+            image_export_format=image_export_format,
+            data_export_format=data_export_format,
             dummy6=dummy6,
             verbose=verbose,
             hot_reload=hot_reload,
@@ -327,8 +327,8 @@ def radar(
                 paper_background_color=paper_background_color,
                 screenshot=screenshot,
                 graph_background_color=graph_background_color,
-                export_image_format=export_image_format,
-                export_data_format=export_data_format,
+                image_export_format=image_export_format,
+                data_export_format=data_export_format,
                 dummy6=dummy6,
                 verbose=verbose,
             ),
