+++ parmoo/viz/dashboard.py
@@ -11,6 +11,7 @@ from dash import (
     html,
     dcc,
     exceptions,
+    no_update,
 )
 from .graph import (
     generate_scatter,
@@ -91,8 +92,6 @@ class Dash_App:
         self.graph = self.generate_graph()
         self.config = self.configure()
 
-        print(image_export_format)
-        print(self.image_export_format)
     # ! LAYOUT
 
         # * initialize app
@@ -103,86 +102,6 @@ class Dash_App:
             # html.Img(
             #     src='data:image/png;base64,{}'.format(parmoo_logo_encoded)
             # ),
-            # * main plot
-            dcc.Graph(
-                id='parmoo_graph',
-                figure=self.graph,
-                config=self.config,
-            ),
-            dcc.Store(
-                id='selection',
-            ),
-            html.Button(
-                children='Download dataset',
-                id='download_dataset_button',
-            ),
-            dcc.Download(
-                id='download_dataset_dcc',
-            ),
-            html.Button(
-                children='Download selection',
-                id='download_selection_button',
-            ),
-            dcc.Download(
-                id='download_selection_dcc',
-            ),
-            html.Button(
-                children='Export image to working directory',
-                id='download_image_button',
-            ),
-            dcc.Store(
-                id='image'
-            ),
-            dcc.Input(
-                id='font_selection_input',
-                placeholder='Select font',
-                type='text',
-                value='',
-                debounce=True,
-            ),
-            dcc.Input(
-                id='font_size_input',
-                placeholder='Select font size',
-                type='number',
-                value='',
-                min=1,
-                max=100,
-                debounce=True,
-            ),
-            dcc.Input(
-                id='graph_width_input',
-                placeholder='Select graph width',
-                type='number',
-                value='',
-                min=10,
-                step=1,
-            ),
-            dcc.Input(
-                id='graph_height_input',
-                placeholder='Select graph height',
-                type='number',
-                value='',
-                min=10,
-                step=1,
-            ),
-            dcc.Input(
-                id='plot_name_input',
-                placeholder='Select plot name',
-                type='text',
-                value='',
-                debounce=True,
-            ),
-            dcc.Dropdown(
-                id='screenshot_dropdown',
-                options=[
-                    'svg',
-                    'png',
-                    'jpeg',
-                    'webp'
-                ],
-                placeholder='Select screenshot format',
-                disabled=True,
-            ),
             dcc.Dropdown(
                 id='plot_type_dropdown',
                 options=['Scatterplot',
@@ -198,6 +117,99 @@ class Dash_App:
                 ],
                 placeholder='Select database',
             ),
+            html.Button(
+                children='Export dataset',
+                id='download_dataset_button',
+            ),
+            dcc.Download(
+                id='download_dataset_dcc',
+            ),
+            html.Button(
+                children='Export selected data',
+                id='download_selection_button',
+            ),
+            dcc.Download(
+                id='download_selection_dcc',
+            ),
+            html.Button(
+                children='Export image to working directory',
+                id='download_image_button',
+            ),
+            dcc.Store(
+                id='image'
+            ),
+            html.Br(),
+            # * main plot
+            dcc.Graph(
+                id='parmoo_graph',
+                figure=self.graph,
+                config=self.config,
+            ),
+            html.Button(
+                children='Show graph customization options',
+                id='show_customization_options',
+            ),
+            html.Button(
+                children='Hide graph customization options',
+                id='hide_customization_options',
+                style=dict(display='none'),
+            ),
+            html.Button(
+                children='Show export options',
+                id='show_export_options',
+            ),
+            html.Button(
+                children='Hide export options',
+                id='hide_export_options',
+                style=dict(display='none'),
+            ),
[…]
