### 2022-08-01 7ba1d04
unify background color control
it's unnecessary to have separate controls for the paper and plot
background color. there is no use case we can see where that would be a
critical feature, and having the feature there makes setting colors more
confusing
