+++ parmoo/viz/plot.py
@@ -67,9 +67,8 @@ def scatter(
     width='auto',
     font='auto',
     fontsize='auto',
-    paper_background_color='auto',
+    background_color='auto',
     screenshot='svg',
-    graph_background_color='auto',
     image_export_format='svg',
     data_export_format='csv',
     dummy6='auto',
@@ -89,9 +88,8 @@ def scatter(
             width=width,
             font=font,
             fontsize=fontsize,
-            paper_background_color=paper_background_color,
+            background_color=background_color,
             screenshot=screenshot,
-            graph_background_color=graph_background_color,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
             dummy6=dummy6,
@@ -112,9 +110,8 @@ def scatter(
                 width=width,
                 font=font,
                 fontsize=fontsize,
-                paper_background_color=paper_background_color,
+                background_color=background_color,
                 screenshot=screenshot,
-                graph_background_color=graph_background_color,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
                 dummy6=dummy6,
@@ -175,9 +172,8 @@ def parallel_coordinates(
     width='auto',
     font='auto',
     fontsize='auto',
-    paper_background_color='auto',
+    background_color='auto',
     screenshot='svg',
-    graph_background_color='auto',
     image_export_format='svg',
     data_export_format='csv',
     dummy6='auto',
@@ -198,9 +194,8 @@ def parallel_coordinates(
             width=width,
             font=font,
             fontsize=fontsize,
-            paper_background_color=paper_background_color,
+            background_color=background_color,
             screenshot=screenshot,
-            graph_background_color=graph_background_color,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
             dummy6=dummy6,
@@ -221,9 +216,8 @@ def parallel_coordinates(
                 width=width,
                 font=font,
                 fontsize=fontsize,
-                paper_background_color=paper_background_color,
+                background_color=background_color,
                 screenshot=screenshot,
-                graph_background_color=graph_background_color,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
                 dummy6=dummy6,
@@ -278,9 +272,8 @@ def radar(
     width='auto',
     font='auto',
     fontsize='auto',
-    paper_background_color='auto',
+    background_color='auto',
     screenshot='svg',
-    graph_background_color='auto',
     image_export_format='svg',
     data_export_format='csv',
     dummy6='auto',
@@ -301,9 +294,8 @@ def radar(
             width=width,
             font=font,
             fontsize=fontsize,
-            paper_background_color=paper_background_color,
+            background_color=background_color,
             screenshot=screenshot,
-            graph_background_color=graph_background_color,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
             dummy6=dummy6,
@@ -324,9 +316,8 @@ def radar(
                 width=width,
                 font=font,
                 fontsize=fontsize,
-                paper_background_color=paper_background_color,
+                background_color=background_color,
                 screenshot=screenshot,
-                graph_background_color=graph_background_color,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
                 dummy6=dummy6,
