+++ parmoo/viz/graph.py
@@ -29,9 +29,8 @@ def generate_scatter(
     width,
     font,
     fontsize,
-    paper_background_color,
+    background_color,
     screenshot,
-    graph_background_color,
     image_export_format,
     data_export_format,
     dummy6,
@@ -83,9 +82,8 @@ def generate_parallel(
     width,
     font,
     fontsize,
-    paper_background_color,
+    background_color,
     screenshot,
-    graph_background_color,
     image_export_format,
     data_export_format,
     dummy6,
@@ -137,9 +135,8 @@ def generate_radar(
     width,
     font,
     fontsize,
-    paper_background_color,
+    background_color,
     screenshot,
-    graph_background_color,
     image_export_format,
     data_export_format,
     dummy6,
