### 2022-08-01 fc7b66d
minor update, allowing infeasible starts for algorithms that support it
