+++ parmoo/tests/unit_tests/test_weighted_sum.py
@@ -16,28 +16,26 @@ def test_UniformWeights():
     assert(np.all(acqu.lb[:] == 0.0) and np.all(acqu.ub[:] == 1.0))
     # Set some bad targets to check error handling
     with pytest.raises(ValueError):
-        acqu.setTarget(5, lambda x: np.zeros(1), {})
+        acqu.setTarget(5, lambda x: np.zeros(3), {})
     with pytest.raises(AttributeError):
-        acqu.setTarget({'x_vals': []}, lambda x: np.zeros(0), {})
+        acqu.setTarget({'x_vals': []}, lambda x: np.zeros(3), {})
     with pytest.raises(ValueError):
         acqu.setTarget({'x_vals': np.ones(1), 'f_vals': np.ones(2)},
-                       lambda x: np.zeros(0), {})
+                       lambda x: np.zeros(3), {})
     with pytest.raises(ValueError):
         acqu.setTarget({'x_vals': np.ones((1, 1)), 'f_vals': np.ones((1, 3))},
-                       lambda x: np.zeros(0), {})
+                       lambda x: np.zeros(3), {})
     with pytest.raises(ValueError):
         acqu.setTarget({'x_vals': np.ones((1, 4)), 'f_vals': np.ones((1, 1))},
-                       lambda x: np.zeros(0), {})
+                       lambda x: np.zeros(3), {})
     with pytest.raises(ValueError):
         acqu.setTarget({}, 4, {})
     with pytest.raises(ValueError):
         acqu.setTarget({}, 4, {})
     with pytest.raises(ValueError):
-        acqu.setTarget({}, lambda x, y: np.zeros(1), {})
-    with pytest.raises(ValueError):
-        acqu.setTarget({}, lambda x: np.ones(1), {})
+        acqu.setTarget({}, lambda x, y: np.zeros(3), {})
     # Set a good target for future usage
-    x0 = acqu.setTarget({}, lambda x: np.zeros(1), {})
+    x0 = acqu.setTarget({}, lambda x: np.zeros(3), {})
     assert(abs(sum(acqu.weights) - 1.0) < 0.00000001)
     assert(np.all(x0[:] <= acqu.ub) and np.all(x0[:] >= acqu.lb))
     # Try some bad scalarizations to test error handling
@@ -47,12 +45,12 @@ def test_UniformWeights():
         acqu.scalarize(np.ones(2))
     # Generate 3 random weight vector
     acqu1 = UniformWeights(3, np.zeros(4), np.ones(4), {})
-    acqu1.setTarget({}, lambda x: np.zeros(1), {})
+    acqu1.setTarget({}, lambda x: np.zeros(3), {})
     acqu2 = UniformWeights(3, np.zeros(4), np.ones(4), {})
-    acqu2.setTarget({}, lambda x: np.zeros(1), {})
+    acqu2.setTarget({}, lambda x: np.zeros(3), {})
     acqu3 = UniformWeights(3, np.zeros(4), np.ones(4), {})
     acqu3.setTarget({'x_vals': None, 'f_vals': None},
-                    lambda x: np.zeros(1), {})
+                    lambda x: np.zeros(3), {})
     # Check that the weights are all greater than 0
     assert(all(acqu1.weights[:] >= 0.0))
     assert(all(acqu2.weights[:] >= 0.0))
@@ -108,36 +106,34 @@ def test_FixedWeights():
     assert(np.all(acqu.weights[:] - (1.0 / 3.0) < 0.00000001))
     # Set some bad targets to check error handling
     with pytest.raises(ValueError):
-        acqu.setTarget(5, lambda x: np.zeros(1), {})
+        acqu.setTarget(5, lambda x: np.zeros(3), {})
     with pytest.raises(AttributeError):
[…]
