+++ parmoo/viz/plot.py
@@ -71,7 +71,7 @@ def scatter(
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
-    dummy6='auto',
+    points='satisfied',
     verbose=True,
     hot_reload=False,
     pop_up=True,
@@ -92,7 +92,7 @@ def scatter(
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
-            dummy6=dummy6,
+            points=points,
             verbose=verbose,
             hot_reload=hot_reload,
             pop_up=pop_up,
@@ -114,7 +114,7 @@ def scatter(
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
-                dummy6=dummy6,
+                points=points,
                 verbose=verbose,
             ),
             plot_name=plot_name,
@@ -176,7 +176,7 @@ def parallel_coordinates(
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
-    dummy6='auto',
+    points='satisfied',
     verbose=True,
     hot_reload=False,
     pop_up=True,
@@ -198,7 +198,7 @@ def parallel_coordinates(
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
-            dummy6=dummy6,
+            points=points,
             verbose=verbose,
             hot_reload=hot_reload,
             pop_up=pop_up,
@@ -220,7 +220,7 @@ def parallel_coordinates(
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
-                dummy6=dummy6,
+                points=points,
                 verbose=verbose,
             ),
             plot_name=plot_name,
@@ -276,7 +276,7 @@ def radar(
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
-    dummy6='auto',
+    points='satisfied',
     verbose=True,
     hot_reload=False,
     pop_up=True,
@@ -298,7 +298,7 @@ def radar(
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
-            dummy6=dummy6,
+            points=points,
             verbose=verbose,
             hot_reload=hot_reload,
             pop_up=pop_up,
@@ -320,7 +320,7 @@ def radar(
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
-                dummy6=dummy6,
+                points=points,
                 verbose=verbose,
             ),
             plot_name=plot_name,
