### 2022-08-02 510ab56
add axis color control + default axis color is black
axis color control only sort of works
