+++ parmoo/viz/dashboard.py
@@ -51,6 +51,7 @@ class Dash_App:
         font,
         fontsize,
         background_color,
+        axis_color,
         screenshot,
         image_export_format,
         data_export_format,
@@ -60,7 +61,6 @@ class Dash_App:
         pop_up,
         port,
     ):
-
     # ! STATE
 
         # * define independent state
@@ -74,6 +74,7 @@ class Dash_App:
         self.font = font
         self.fontsize = fontsize
         self.background_color = background_color
+        self.axis_color = axis_color
         self.screenshot = screenshot
         self.image_export_format = image_export_format
         self.data_export_format = data_export_format
@@ -82,7 +83,6 @@ class Dash_App:
         self.hot_reload = hot_reload
         self.pop_up = pop_up
         self.port = port
-
         # * define dependent state
         self.selection_indexes = []
         self.plot_name = set_plot_name(db=self.db)
@@ -238,6 +238,23 @@ class Dash_App:
                 style=dict(display='none'),
             ),
             dcc.Dropdown(
+                id='axis_color_dropdown',
+                options=[
+                    'White',
+                    'Grey',
+                    'Black',
+                    'Transparent',
+                    'Red',
+                    'Yellow',
+                    'Blue',
+                    'Green',
+                    'Orange',
+                    'Purple'
+                ],
+                placeholder='Set axis color',
+                style=dict(display='none'),
+            ),
+            dcc.Dropdown(
                 id='image_export_format_dropdown',
                 options=[
                     'PNG',
@@ -297,6 +314,9 @@ class Dash_App:
                 component_property='style',),
             Output(
                 component_id='background_color_dropdown',
+                component_property='style',),
+            Output(
+                component_id='axis_color_dropdown',
                 component_property='style',),
             Input(
                 component_id='show_customization_options',
@@ -375,6 +395,10 @@ class Dash_App:
             Input(
                 component_id='background_color_dropdown',
                 component_property='value',),
+            # axis color - update
+            Input(
+                component_id='axis_color_dropdown',
+                component_property='value',),
             # plot name - update
             Input(
                 component_id='plot_name_input',
@@ -399,6 +423,7 @@ class Dash_App:
             font_value,
             font_size_value,
             background_color_value,
+            axis_color_value,
             plot_name_value,
             plot_type_value,
             database_value,
@@ -415,6 +440,8 @@ class Dash_App:
                 return self.evaluate_font_size(font_size_value)
             elif 'background_color_dropdown.value' == triggered_id:
                 return self.evaluate_background_color(background_color_value)
+            elif 'axis_color_dropdown.value' == triggered_id:
+                return self.evaluate_axis_color(axis_color_value)
             elif 'plot_name_input.value' == triggered_id:
                 return self.evaluate_plot_name(plot_name_value)
             elif 'plot_type_dropdown.value' == triggered_id:
@@ -534,6 +561,7 @@ class Dash_App:
                 fontsize=self.fontsize,
                 screenshot=self.screenshot,
                 background_color=self.background_color,
+                axis_color=self.axis_color,
                 image_export_format=self.image_export_format,
                 data_export_format=self.data_export_format,
                 points=self.points,
@@ -551,6 +579,7 @@ class Dash_App:
                 fontsize=self.fontsize,
                 screenshot=self.screenshot,
                 background_color=self.background_color,
+                axis_color=self.axis_color,
                 image_export_format=self.image_export_format,
                 data_export_format=self.data_export_format,
                 points=self.points,
@@ -568,6 +597,7 @@ class Dash_App:
                 fontsize=self.fontsize,
                 screenshot=self.screenshot,
[…]
