+++ parmoo/viz/plot.py
@@ -68,6 +68,7 @@ def scatter(
     font='auto',
     fontsize='auto',
     background_color='auto',
+    axis_color='auto',
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
@@ -89,6 +90,7 @@ def scatter(
             font=font,
             fontsize=fontsize,
             background_color=background_color,
+            axis_color=axis_color,
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
@@ -111,6 +113,7 @@ def scatter(
                 font=font,
                 fontsize=fontsize,
                 background_color=background_color,
+                axis_color=axis_color,
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
@@ -173,6 +176,7 @@ def parallel_coordinates(
     font='auto',
     fontsize='auto',
     background_color='auto',
+    axis_color='auto',
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
@@ -195,6 +199,7 @@ def parallel_coordinates(
             font=font,
             fontsize=fontsize,
             background_color=background_color,
+            axis_color=axis_color,
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
@@ -217,6 +222,7 @@ def parallel_coordinates(
                 font=font,
                 fontsize=fontsize,
                 background_color=background_color,
+                axis_color=axis_color,
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
@@ -273,6 +279,7 @@ def radar(
     font='auto',
     fontsize='auto',
     background_color='auto',
+    axis_color='auto',
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
@@ -282,7 +289,6 @@ def radar(
     pop_up=True,
     port='http://127.0.0.1:8050/',
 ):
-
     if output == 'dash':
         Dash_App(
             plot_type='radar',
@@ -295,6 +301,7 @@ def radar(
             font=font,
             fontsize=fontsize,
             background_color=background_color,
+            axis_color=axis_color,
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
@@ -317,6 +324,7 @@ def radar(
                 font=font,
                 fontsize=fontsize,
                 background_color=background_color,
+                axis_color=axis_color,
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
