### 2022-08-02 d42294b
kill axis color control
rationale: it only works on radar plot, and there's no need for it as
the default axis color is now black
