+++ parmoo/viz/graph.py
@@ -30,7 +30,6 @@ def generate_scatter(
     font,
     fontsize,
     background_color,
-    axis_color,
     screenshot,
     image_export_format,
     data_export_format,
@@ -87,7 +86,6 @@ def generate_parallel(
     font,
     fontsize,
     background_color,
-    axis_color,
     screenshot,
     image_export_format,
     data_export_format,
@@ -143,7 +141,6 @@ def generate_radar(
     font,
     fontsize,
     background_color,
-    axis_color,
     screenshot,
     image_export_format,
     data_export_format,
