+++ parmoo/viz/plot.py
@@ -68,7 +68,6 @@ def scatter(
     font='auto',
     fontsize='auto',
     background_color='auto',
-    axis_color='auto',
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
@@ -90,7 +89,6 @@ def scatter(
             font=font,
             fontsize=fontsize,
             background_color=background_color,
-            axis_color=axis_color,
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
@@ -113,7 +111,6 @@ def scatter(
                 font=font,
                 fontsize=fontsize,
                 background_color=background_color,
-                axis_color=axis_color,
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
@@ -176,7 +173,6 @@ def parallel_coordinates(
     font='auto',
     fontsize='auto',
     background_color='auto',
-    axis_color='auto',
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
@@ -199,7 +195,6 @@ def parallel_coordinates(
             font=font,
             fontsize=fontsize,
             background_color=background_color,
-            axis_color=axis_color,
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
@@ -222,7 +217,6 @@ def parallel_coordinates(
                 font=font,
                 fontsize=fontsize,
                 background_color=background_color,
-                axis_color=axis_color,
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
@@ -279,7 +273,6 @@ def radar(
     font='auto',
     fontsize='auto',
     background_color='auto',
-    axis_color='auto',
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
@@ -301,7 +294,6 @@ def radar(
             font=font,
             fontsize=fontsize,
             background_color=background_color,
-            axis_color=axis_color,
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
@@ -324,7 +316,6 @@ def radar(
                 font=font,
                 fontsize=fontsize,
                 background_color=background_color,
-                axis_color=axis_color,
                 screenshot=screenshot,
                 image_export_format=image_export_format,
                 data_export_format=data_export_format,
