### 2022-08-02 f0298dd
fix selection bugs
fix bug where for scatterplot, past selection info wasn't cleared from
the selection_db when new selections (not multi-select) were made.
this bug made all selections act like multi-select, even if they weren't

fix bug where parallel coordinates selection didn't work
the code to make parallel coordinates selection work is a fairly large
piece of code with a lot of logic. as best as I can tell, it works as
intended, but it wouldn't hurt to have a second tester make sure
parallel coordinates selection export actually exports what's selected
