### 2022-08-02 5fcb54b
add logging
also changed hot_reload option to dev mode (off by default).
turning dev mode on enables hot_reload and the debugger
