+++ parmoo/viz/dashboard.py
@@ -3,6 +3,7 @@ import plotly.io as pio
 from os import environ
 from webbrowser import open_new
 from warnings import warn
+import logging
 from dash import (
     Dash,
     Input,
@@ -56,10 +57,11 @@ class Dash_App:
         data_export_format,
         points,
         verbose,
-        hot_reload,
+        dev_mode,
         pop_up,
         port,
     ):
+        logging.info('initializing dashboard')
 
         # ! STATE
 
@@ -79,7 +81,7 @@ class Dash_App:
         self.data_export_format = data_export_format
         self.points = points
         self.verbose = verbose
-        self.hot_reload = hot_reload
+        self.dev_mode = dev_mode
         self.pop_up = pop_up
         self.port = port
         # * define dependent state
@@ -312,8 +314,10 @@ class Dash_App:
         ):
             triggered_id = callback_context.triggered[0]['prop_id']
             if triggered_id == 'show_customization_options.n_clicks':
+                logging.info("'show_customization_options.n_clicks' triggered")
                 return self.evaluate_customization_options('show', s_clicks)
             elif triggered_id == 'hide_customization_options.n_clicks':
+                logging.info("'hide_customization_options.n_clicks' triggered")
                 return self.evaluate_customization_options('hide', h_clicks)
 
         # * show export options
@@ -344,8 +348,10 @@ class Dash_App:
         ):
             triggered_id = callback_context.triggered[0]['prop_id']
             if triggered_id == 'show_export_options.n_clicks':
+                logging.info("'show_export_options.n_clicks' triggered")
                 return self.evaluate_export_options('show', s_clicks)
             elif triggered_id == 'hide_export_options.n_clicks':
+                logging.info("'hide_export_options.n_clicks' triggered")
                 return self.evaluate_export_options('hide', h_clicks)
 
         # * regenerate or update graph
@@ -405,22 +411,31 @@ class Dash_App:
         ):
             triggered_id = callback_context.triggered[0]['prop_id']
             if 'graph_height_input.value' == triggered_id:
+                logging.info("'graph_height_input.value' triggered")
                 return self.evaluate_height(height_value)
             elif 'graph_width_input.value' == triggered_id:
+                logging.info("'graph_width_input.value' triggered")
                 return self.evaluate_width(width_value)
             elif 'font_selection_input.value' == triggered_id:
+                logging.info("'font_selection_input.value' triggered")
                 return self.evaluate_font(font_value)
             elif 'font_size_input.value' == triggered_id:
+                logging.info("'font_size_input.value' triggered")
                 return self.evaluate_font_size(font_size_value)
             elif 'background_color_dropdown.value' == triggered_id:
+                logging.info("'background_color_dropdown.value' triggered")
                 return self.evaluate_background_color(background_color_value)
             elif 'plot_name_input.value' == triggered_id:
+                logging.info("'plot_name_input.value' triggered")
                 return self.evaluate_plot_name(plot_name_value)
             elif 'plot_type_dropdown.value' == triggered_id:
+                logging.info("'plot_type_dropdown.value' triggered")
                 return self.evaluate_plot_type(plot_type_value)
             elif 'database_dropdown.value' == triggered_id:
+                logging.info("'database_dropdown.value' triggered")
                 return self.evaluate_database(database_value)
             elif 'constraint_checkboxes.value' == triggered_id:
+                logging.info("'constraint_checkboxes.value' triggered")
                 return self.evaluate_constraint_showr(constraint_showr_value)
 
         # * download dataset
@@ -433,6 +448,7 @@ class Dash_App:
                 component_property='n_clicks'),
         )
         def download_dataset(n_clicks):
[…]
