### 2022-08-02 336901e
add input checking
port ValueErrors in code to check_inputs(), removed them from
everywhere else

remov customize() as it is redundant with update_font() and
update_font_size

remove arguments from graph functions that aren't accessed
