+++ parmoo/viz/utilities.py
@@ -2,26 +2,6 @@ import pandas as pd
 import logging
 
 
-def customize(
-    fig,
-    font,
-    fontsize,
-):
-    if font != 'auto':
-        fig.update_layout(
-            font=dict(
-                family=font
-            )
-        )
-    if fontsize != 'auto':
-        fig.update_layout(
-            font=dict(
-                size=int(fontsize)
-            )
-        )
-    return fig
-
-
 def export_file(fig, plot_name, file_type):
     """ Display MOOP plot.
 
@@ -61,24 +41,15 @@ def export_file(fig, plot_name, file_type):
     elif file_type == 'png':
         fig.write_image(plot_name + ".png")
         logging.info("exported graph as .png")
-    else:
-        message = "ParMOO does not support '" + file_type + "'.\n"
-        message += "Supported outputs:\n"
-        message += "'dash'\n"
-        message += "'no_dash'\n"
-        message += "'html'\n"
-        message += "'pdf'\n"
-        message += "'svg'\n"
-        message += "'webp'\n"
-        message += "'jpeg'\n"
-        message += "'png'\n"
-        raise ValueError(message)
+    elif file_type == 'eps':
+        fig.write_image(plot_name + ".eps")
+        logging.info("exported graph as .eps")
 
 
-def configure(height, width, plot_name):
+def configure(height, width, plot_name, screenshot):
 
     # * set screenshot type
-    screenshot = 'svg'
+    screenshot = screenshot
 
     # * set config based on scale
     if height != 'auto' and width != 'auto':
@@ -110,8 +81,6 @@ def set_plot_name(db):
         plot_name = "Pareto Front"
     elif db == 'obj':
         plot_name = "Objective Data"
-    else:
-        raise ValueError(str(db) + "is invalid value for 'db'")
     return plot_name
 
 
@@ -120,8 +89,6 @@ def set_database(moop, db, points):
         database = pd.DataFrame(moop.getPF())
     elif db == 'obj':
         database = pd.DataFrame(moop.getObjectiveData())
-    else:
-        raise ValueError(str(db) + "is invalid value for 'db'")
     if moop.getConstraintType() is None:
         df = database
     else:
@@ -143,8 +110,6 @@ def set_database(moop, db, points):
             df = database
         elif points == 'none':
             df = database[0:0]
-        else:
-            raise ValueError(str(points) + "is invalid value for 'db'")
     return df
 
 
@@ -158,3 +123,184 @@ def set_hover_info(database, i):
         # since plotly is JavaScript-based
         # it uses HTML string formatting
     return hover_info
+
+
+def check_inputs(
+    db,
+    output,
+    axes,
+    points,
+    height,
+    width,
+    font,
+    fontsize,
+    background_color,
+    screenshot,
+    image_export_format,
+    data_export_format,
+    verbose,
+    dev_mode,
+    pop_up,
+    port,
+):
+    try:
+        if (db == 'pf' or
+            db == 'obj'):
+            pass
+        else:
+            raise ValueError(str(db) + " is an invalid value for 'db'")
+    except:
+        raise ValueError(str(db) + " is an invalid value for 'db'")
+
+    try:
+        if (output == 'dash' or
+            output == 'html' or
+            output == 'svg' or
+            output == 'pdf' or
+            output == 'eps' or
+            output == 'jpeg' or
+            output == 'png' or
+            output == 'webp'):
+            pass
+        else:
+            raise ValueError(str(output) + " is an invalid value for 'output'")
+    except:
+        raise ValueError(str(output) + " is an invalid value for 'output'")
+
+    try:
+        if (axes == 'objectives'):
+            pass
+        else:
[…]
