+++ parmoo/viz/plot.py
@@ -5,6 +5,7 @@ from .dashboard import (
 from .utilities import (
     export_file,
     set_plot_name,
+    check_inputs,
 )
 from .graph import (
     generate_scatter,
@@ -63,7 +64,7 @@ def scatter(
     db='pf',
     output='dash',
     axes='objectives',
-    specificaxes='auto',
+    points='satisfied',
     height='auto',
     width='auto',
     font='auto',
@@ -72,12 +73,29 @@ def scatter(
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
-    points='satisfied',
     verbose=True,
     dev_mode=False,
     pop_up=True,
     port='http://127.0.0.1:8050/',
 ):
+    check_inputs(
+        db=db,
+        output=output,
+        axes=axes,
+        points=points,
+        height=height,
+        width=width,
+        font=font,
+        fontsize=fontsize,
+        background_color=background_color,
+        screenshot=screenshot,
+        image_export_format=image_export_format,
+        data_export_format=data_export_format,
+        verbose=verbose,
+        dev_mode=dev_mode,
+        pop_up=pop_up,
+        port=port,
+    )
     logging.info('initialized scatter() wrapper')
     if output == 'dash':
         Dash_App(
@@ -85,7 +103,7 @@ def scatter(
             moop=moop,
             db=db,
             axes=axes,
-            specificaxes=specificaxes,
+            points=points,
             height=height,
             width=width,
             font=font,
@@ -94,7 +112,6 @@ def scatter(
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
-            points=points,
             verbose=verbose,
             dev_mode=dev_mode,
             pop_up=pop_up,
@@ -107,17 +124,9 @@ def scatter(
                 moop,
                 db=db,
                 axes=axes,
-                specificaxes=specificaxes,
-                height=height,
-                width=width,
+                points=points,
                 font=font,
                 fontsize=fontsize,
-                background_color=background_color,
-                screenshot=screenshot,
-                image_export_format=image_export_format,
-                data_export_format=data_export_format,
-                points=points,
-                verbose=verbose,
             ),
             plot_name=plot_name,
             file_type=output,
@@ -169,7 +178,7 @@ def parallel_coordinates(
     db='pf',
     output='dash',
     axes='objectives',
-    specificaxes='auto',
+    points='satisfied',
     height='auto',
     width='auto',
     font='auto',
@@ -178,12 +187,29 @@ def parallel_coordinates(
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
-    points='satisfied',
     verbose=True,
     dev_mode=False,
     pop_up=True,
     port='http://127.0.0.1:8050/',
 ):
+    check_inputs(
+        db=db,
+        output=output,
+        axes=axes,
+        points=points,
+        height=height,
+        width=width,
+        font=font,
+        fontsize=fontsize,
+        background_color=background_color,
+        screenshot=screenshot,
+        image_export_format=image_export_format,
+        data_export_format=data_export_format,
+        verbose=verbose,
+        dev_mode=dev_mode,
+        pop_up=pop_up,
+        port=port,
+    )
     logging.info('initialized parallel_coordinates() wrapper')
     if output == 'dash':
         Dash_App(
@@ -191,7 +217,7 @@ def parallel_coordinates(
             moop=moop,
             db=db,
             axes=axes,
-            specificaxes=specificaxes,
+            points=points,
             height=height,
             width=width,
             font=font,
@@ -200,7 +226,6 @@ def parallel_coordinates(
             screenshot=screenshot,
             image_export_format=image_export_format,
[…]
