+++ parmoo/tests/unit_tests/test_viz.py
@@ -1,0 +1,6 @@
+def test_parallel_coordinates_basic():
+
+    import parmoo.viz
+    import examples
+    import pytest
+    import os
