+++ parmoo/viz/test_viz.py
@@ -1,7 +1,0 @@
-test_description():
-    pass
-
-bugs
-dependency list
-tests
-docs
