+++ parmoo/viz/__init__.py
@@ -4,7 +4,46 @@ from .plot import (
     radar,
 )
 from .dashboard import (
-    Dash_App
+    Dash_App,
+    __init__,
+    update_customization_components,
+    update_export_components,
+    update_graph,
+    download_dataset,
+    store_selection,
+    download_selection,
+    update_image_export_format,
+    update_data_export_format,
+    download_image,
+    generate_graph,
+    configure,
+    update_height,
+    update_width,
+    update_font,
+    update_font_size,
+    update_plot_name,
+    update_background_color,
+    update_plot_type,
+    update_database,
+    evaluate_height,
+    evaluate_width,
+    evaluate_font,
+    evaluate_font_size,
+    evaluate_background_color,
+    evaluate_plot_name,
+    evaluate_plot_type,
+    evaluate_database,
+    evaluate_dataset_download,
+    evaluate_selected_data,
+    set_constraint_range,
+    update_constraint_range,
+    evaluate_selection_download,
+    evaluate_image_export_format,
+    evaluate_data_export_format,
+    evaluate_image_download,
+    evaluate_customization_options,
+    evaluate_export_options,
+    evaluate_constraint_showr,
 )
 from .graph import (
     generate_scatter,
@@ -17,4 +56,5 @@ from .utilities import (
     set_plot_name,
     set_database,
     set_hover_info,
+    check_inputs,
 )
