### 2022-08-03 0bf8c79
remove axes keyword
because it doesn't do anything.
originally this was meant to specify whether constraints
would be plotted as axes or only objectives, and provide a port into
using arbitary-axes wrapper functions
we didn't have time to implement that functionality
