+++ parmoo/viz/plot.py
@@ -63,7 +63,6 @@ def scatter(
     moop,
     db='pf',
     output='dash',
-    axes='objectives',
     points='satisfied',
     height='auto',
     width='auto',
@@ -81,7 +80,6 @@ def scatter(
     check_inputs(
         db=db,
         output=output,
-        axes=axes,
         points=points,
         height=height,
         width=width,
@@ -102,7 +100,6 @@ def scatter(
             plot_type='scatter',
             moop=moop,
             db=db,
-            axes=axes,
             points=points,
             height=height,
             width=width,
@@ -123,10 +120,7 @@ def scatter(
             fig=generate_scatter(
                 moop,
                 db=db,
-                axes=axes,
                 points=points,
-                font=font,
-                fontsize=fontsize,
             ),
             plot_name=plot_name,
             file_type=output,
@@ -177,7 +171,6 @@ def parallel_coordinates(
     moop,
     db='pf',
     output='dash',
-    axes='objectives',
     points='satisfied',
     height='auto',
     width='auto',
@@ -195,7 +188,6 @@ def parallel_coordinates(
     check_inputs(
         db=db,
         output=output,
-        axes=axes,
         points=points,
         height=height,
         width=width,
@@ -216,7 +208,6 @@ def parallel_coordinates(
             plot_type='parallel',
             moop=moop,
             db=db,
-            axes=axes,
             points=points,
             height=height,
             width=width,
@@ -237,10 +228,7 @@ def parallel_coordinates(
             fig=generate_parallel(
                 moop=moop,
                 db=db,
-                axes=axes,
                 points=points,
-                font=font,
-                fontsize=fontsize,
             ),
             plot_name=plot_name,
             file_type=output
@@ -285,7 +273,6 @@ def radar(
     moop,
     db='pf',
     output='dash',
-    axes='objectives',
     points='satisfied',
     height='auto',
     width='auto',
@@ -303,7 +290,6 @@ def radar(
     check_inputs(
         db=db,
         output=output,
-        axes=axes,
         points=points,
         height=height,
         width=width,
@@ -324,7 +310,6 @@ def radar(
             plot_type='radar',
             moop=moop,
             db=db,
-            axes=axes,
             points=points,
             height=height,
             width=width,
@@ -345,10 +330,7 @@ def radar(
             fig=generate_radar(
                 moop,
                 db=db,
-                axes=axes,
                 points=points,
-                font=font,
-                fontsize=fontsize,
             ),
             plot_name=plot_name,
             file_type=output
