### 2022-08-03 abd3b06
added assert statement and cleanup to test_parallel_coordinates.py
