### 2022-08-03 513d872
goodbye utilities configure
function redundant and not used
