### 2022-08-03 5c422fa
polishing tweaks
remove 'verbose' keyword because it doesn't do anything

remove colorful background color options because they're unprofessional,
bright, and unlikely to be used

rename 'points' options to make name meanings more self-evident

flake some input checker line lengths
