+++ parmoo/viz/plot.py
@@ -63,7 +63,7 @@ def scatter(
     moop,
     db='pf',
     output='dash',
-    points='satisfied',
+    points='constraint_satisfying',
     height='auto',
     width='auto',
     font='auto',
@@ -72,7 +72,6 @@ def scatter(
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
-    verbose=True,
     dev_mode=False,
     pop_up=True,
     port='http://127.0.0.1:8050/',
@@ -89,7 +88,6 @@ def scatter(
         screenshot=screenshot,
         image_export_format=image_export_format,
         data_export_format=data_export_format,
-        verbose=verbose,
         dev_mode=dev_mode,
         pop_up=pop_up,
         port=port,
@@ -109,7 +107,6 @@ def scatter(
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
-            verbose=verbose,
             dev_mode=dev_mode,
             pop_up=pop_up,
             port=port,
@@ -171,7 +168,7 @@ def parallel_coordinates(
     moop,
     db='pf',
     output='dash',
-    points='satisfied',
+    points='constraint_satisfying',
     height='auto',
     width='auto',
     font='auto',
@@ -180,7 +177,6 @@ def parallel_coordinates(
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
-    verbose=True,
     dev_mode=False,
     pop_up=True,
     port='http://127.0.0.1:8050/',
@@ -197,7 +193,6 @@ def parallel_coordinates(
         screenshot=screenshot,
         image_export_format=image_export_format,
         data_export_format=data_export_format,
-        verbose=verbose,
         dev_mode=dev_mode,
         pop_up=pop_up,
         port=port,
@@ -217,7 +212,6 @@ def parallel_coordinates(
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
-            verbose=verbose,
             dev_mode=dev_mode,
             pop_up=pop_up,
             port=port,
@@ -273,7 +267,7 @@ def radar(
     moop,
     db='pf',
     output='dash',
-    points='satisfied',
+    points='constraint_satisfying',
     height='auto',
     width='auto',
     font='auto',
@@ -282,7 +276,6 @@ def radar(
     screenshot='svg',
     image_export_format='svg',
     data_export_format='csv',
-    verbose=True,
     dev_mode=False,
     pop_up=True,
     port='http://127.0.0.1:8050/',
@@ -299,7 +292,6 @@ def radar(
         screenshot=screenshot,
         image_export_format=image_export_format,
         data_export_format=data_export_format,
-        verbose=verbose,
         dev_mode=dev_mode,
         pop_up=pop_up,
         port=port,
@@ -319,7 +311,6 @@ def radar(
             screenshot=screenshot,
             image_export_format=image_export_format,
             data_export_format=data_export_format,
-            verbose=verbose,
             dev_mode=dev_mode,
             pop_up=pop_up,
             port=port,
