### 2022-08-03 8a9f46d
add tests
tests don't cover the dash app code
