+++ parmoo/tests/unit_tests/test_viz.py
@@ -1,271 +1,0 @@
-def test_parallel_coordinates_static():
-
-    from parmoo.viz.plot import parallel_coordinates
-    import pytest
-    import os
-    parallel_coordinates(run_quickstart(), output='png')
-    parallel_coordinates(run_quickstart(), output='png')
-    parallel_coordinates(run_quickstart(), output='png')
-    parallel_coordinates(run_quickstart(), output='png')
-    assert(os.path.exists("Pareto front.png"))
-    os.remove("Pareto front.png")
-
-
-def run_quickstart():
-
-    import numpy as np
-    from parmoo import MOOP
-    from parmoo.searches import LatinHypercube
-    from parmoo.surrogates import GaussRBF
-    from parmoo.acquisitions import UniformWeights
-    from parmoo.optimizers import LocalGPS
-
-    my_moop = MOOP(LocalGPS)
-
-    my_moop.addDesign({
-        'name': "x1",
-        'des_type': "continuous",
-        'lb': 0.0, 'ub': 1.0
-    })
-    my_moop.addDesign({
-        'name': "x2",
-        'des_type': "categorical",
-        'levels': 3
-    })
-
-    def sim_func(x):
-        if x["x2"] == 0:
-            return np.array([(x["x1"] - 0.2) ** 2, (x["x1"] - 0.8) ** 2])
-        else:
-            return np.array([99.9, 99.9])
-
-    my_moop.addSimulation({
-        'name': "MySim",
-        'm': 2,
-        'sim_func': sim_func,
-        'search': LatinHypercube,
-        'surrogate': GaussRBF,
-        'hyperparams': {'search_budget': 20}
-    })
-
-    my_moop.addObjective({
-        'name': "Cost of driving",
-        'obj_func': lambda x, s: s["MySim"][0]
-    })
-    my_moop.addObjective({
-        'name': "Time spent driving",
-        'obj_func': lambda x, s: s["MySim"][1]
-    })
-
-    my_moop.addConstraint({
-        'name': "c1",
-        'constraint': lambda x, s: 0.1 - x["x1"]
-    })
-
-    for i in range(3):
-        my_moop.addAcquisition({
-            'acquisition': UniformWeights,
-            'hyperparams': {}
-        })
-
-    my_moop.solve(5)
-
-    return my_moop
-
-
-def run_dtlz2():
-    from parmoo import MOOP
-    from parmoo.acquisitions import RandomConstraint
-    from parmoo.searches import LatinHypercube
-    from parmoo.surrogates import GaussRBF
-    from parmoo.optimizers import LBFGSB
-    from parmoo.objectives.dtlz import dtlz2_obj
-    from parmoo.simulations.dtlz import g2_sim
-
-    n = 6  # number of design variables
-    o = 5  # number of objectives
-    q = 4  # batch size (number of acquisitions)
-
-    # Create MOOP
-    moop = MOOP(LBFGSB)
-    # Add n design variables
-    for i in range(n):
-        moop.addDesign({
-            'name': f"Input {i+1}",
-            'des_type': 'continuous',
-            'lb': 0.0,
-            'ub': 1.0,
-            'des_tol': 1.0e-8
-        })
-
-    # Create the g2 simulation
-    moop.addSimulation({
-        'name': "g2",
-        'm': 1,
-        'sim_func': g2_sim(
-            moop.getDesignType(),
-            num_obj=o,
-            offset=0.5
-        ),
-        'search': LatinHypercube,
-        'surrogate': GaussRBF,
-        'hyperparams': {'search_budget': 10*n}
-    })
-    # Add o objectives
-    for i in range(o):
-        moop.addObjective({
-            'name': f"Objective {i+1}",
[…]
