+++ parmoo/tests/unit_tests/test_viz_without_dash.py
@@ -1,0 +1,891 @@
+def test_static_export():
+
+    # tests all export formats except eps
+    # (because eps may not be supported on all machines)
+
+    from parmoo.viz.plot import (
+        scatter,
+        parallel_coordinates,
+        radar,
+    )
+    import os
+
+    # * html output
+    scatter(moop=run_quickstart(), output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+    # * svg output
+    parallel_coordinates(moop=run_quickstart(), output='svg')
+    assert(os.path.exists("Pareto front.svg"))
+    os.remove("Pareto front.svg")
+
+    # * pdf output
+    radar(moop=run_quickstart(), output='pdf')
+    assert(os.path.exists("Pareto front.pdf"))
+    os.remove("Pareto front.pdf")
+
+    # * jpeg output
+    scatter(moop=run_quickstart(), output='jpeg')
+    assert(os.path.exists("Pareto front.jpeg"))
+    os.remove("Pareto front.jpeg")
+
+    # * png output
+    parallel_coordinates(moop=run_quickstart(), output='png')
+    assert(os.path.exists("Pareto front.png"))
+    os.remove("Pareto front.png")
+
+    # * webp output
+    radar(moop=run_quickstart(), output='webp')
+    assert(os.path.exists("Pareto front.webp"))
+    os.remove("Pareto front.webp")
+
+
+def test_quantity_constraints_objectives():
+
+    from parmoo.viz.plot import (
+        scatter,
+        parallel_coordinates,
+        radar,
+    )
+    import os
+
+    # * 2 objective scatter with constraint
+    scatter(moop=run_quickstart(), output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+    # * 2 objective parallel with constraint
+    parallel_coordinates(moop=run_quickstart(), output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+    # * 2 objective radar with constraint
+    radar(moop=run_quickstart(), output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+    # * 5 objective scatter without constraint
+    scatter(moop=run_dtlz2(), output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+    # * 5 objective parallel without constraint
+    parallel_coordinates(moop=run_dtlz2(), output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+    # * 5 objective radar without constraint
+    radar(moop=run_dtlz2(), output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+
+def test_database_options():
+
+    # test all database-plotting options for all plots
+
+    from parmoo.viz.plot import (
+        scatter,
+        parallel_coordinates,
+        radar,
+    )
+    import os
+
+    # * pf x constraint_satisfying x constraints in MOOP
+    scatter(
+        moop=run_quickstart(),
+        db='pf',
+        points='constraint_satisfying',
+        output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+    parallel_coordinates(
+        moop=run_quickstart(),
+        db='pf',
+        points='constraint_satisfying',
+        output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
+    radar(
+        moop=run_quickstart(),
+        db='pf',
+        points='constraint_satisfying',
+        output='html')
+    assert(os.path.exists("Pareto front.html"))
+    os.remove("Pareto front.html")
+
[…]
