### 2022-08-03 1445c9e
pre-computed moop objects for viz tests to save work during testing, and changed header comments for viz unit-tests to docstrings
