+++ parmoo/tests/unit_tests/test_viz_without_dash.py
@@ -1,7 +1,10 @@
 def test_static_export():
+    """ Create a MOOP object and export various plot types.
 
-    # tests all export formats except eps
-    # (because eps may not be supported on all machines)
+    Tests all export formats except eps
+    (because eps may not be supported on all machines).
+
+    """
 
     from parmoo.viz.plot import (
         scatter,
@@ -10,38 +13,42 @@ def test_static_export():
     )
     import os
 
+    # Pre-calculate a moop object
+    moop1 = run_quickstart()
+
     # * html output
-    scatter(moop=run_quickstart(), output='html')
+    scatter(moop1, output='html')
     assert(os.path.exists("Pareto front.html"))
     os.remove("Pareto front.html")
 
     # * svg output
-    parallel_coordinates(moop=run_quickstart(), output='svg')
+    parallel_coordinates(moop1, output='svg')
     assert(os.path.exists("Pareto front.svg"))
     os.remove("Pareto front.svg")
 
     # * pdf output
-    radar(moop=run_quickstart(), output='pdf')
+    radar(moop1, output='pdf')
     assert(os.path.exists("Pareto front.pdf"))
     os.remove("Pareto front.pdf")
 
     # * jpeg output
-    scatter(moop=run_quickstart(), output='jpeg')
+    scatter(moop1, output='jpeg')
     assert(os.path.exists("Pareto front.jpeg"))
     os.remove("Pareto front.jpeg")
 
     # * png output
-    parallel_coordinates(moop=run_quickstart(), output='png')
+    parallel_coordinates(moop1, output='png')
     assert(os.path.exists("Pareto front.png"))
     os.remove("Pareto front.png")
 
     # * webp output
-    radar(moop=run_quickstart(), output='webp')
+    radar(moop1, output='webp')
     assert(os.path.exists("Pareto front.webp"))
     os.remove("Pareto front.webp")
 
 
 def test_quantity_constraints_objectives():
+    """ Create a MOOP object and plot PF with and w/o constraint data. """
 
     from parmoo.viz.plot import (
         scatter,
@@ -50,40 +57,47 @@ def test_quantity_constraints_objectives():
     )
     import os
 
+    # Pre-calculate two moop objects
+    moop1 = run_quickstart()
+    moop2 = run_dtlz2()
+
     # * 2 objective scatter with constraint
-    scatter(moop=run_quickstart(), output='html')
+    scatter(moop1, output='html')
     assert(os.path.exists("Pareto front.html"))
     os.remove("Pareto front.html")
 
     # * 2 objective parallel with constraint
-    parallel_coordinates(moop=run_quickstart(), output='html')
+    parallel_coordinates(moop1, output='html')
     assert(os.path.exists("Pareto front.html"))
     os.remove("Pareto front.html")
 
     # * 2 objective radar with constraint
-    radar(moop=run_quickstart(), output='html')
+    radar(moop1, output='html')
     assert(os.path.exists("Pareto front.html"))
     os.remove("Pareto front.html")
 
     # * 5 objective scatter without constraint
-    scatter(moop=run_dtlz2(), output='html')
+    scatter(moop2, output='html')
     assert(os.path.exists("Pareto front.html"))
     os.remove("Pareto front.html")
 
     # * 5 objective parallel without constraint
-    parallel_coordinates(moop=run_dtlz2(), output='html')
+    parallel_coordinates(moop2, output='html')
     assert(os.path.exists("Pareto front.html"))
     os.remove("Pareto front.html")
 
     # * 5 objective radar without constraint
-    radar(moop=run_dtlz2(), output='html')
+    radar(moop2, output='html')
     assert(os.path.exists("Pareto front.html"))
     os.remove("Pareto front.html")
 
 
 def test_database_options():
+    """ Create a MOOP object and plot the full database.
 
[…]
