### 2022-08-03 b946b6b
edited error messages when improper return format is given, and removed unused imports from viz tests
