### 2022-08-05 7cdecd6
added viz.rst template to sphinx autodocs
