+++ docs/modules/viz.rst
@@ -1,0 +1,25 @@
+Parallel Solvers using the libE_MOOP Class
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Solve MOOPs in parallel using ParMOO together with libEnsemble.
+
+.. code-block:: python
+
+    from parmoo.extras import libE_MOOP
+
+The ``libE_MOOP.solve(...)`` method will distribute simulations across
+parallel resources using libEnsemble for this class.
+
+.. automodule:: viz.file1
+..    :members: viz/file1
+
+.. autofunction:: my_func
+
+.. automodule:: viz.file2
+..    :members: viz/file2
+
+.. autoclass:: Class2
+   :member-order: bysource
+   :members:
+
+   .. automethod:: __init__
