+++ parmoo/viz/utilities.py
@@ -120,21 +120,14 @@ def set_hover_info(database, i):
     """ Customize information in hover label for trace i.
 
         Args:
-            database (Pandas dataframe):
+            database (Pandas dataframe): A 2D dataframe containing the
+                traces to be graphed.
             i (int): An index indicating the row where the trace
                 we're labeling is located.
 
-            points (string): Filter traces from dataset by constraint score.
-                'constraint_satisfying' - Include only points that
-                    satisfy every constraint.
-                'constraint_violating' - Include only points that
-                    violate any constraint.
-                'all' - Include all points in dataset.
-                'none' - Include no points in dataset.
-
         Returns:
-            df (Pandas dataframe): A 2D dataframe containing post-filter
-                data from the MOOP.
+        hover_info (string): An HTML-format string to display when
+        users hover over trace i.
     """
 
     hover_info = ""
@@ -164,9 +157,85 @@ def check_inputs(
     pop_up,
     port,
 ):
+
+    """ Check keyword inputs to user-facing functions for validity
+
+        Args:
+            db: The item passed to the 'db' keyword in a user-facing function.
+                If db cannot be cast to a string valued 'pf' or 'obj',
+                a ValueError is raised.
+            output: The item passed to the 'output' keyword in a
+                user-facing function.
+                If output cannot be cast to a string corresponding to one of
+                the supported output filetypes, a ValueError is raised.
+            points: The item passed to the 'points' keyword in a
+                user-facing function.
+                If points cannot be cast to a string corresponding to one of
+                the supported contraint filters, a ValueError is raised.
+            height: The item passed to the 'height' keyword in a user-facing
+                function.
+                If height is not the default string 'auto' or cannot be cast
+                to an int
+                of value greater than one, a ValueError is raised.
+            width: The item passed to the 'width' keyword in a user-facing
+                function.
+                If width is not the default string 'auto' or cannot be cast
+                to an int
+                of value greater than one, a ValueError is raised.
+            font: The item passed to the 'font' keyword in a user-facing
+                function.
+                If font cannot be cast to a string, a ValueError is raised.
+            fontsize: The item passed to the 'fontsize' keyword in a
+                user-facing function.
+                If fontsize is not the default value 'auto' or cannot be cast
+                to an int
+                of value between 1 and 100 inclusive, a ValueError is raised.
+            background_color: The item passed to the 'background_color'
+                keyword in a user-facing function.
+                If background_color cannot be cast to a string, a ValueError
+                is raised.
+            screenshot: The item passed to the 'screenshot' keyword in a
+                user-facing function.
+                If screenshot cannot be cast to a string corresponding to one
+                of the supported
+                screenshot filetypes, a ValueError is raised.
+            image_export_format: The item passed to the 'image_export_format'
+                keyword in a user-facing function.
+                If image_export_format cannot be cast to a string
+                corresponding to one of the supported
+                image_export_format filetypes, a ValueError is raised.
+            data_export_format: The item passed to the 'data_export_format'
+                keyword in a user-facing function.
+                If data_export_format cannot be cast to a string corresponding
+                to one of the supported
+                data_export_format filetypes, a ValueError is raised.
+            data_export_format: The item passed to the 'data_export_format'
+                keyword in a user-facing function.
+                If data_export_format cannot be cast to a string corresponding
+                to one of the supported
+                data_export_format filetypes, a ValueError is raised.
+            dev_mode: The item passed to the 'dev_mode' keyword in a
+                user-facing function.
[…]
