### 2022-08-22 14e1bd2
started on Viech's suggestions about the code
