### 2022-08-26 fc7280a
working on filling in extra refs to paper
