### 2022-08-29 a4fd39a
added dois to bibs, filling in refs, and revising intro section
