+++ docs/papers/joss/paper.bib
@@ -5,6 +5,7 @@
   journal  = {IEEE Access},
   volume   = {8},
   pages    = {89497--89509},
+  doi      = {10.1109/ACCESS.2020.2990567},
   software = {http://github.com/anyoptimization/pymoo}
 }
 
@@ -17,6 +18,7 @@
   volume   = {71},
   number   = {1},
   pages    = {53--72},
+  doi      = {10.1007/s10589-017-9955-0},
   sofware  = {http://www.iasi.cnr.it/~liuzzi/DFL/index.php/home}
 }
 
@@ -27,10 +29,8 @@
               Optimization Problems},
   year     = {2022},
   journal  = {{ACM} Transactions on Mathematical Software},
-  volume   = {},
-  number   = {},
-  pages    = {},
-  note     = {To appear},
+  volume   = {Just accepted},
+  doi      = {10.1145/3529258},
   software = {https://github.com/Libensemble/libe-community-examples/tree/main/vtmop}
 }
 
@@ -43,6 +43,7 @@
   volume   = {32},
   number   = {4},
   pages    = {1101--1108},
+  doi      = {10.1287/ijoc.2019.0902},
   software = {http://github.com/pymoso/PyMOSO}
 }
 
@@ -57,28 +58,18 @@
   volume   = {21},
   number   = {81},
   pages    = {1--27},
+  url      = {http://jmlr.org/papers/v21/18-223.html},
   software = {http://github.com/dragonfly/dragonfly}
-}
-
-@article{edbo,
-  author   = {Shields, B. J. and Stevens, J. and Li, J. and Parasram, M. and
-              Damani, F. and Alvarado, J. I. M. and Janey, J. M. and
-              Adams, R. P. and Doyle, A. G.},
-  title    = {Bayesian reaction optimization as a tool for chemical synthesis},
-  year     = {2021},
-  journal  = {Nature},
-  volume   = {590},
-  number   = {7844},
-  pages    = {89--96},
-  software = {http://github.com/b-shields/edbo}
 }
 
 @misc{boostdfo,
   author   = {Tavares, S. and Br{\'a}s, C. P. and Cust{\'o}dio, A. L. and
               Duarte, V. and Medeiros, P.},
   title    = {Parallel strategies for Direct Multisearch},
-  year     = {2021},
-  journal  = {arXiv preprint arXiv:2105.03000},
+  year     = {2022},
+  journal  = {Numerical Algorithms},
+  volume   = {Online first},
+  doi = {10.1007/s11075-022-01364-1},
   software = {https://docentes.fct.unl.pt/algb/pages/boostdfo}
 }
 
@@ -91,5 +82,71 @@
   volume   = {33},
   number   = {4},
   pages    = {977--988},
+  doi     = {10.1109/tpds.2021.3082815},
   software = {https://github.com/Libensemble/libensemble}
+}
+
+@techreport{platypus,
+  title={Platypus - multiobjective optimization in {P}ython},
+  author={Hadka, David},
+  year={2015},
+  institution={GitHub},
+  number={Version 1.0.4},
+  url={https://platypus.readthedocs.io/en/latest}
+}
+
+@article{jMetalPy,
+  title = {{jMetalPy}: A {P}ython framework for multi-objective optimization with metaheuristics},
+  author = {Antonio Ben{\'i}tez-Hidalgo and Antonio J. Nebro and Jos{\'e} Garc{\'i}a-Nieto and Izaskun Oregi and Javier {Del Ser}},
+  year = {2019},
+  journal = {Swarm and Evolutionary Computation},
+  volume = {51},
+  pages = {100598},
+  doi = {10.1016/j.swevo.2019.100598},
+}
+
[…]
