### 2022-08-29 e30991e
more revisions to joss paper - statemnt of need and methodology
