+++ docs/papers/joss/paper.md
@@ -39,19 +39,16 @@ Such problems are ubiquitous in science and engineering.
 ParMOO is a Python framework and library of solver components for building
 and deploying highly customized multiobjective simulation optimization solvers.
 ParMOO is designed to help engineers, practitioners, and optimization experts
-to easily exploit available structure in how the simulation outputs are used
-to formulate the objectives for the optimization problem.
+exploit available structures in how simulation outputs are used to formulate
+the objectives for a MOOP.
 We will elaborate on these structures, and provide two examples in future
 sections.
+
 Additionally, ParMOO is:
 
  - an open-source project on [GitHub](https://github.com/parmoo/parmoo),
  - ``pip``-installable via [PyPI](https://pypi.org/project/parmoo), and
  - fully [documented](https://parmoo.readthedocs.io).
-
-# Target Audience
-
-For this paper, the target audience includes 
 
 # Multiobjective Simulation Optimization Software
 
@@ -96,25 +93,76 @@ However, ParMOO is distinct for the following reasons:
     solvers, such as checkpointing and logging.
  4) By layering on top of libEnsemble [@libensemble], ParMOO provides an 
     easy-to-use interface for distributing expensive simulation calculations
-    over HPC resources.
+    over high performance computing (HPC) resources.
  5) ParMOO provides complete documentation, including instructions for
     potential contributors.
  6) ParMOO is designed around extensibility and continuous-integration, with
     the intention of adding support for new features, solvers, techniques,
     and problem types, some of which may be beyond what we originally
     envisioned.
- 7) In situations where there is an exploitable simulation-based structure
-    in the problem definition
-    (e.g., one or more objectives/constraints does not depend on any
-    simulation outputs or the objectives are sum-of-squares of simulation
-    outputs),
-    then ParMOO can exploit this structure by collecting simulation data
-    and modeling simulation outputs independently from the objectives and
-    constraints.
+ 7) In situations where there is an exploitable structure in how the
+    simulation outputs are used to define the objectives and/or constraints,
+    ParMOO can exploit this structure by modeling simulation outputs
+    independently.
 
 While many existing solvers provide one or more of properties 1-5,
 at this time, no other solver has *all* of these properties at once.
 Additionally, to our knowledge, properties 6 and 7 are unique to ParMOO.
+
+The target audience for ParMOO includes scientists, engineers, practitioners,
+and optimizers, who are looking to build or use custom solvers for
+simulation- or experimentation-based MOOPs.
+
+# Our Methodology
+
+In our *statement of need*, we outlined the properties that make ParMOO
+unique.
+In this section, we outline our strategy for achieving these goals.
+In particular, properties 1, 6, and 7 are non trivial.
+
+First, in order to achieve flexibility and customizability without sacrificing
+ease of use, we have focused on implementing a multiobjective response surface
+methodology (RSM) framework, which encompasses a wide range of existing
+techniques.
+Using the RSM framework, we decompose multiobjective simulation optimization
+problems into four central components:
+  i)  an initial search/design of experiments, used to generate the initial
+      data set;
+ ii)  multiple surrogates, used to model the simulation outputs based on
+      existing data;
+ iii) one or more families of acquisition functions, used to scalarize the
+      problem and guide the optimization solver to multiple distinct solutions;
+      and
+ iv)  a single-objective optimization solver, used to solve the scalarized
+      surrogate problems, in order to produce batches of candidate solution
+      points.
+
+In order to achieve property 1, we provide a customizable embedding layer,
+which can be used to embed categorical, integer, mixed-variable, and other
+input types into a continuous latent space, where the above components can
+be easily applied.
+We also support nonlinear relaxable constraints, by using a multiobjective
+progressive barrier method.
+
[…]
