### 2022-08-29 2e66158
finished updating viz docs, and credited hyrum in several places, and referenced viz docs in several places in the how-to-write-a-moop section
