+++ parmoo/viz/plot.py
@@ -1,3 +1,50 @@
+
+""" A plotting library for interactive visualization of MOOP objects.
+
+To generate a plot, generate create a MOOP object and use it to solve a problem.
+Then call one of the following function, passing the MOOP object as the
+first argument:
+
+ - ``viz.scatter(moop)``
+ - ``viz.parallel_coordinates(moop)``
+ - ``viz.radar(moop)``
+
+Please take note of the following:
+
+ * Via the interactivity features, all of the arguments except for the
+   ``moop`` field can be adjusted interactively through the pop-up GUI.
+   The plot type can also be adjusted from the dropdown menu while running
+   the GUI..
+ * The viz tool hosts a Dash app through the local port
+   ``http://127.0.0.1:8050/``.
+   This means that only one visualization can be hosted at a time.
+   Also, killing the Dash app will end the interactivity in the browser,
+   although a static plot may remain in your broswer window.
+ * Finally, note that there is a known issue when using the Chrome browser.
+   The Firefox or Safari browsers are recommended.
+
+To interact with the plots:
+
+ 1. To multi-select in scatterplots and radar plots, hold down ``SHIFT``
+    while making additional selections.
+ 2. To remove all selections from a scatterplot and radar plot, select without
+    holding down the ``SHIFT`` key or double click.
+ 3. To select in parallel coordinates plots, click along an axis and then
+    drag the cursor elsewhere along the same axis.
+    The highlighted part of the axis is the area that is selected.
+    There is no need to hold down SHIFT when making multiple selections in
+    parallel coordinates plots.
+ 4. To remove a parallel coordinates plot selection, click on the highlighted
+    selection bar you wish to delete.
+ 5. To reset plot interactions made through the toolbar in the top right of a
+    graph (not interactions made through buttons, dropdowns, toggles, or input
+    boxes), double click on the plot. This will undo selections, zoom, pan,
+    etc.
+
+The three basic plot options are detailed below.
+
+"""
+
 import logging
 from .dashboard import (
     Dash_App
@@ -14,23 +61,94 @@ from .graph import (
 )
 
 
-def scatter(
-    moop,
-    db='pf',
-    output='dash',
-    points='constraint_satisfying',
-    height='auto',
-    width='auto',
-    font='auto',
-    fontsize='auto',
-    background_color='auto',
-    screenshot='svg',
-    image_export_format='svg',
-    data_export_format='csv',
-    dev_mode=False,
-    pop_up=True,
-    port='http://127.0.0.1:8050/',
-):
+def scatter(moop,
+            db='pf',
+            output='dash',
+            points='constraint_satisfying',
+            height='auto',
+            width='auto',
+            font='auto',
+            fontsize='auto',
+            background_color='auto',
+            screenshot='svg',
+            image_export_format='svg',
+            data_export_format='csv',
+            dev_mode=False,
+            pop_up=True,
+            port='http://127.0.0.1:8050/',
+           ):
+    """ Create a scatter plot matrix to visualize the results of a MOOP.
+
+    Args:
+        moop (MOOP): The MOOP results that you would like to visualize.
+
+        db (str): Either 'pf' to plot just the Pareto front, or 'obj'
+            to plot the complete objective database. Defaults to 'pf'.
+
+        output (str): Either 'dash' to generate an interactive plot
+            running in your browser using the dash app, or anything
+            else to save a static plot to the desktop. Defaults to
+            'dash'.
+
+        points (str): Plot only
+            constraint satisfying points ('constraint_satisfying'),
+            constraint violating points ('constraint_violating'),
+            all points ('all'),
+            or no points ('none').
+
+        height (str): The height in pixels of the resulting figure.
+            Defaults to 'auto', which matches your screen size.
+
[…]
