+++ README.rst
@@ -246,7 +246,7 @@ Citing ParMOO:
 
     @techreport{parmoo-docs,
         title       = {{ParMOO}: {P}ython library for parallel multiobjective simulation optimization},
-        author      = {Chang, Tyler H. and Wild, Stefan M.},
+        author      = {Chang, Tyler H. and Wild, Stefan M. and Dickinson, Hyrum},
         institution = {Argonne National Laboratory},
         number      = {Version 0.1.0},
         year        = {2022},

[…]
