### 2022-09-05 090fb14
Update viz.rst
Minor clarification in viz module
