### 2022-09-07 #4468 @kellyrowland
Hi @openjournals/joss-eics @thchang @Viech @jonjoncardoso  - I just wanted to let you all know that I will be out of office and not checking emails, Github, etc. starting September 10 and will be returning on October 3. Happy to field any questions through the end of the week.
