### 2022-09-07 dc21515
updated docs with additional dependencies, added viz to setup.py package list, and changed filename for check in viz tests
