+++ parmoo/tests/unit_tests/test_viz_without_dash.py
@@ -18,33 +18,33 @@ def test_static_export():
 
     # * html output
     scatter(moop1, output='html')
-    assert(os.path.exists("Pareto front.html"))
-    os.remove("Pareto front.html")
+    assert(os.path.exists("Pareto Front.html"))
+    os.remove("Pareto Front.html")
 
     # * svg output
     parallel_coordinates(moop1, output='svg')
-    assert(os.path.exists("Pareto front.svg"))
-    os.remove("Pareto front.svg")
+    assert(os.path.exists("Pareto Front.svg"))
+    os.remove("Pareto Front.svg")
 
     # * pdf output
     radar(moop1, output='pdf')
-    assert(os.path.exists("Pareto front.pdf"))
-    os.remove("Pareto front.pdf")
+    assert(os.path.exists("Pareto Front.pdf"))
+    os.remove("Pareto Front.pdf")
 
     # * jpeg output
     scatter(moop1, output='jpeg')
-    assert(os.path.exists("Pareto front.jpeg"))
-    os.remove("Pareto front.jpeg")
+    assert(os.path.exists("Pareto Front.jpeg"))
+    os.remove("Pareto Front.jpeg")
 
     # * png output
     parallel_coordinates(moop1, output='png')
-    assert(os.path.exists("Pareto front.png"))
-    os.remove("Pareto front.png")
+    assert(os.path.exists("Pareto Front.png"))
+    os.remove("Pareto Front.png")
 
     # * webp output
     radar(moop1, output='webp')
-    assert(os.path.exists("Pareto front.webp"))
-    os.remove("Pareto front.webp")
+    assert(os.path.exists("Pareto Front.webp"))
+    os.remove("Pareto Front.webp")
 
 
 def test_quantity_constraints_objectives():
@@ -63,33 +63,33 @@ def test_quantity_constraints_objectives():
 
     # * 2 objective scatter with constraint
     scatter(moop1, output='html')
-    assert(os.path.exists("Pareto front.html"))
-    os.remove("Pareto front.html")
+    assert(os.path.exists("Pareto Front.html"))
+    os.remove("Pareto Front.html")
 
     # * 2 objective parallel with constraint
     parallel_coordinates(moop1, output='html')
-    assert(os.path.exists("Pareto front.html"))
-    os.remove("Pareto front.html")
+    assert(os.path.exists("Pareto Front.html"))
+    os.remove("Pareto Front.html")
 
     # * 2 objective radar with constraint
     radar(moop1, output='html')
-    assert(os.path.exists("Pareto front.html"))
-    os.remove("Pareto front.html")
+    assert(os.path.exists("Pareto Front.html"))
+    os.remove("Pareto Front.html")
 
     # * 5 objective scatter without constraint
     scatter(moop2, output='html')
-    assert(os.path.exists("Pareto front.html"))
-    os.remove("Pareto front.html")
+    assert(os.path.exists("Pareto Front.html"))
+    os.remove("Pareto Front.html")
 
     # * 5 objective parallel without constraint
     parallel_coordinates(moop2, output='html')
-    assert(os.path.exists("Pareto front.html"))
-    os.remove("Pareto front.html")
+    assert(os.path.exists("Pareto Front.html"))
+    os.remove("Pareto Front.html")
 
     # * 5 objective radar without constraint
     radar(moop2, output='html')
-    assert(os.path.exists("Pareto front.html"))
-    os.remove("Pareto front.html")
+    assert(os.path.exists("Pareto Front.html"))
+    os.remove("Pareto Front.html")
 
 
 def test_database_options():
@@ -116,24 +116,24 @@ def test_database_options():
         db='pf',
         points='constraint_satisfying',
         output='html')
[…]
