### 2022-09-09 70765eb
updated docs to reflect infesible starts in src code docstrings
