+++ parmoo/acquisitions/weighted_sum.py
@@ -76,8 +76,10 @@ class UniformWeights(AcquisitionFunction):
             history (dict): Another unused argument for this function.
 
         Returns:
-            numpy.ndarray: A 1d array containing a feasible starting point
-            for the scalarized problem.
+            numpy.ndarray: A 1d array containing the 'best' feasible starting
+            point for the scalarized problem (if any previous evaluations
+            were feasible) or the point in the existing database that is
+            most nearly feasible.
 
         """
 
@@ -271,8 +273,10 @@ class FixedWeights(AcquisitionFunction):
             history (dict): Another unused argument for this function.
 
         Returns:
-            numpy.ndarray: A 1d array containing a feasible starting point
-            for the scalarized problem.
+            numpy.ndarray: A 1d array containing the 'best' feasible starting
+            point for the scalarized problem (if any previous evaluations
+            were feasible) or the point in the existing database that is
+            most nearly feasible.
 
         """
 
