### 2022-09-09 f002968
updated docs requirements
