### 2022-09-12 a54a6ef
updated function names and docstrings, replacing incorrect usage of 'Lagrangian' with more generic references to a penalized objective score
