+++ docs/papers/joss/paper.bib
@@ -25,13 +25,16 @@
 @article{vtmop,
   author   = {Chang, T. H. and Watson, L. T. and Larson, J. and Neveu, N. and
               Thacker, W. I. and Deshpande, S. and Lux, T. C. H.},
-  title    = {{Algorithm XXXX}: {VTMOP}: Solver for Blackbox Multiobjective
+  title    = {{Algorithm 1028}: {VTMOP}: Solver for Blackbox Multiobjective
               Optimization Problems},
   year     = {2022},
   journal  = {{ACM} Transactions on Mathematical Software},
-  volume   = {Just accepted},
+  volume   = {48},
+  number   = {3},
+  articleno = {36},
+  numpages = {34},
   doi      = {10.1145/3529258},
-  software = {https://github.com/Libensemble/libe-community-examples/tree/main/vtmop}
+  software = {https://github.com/vtopt/vtmop}
 }
 
 @article{pymoso,
