### 2022-09-15 cb5100e
updated ABC classes to use @abstractmethod decorator, still need to update test cases to account for inability to initialize these classes without defining abstract methods
