+++ parmoo/structs.py
@@ -13,7 +13,7 @@ The classes include:
 
 """
 
-from abc import ABC
+from abc import ABC, abstractmethod
 
 
 class AcquisitionFunction(ABC):
@@ -28,6 +28,7 @@ class AcquisitionFunction(ABC):
 
     """
 
+    @abstractmethod
     def __init__(self, o, lb, ub, hyperparams):
         """ Constructor for the AcquisitionFunction class.
 
@@ -50,6 +51,7 @@ class AcquisitionFunction(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def setTarget(self, data, constraint_func, history):
         """ Set a new target value or region for the AcquisitionFunction.
 
@@ -84,6 +86,7 @@ class AcquisitionFunction(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def scalarize(self, f_vals):
         """ Scalarize a vector-valued function using the AcquisitionFunction.
 
@@ -157,6 +160,7 @@ class GlobalSearch(ABC):
 
     """
 
+    @abstractmethod
     def __init__(self, o, lb, ub, hyperparams):
         """ Constructor for the GlobalSearch class.
 
@@ -180,6 +184,7 @@ class GlobalSearch(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def startSearch(self, lb, ub):
         """ Begin a new global search.
 
@@ -255,6 +260,7 @@ class SurrogateFunction(ABC):
 
     """
 
+    @abstractmethod
     def __init__(self, m, lb, ub, hyperparams):
         """ Constructor for the SurrogateFunction class.
 
@@ -283,6 +289,7 @@ class SurrogateFunction(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def fit(self, x, f):
         """ Fit a new surrogate to the given data.
 
@@ -297,6 +304,7 @@ class SurrogateFunction(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def update(self, x, f):
         """ Update an existing surrogate model using new data.
 
@@ -312,6 +320,7 @@ class SurrogateFunction(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def setCenter(self, center):
         """ Set the center for the fit, if this is a local method.
 
@@ -323,6 +332,7 @@ class SurrogateFunction(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def evaluate(self, x):
         """ Evaluate the surrogate at a design point.
 
@@ -353,6 +363,7 @@ class SurrogateFunction(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def improve(self, x, global_imrpov):
         """ Suggests a design to evaluate to improve the surrogate near x.
 
@@ -417,6 +428,7 @@ class SurrogateOptimizer(ABC):
 
     """
 
+    @abstractmethod
     def __init__(self, o, lb, ub, hyperparams):
         """ Constructor for the SurrogateOptimizer class.
 
@@ -439,6 +451,7 @@ class SurrogateOptimizer(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
     def setObjective(self, obj_func):
         """ Add a vector-valued objective function that will be solved.
 
@@ -461,6 +474,7 @@ class SurrogateOptimizer(ABC):
 
         raise NotImplementedError("This class method has not been implemented")
 
+    @abstractmethod
[…]
