### 2022-09-15 4aa0a78
removed exception raising from inside abstract methods, and changed test_structs to check for TypeErrors
