### 2022-09-16 92dc92a
updated usage of Value/Type errors, and the definition of the xerror subroutine
