+++ parmoo/moop.py
@@ -580,13 +580,13 @@ class MOOP:
             if not isinstance(arg, dict):
                 raise TypeError("Each argument must be a Python dict")
             # Check for design variable type
-            if 'des_type' in arg.keys():
+            if 'des_type' in arg:
                 if not isinstance(arg['des_type'], str):
                     raise TypeError("args['des_type'] must be a str")
             # Append a new continuous design variable (default) to the list
-            if 'des_type' not in arg.keys() or \
+            if 'des_type' not in arg or \
                arg['des_type'] in ["continuous", "cont", "real"]:
-                if 'des_tol' in arg.keys():
+                if 'des_tol' in arg:
                     if isinstance(arg['des_tol'], float):
                         if arg['des_tol'] > 0.0:
                             des_tol = arg['des_tol']
@@ -597,7 +597,7 @@ class MOOP:
                         raise TypeError("args['des_tol'] must be a float")
                 else:
                     des_tol = 1.0e-8
-                if 'lb' in arg.keys() and 'ub' in arg.keys():
+                if 'lb' in arg and 'ub' in arg:
                     if not (isinstance(arg['lb'], float) and
                             isinstance(arg['ub'], float)):
                         raise TypeError("args['lb'] and args['ub'] must "
@@ -610,7 +610,7 @@ class MOOP:
                     raise AttributeError("'lb' and 'ub' keys must be "
                                          + "present when 'des_type' is "
                                          + "'continuous'")
-                if 'name' in arg.keys():
+                if 'name' in arg:
                     if isinstance(arg['name'], str):
                         if any([arg['name'] == dname[0]
                                 for dname in self.des_names]):
@@ -636,7 +636,7 @@ class MOOP:
                 self.cont_ub.append(arg['ub'])
             # Append a new categorical design variable to the list
             elif arg['des_type'] in ["categorical", "cat"]:
-                if 'levels' in arg.keys():
+                if 'levels' in arg:
                     if isinstance(arg['levels'], int):
                         if arg['levels'] < 2:
                             raise ValueError("args['levels'] must be at "
@@ -655,7 +655,7 @@ class MOOP:
                 else:
                     raise AttributeError("'levels' must be present when "
                                          + "'des_type' is 'categorical'")
-                if 'name' in arg.keys():
+                if 'name' in arg:
                     if not isinstance(arg['name'], str):
                         raise TypeError("When present, 'name' must be a "
                                         + "str type")
@@ -677,19 +677,19 @@ class MOOP:
                 if isinstance(arg['levels'], int):
                     self.n_lvls.append(arg['levels'])
                     self.cat_names.append([])
-                    if 'name' in arg.keys():
+                    if 'name' in arg:
                         self.des_names.append((arg['name'], 'i4'))
                 else:
                     self.n_lvls.append(len(arg['levels']))
                     self.cat_names.append(arg['levels'])
-                    if 'name' in arg.keys():
+                    if 'name' in arg:
                         self.des_names.append((arg['name'], 'U25'))
                 self.__generate_encoding__()
             # Add an integer design variable
             elif arg['des_type'] in ["integer", "int"]:
                 # Relax to a continuous design variable with des_tol = 0.5
                 des_tol = 0.5
-                if 'lb' in arg.keys() and 'ub' in arg.keys():
+                if 'lb' in arg and 'ub' in arg:
                     if not (isinstance(arg['lb'], int) and
                             isinstance(arg['ub'], int)):
                         raise TypeError("args['lb'] and args['ub'] must "
@@ -702,7 +702,7 @@ class MOOP:
                     raise AttributeError("'lb' and 'ub' keys must be "
[…]
