+++ parmoo/optimizers/lbfgsb.py
@@ -16,6 +16,7 @@ The classes include:
 import numpy as np
 import inspect
 from parmoo.structs import SurrogateOptimizer, AcquisitionFunction
+from parmoo.util import xerror
 
 
 class LBFGSB(SurrogateOptimizer):
@@ -52,8 +53,6 @@ class LBFGSB(SurrogateOptimizer):
             SurrogateOptimizer: A new SurrogateOptimizer object.
 
         """
-
-        from parmoo.util import xerror
 
         # Check inputs
         xerror(o=o, lb=lb, ub=ub, hyperparams=hyperparams)
@@ -266,8 +265,6 @@ class TR_LBFGSB(SurrogateOptimizer):
             SurrogateOptimizer: A new SurrogateOptimizer object.
 
         """
-
-        from parmoo.util import xerror
 
         # Check inputs
         xerror(o=o, lb=lb, ub=ub, hyperparams=hyperparams)
