+++ parmoo/optimizers/random_search.py
@@ -15,6 +15,7 @@ The classes include:
 import numpy as np
 import inspect
 from parmoo.structs import SurrogateOptimizer, AcquisitionFunction
+from parmoo.util import xerror
 
 
 class RandomSearch(SurrogateOptimizer):
@@ -50,8 +51,6 @@ class RandomSearch(SurrogateOptimizer):
             SurrogateOptimizer: A new SurrogateOptimizer object.
 
         """
-
-        from parmoo.util import xerror
 
         # Check inputs
         xerror(o=o, lb=lb, ub=ub, hyperparams=hyperparams)
