### 2022-09-20 e4eeaf6
updated docs to explain results of quickstart example
