### 2022-09-20 1262f29
updated paper and bib with a description of simulation-based structures in 2 example problems, including a reference to POUNDERS for those interested in learning more
