+++ docs/papers/joss/paper.bib
@@ -153,3 +153,17 @@
   pages={73-87},
   doi={10.1109/MCI.2017.2742868}
 }
+
+@incollection{pounders,
+  author    = {Stefan M. Wild},
+  title     = {Solving Derivative-Free Nonlinear Least Squares Problems with {POUNDERS}},
+  year      = {2017},
+  booktitle = {Advances and Trends in Optimization with Engineering Applications},
+  publisher = {SIAM},
+  editor    = {Terlaky, Tamas and Anjos, Miguel F. and Ahmed, Shabbir},
+  pages     = {529--540},
+  aisbn     = {978-1-611974-67-6},
+  anumber   = {ANL/MCS-P5120-0414},
+  aurl      = {http://www.mcs.anl.gov/papers/P5120-0414.pdf},
+  doi       = {10.1137/1.9781611974683.ch40},
+}
