+++ docs/install.rst
@@ -60,6 +60,36 @@ To check the installation by running the full test suite, use:
 which will also install the test dependencies (pytest_, pytest-cov_, and
 flake8_).
 
+Conda Forge
+-----------
+
+For some users (in particular, this is the recommended method for Windows
+users), the preferred method for obtaining the latest release of
+ParMOO may be through the ``conda`` package manager.
+The latest release of ParMOO is available through the ``conda-forge`` channel.
+Note that ``conda`` does not support optional dependencies, so the following
+command will automatically fetch all required and optional dependencies:
+
+.. code-block:: bash
+
+   conda install --channel=conda-forge parmoo
+
+Before running the above command, it is recommended to create a new conda
+environment using to avoid conflicts.
+Do so using:
+
+.. code-block:: bash
+
+   conda create --name channel-name
+   conda activate channel-name
+
+After performing a ``conda-forge`` installation of ParMOO, you can run
+our unit tests to make sure your installation is working:
+
+.. code-block:: bash
+
+   py.test --pyargs parmoo.tests.unit_tests
+
 Install from GitHub source
 --------------------------
 
