+++ docs/install.rst
@@ -39,7 +39,7 @@ To install the latest release:
 
     pip install < --user > parmoo
 
-where the braces around ``< --user >`` indicates that the ``--user`` flag is
+where the braces around ``< --user >`` indicate that the ``--user`` flag is
 optional.
 
 Note that the default install will not install the extra dependencies,
