### 2022-09-21 81bd5b8
Update install.rst
minor grammar issue
