### 2022-09-21 7bf0752
Update paper.bib
Also trying to fix JOSS' lack of casing for booktitles
