+++ docs/papers/joss/paper.bib
@@ -32,7 +32,7 @@
   volume   = {48},
   number   = {3},
   articleno = {36},
-  numpages = {34},
+  pages = {36:1–-34},
   doi      = {10.1145/3529258},
   software = {https://github.com/vtopt/vtmop}
 }
@@ -124,7 +124,7 @@
  author = {Balandat, Maximilian and Karrer, Brian and Jiang, Daniel and Daulton, Samuel and Letham, Ben and Wilson, Andrew G and Bakshy, Eytan},
  title = {{BoTorch}: A Framework for Efficient {M}onte-{C}arlo {B}ayesian Optimization},
  year = {2020},
- booktitle = {Advances in Neural Information Processing Systems},
+ booktitle = {{Advances in Neural Information Processing Systems}},
  volume = {33},
  editor = {H. Larochelle and M. Ranzato and R. Hadsell and M.F. Balcan and H. Lin},
  pages = {21524--21538},
@@ -133,7 +133,7 @@
 }
 
 @article{deap,
-    author    = {Fortin, F\'elix-Antoine and {De Rainville}, Fran\c{c}ois-Michel and Gardner, Marc-Andr\'e and Parizeau, Marc and Gagn\'e ", Christian},
+    author    = {Fortin, F\'elix-Antoine and {De Rainville}, Fran\c{c}ois-Michel and Gardner, Marc-Andr\'e and Parizeau, Marc and Gagn\'e, Christian},
     title     = {{DEAP}: Evolutionary Algorithms Made Easy},
     year      = {2012},
     journal   = {Journal of Machine Learning Research},
@@ -158,7 +158,7 @@
   author    = {Stefan M. Wild},
   title     = {Solving Derivative-Free Nonlinear Least Squares Problems with {POUNDERS}},
   year      = {2017},
-  booktitle = {Advances and Trends in Optimization with Engineering Applications},
+  booktitle = {{Advances and Trends in Optimization with Engineering Applications}},
   publisher = {SIAM},
   editor    = {Terlaky, Tamas and Anjos, Miguel F. and Ahmed, Shabbir},
   pages     = {529--540},
