### 2022-09-22 5be9019
added conda-installable alongside pip-installable
