### 2022-09-22 #4468 @thchang
@Viech here is our response to your review.  Your review was very helpful to us, thank-you for your thoroughness.

FYI, all changes to our source code and docs that you suggested are now active in the develop branch, and were addressed in the following PR, so you can conveniently view the diffs here:  https://github.com/parmoo/parmoo/pull/39

Additionally, we have not yet updated the paper, but we have added a joss-rev branch, where the changes to the paper have been implemented.  You can view the latest paper build here:  https://github.com/parmoo/parmoo/actions/runs/3107253388 or the latest paper source here:  https://github.com/parmoo/parmoo/tree/joss-rev/docs/papers/joss

Responses to your individual comments are given below:

## General Comments
 - Nitpick: Your copyright line reads Copyright (c) 2022, parmoo. I'm not sure if mentioning the program name here is appropriate. Maybe Copyright (c) 2022, parmoo developers, or using names or institutions, would be better?
    - We changed the copyright to “parmoo developers”

## Functionality/Installation

 - Nitpick: Writing pip install [--user] parmoo[extras] might confuse inexperienced users as the [--user] is optional (and the brackets must be removed) while for [extras] the brackets are part of the input. Also, for some shells (e.g. zsh) one would need to write "parmoo[extras]" instead of parmoo[extras] (the former should work for all shells).
    - We are now more explicit in terms of optional arguments and have added quotes. These changed the README and other doc files.
 - Optional suggestion: To support Windows and also Mac users, I've found Anaconda to be a useful distribution system in addition to PyPI/pip.
    - We have now added a parmoo recipe to conda-forge channel and updated our docs with the conda install instructions.

## Documentation

 - In the Quickstart section, the concepts of design variables, simulations, objectives and constraints are well explained but the surrogate model, its scalarazation, and the acquisition function remain somewhat enigmatic. In particular, I did not initially understand why "we must add one or more acquisition functions". I suggest to either explain these concepts already in Quickstart or to refer the reader to explanations within other sections where appropriate.
    - We have now added a short summary of the framework, linked to the detailed documentation within the README/quickstart.
 - The acquisition function is missing from the Glossary subsection of Learn about MOOPs. Non-specialist intuition could also be given in the Adding Acquisitions subsection of Write a ParMOO Script.
    - Thanks for catching this, we have now added this.
 - In the Minimum Working Example, it would be helpful if the output could be explained and/or plotted.
    - We now include a brief description of the results variable, and provide code for plotting a scatterplot of the results, showing the nondominated points.  This has been updated in the README, quickstart section of the docs, and the basic tutorials.
 - JOSS requirement: You should mention the target audience in the documentation.
    - We have added a paragraph about our intended audience and usage to the landing page of the readthedocs site.
 - JOSS requirement: You should mention how to report issues and seek support.
    - In the resources in documentation we now explicitly say that one can report issues or seek support via github or the provided email.

## Software paper

 - In the summary, which I otherwise find very accessible, it is not clear to me what is meant by a "simulation-based structure". Briefly explaining this concept could make also the examples more accessible to non-specialists (here I do not understand the parts that explain where said structure comes from).
    - We have reworded how this property is introduced in the summary.  Additionally, we have added a methodology section, which both summarizes how ParMOO works and gives further details on how the simulation structure is exploited.  Finally, by clarifying the EDF example (below), we hope that we have further improved clarity through the example.
 - "Scientific software packages for solving multiobjective simulation optimization problems include": Here I would prefer to read the names of the software packages (with authors and years in parentheses). It would also be nice to have a partition into Python packages and other software (unless all of them are for Python; then this should be mentioned).
    - We have added several additional packages to the list that were omitted in the initial submission, added package names, and clarified which packages were written in Python.
 - In the EDF example, I do not quite understand what is meant by ParMOO being used to "minimiz[e] the error between expensive simulation outputs and experimental data".
    - We have updated the description of the EDF calibration and chemical synthesis problems to explicitly specify how the problem is given to ParMOO, and what simulation-based structure is exploitable in each of these contexts.  In the case of the EDF calibration, we also added a reference to the single-objective software POUNDERS, which attempts to exploit a similar structure in single-objective problems.
 - If available, references to the mentioned research using ParMOO would be very interesting.
    - N/A – Unfortunately, we do not have any references for these applications at this time.
 - Nitpick: The HPC abbreviation is not introduced.
    - We have now defined the abbreviation before its first usage.
 - Nitpick: For your ACM Transactions on Mathematical Software paper, in the references, maybe add a note/comment that this paper was just accepted. Otherwise, it looks like volume and issue numbers are missing.
    - The TOMS paper has appeared in print, so we have updated the reference accordingly.
 - Nitpick/unsure: The string "In arXiv preprint arXiv:2105.03000." in the references looks odd (but I do not know the prefered way to cite an arXiv paper in the JOSS).
    - This preprint has now been moved to the “Just Online” section of Numerical Algorithms, so we have changed the reference anyway.
 - JOSS requirement: The statement of need section should mention the target audience.
    - We have added a statement on the target audience (same as in the docs) to the bottom of the statement of need section.
 - (In my experience, the editor will later ask you to investigate the missing DOIs that were mentioned by[ @editorialbot](https://github.com/editorialbot).)
    - We have now added either DOIs or URLs (when no DOI is available) to all refs.

## Source Code:

 - You do not need to have a return statement at the end of a function. (Python automatically does return None at the end of a non-generator function.)
    - Your comment is well taken but we prefer to explicitly include the return statements for clarity reasons.  In our experience, this helps with maintainability, especially with complicated functions containing multiple return statements, by explicitly marking the possible return locations
 - When using the ABC base class, you should use the abc.abstractmethod decorator instead of raising NotImplementedError(e.g. in sim_func add @abstractmethod above __call__, then inside __call__, just pass). Otherwise I do not think that adding the ABC base class has any effect.
    - Thanks for the tip, we weren’t sure if we were using the ABC library right, so that makes sense now.  There are a few methods in each class that either (1) have a default implementation or (2) need not be implemented by the child class, depending on the solver techniques being used.  To allow for this, we left the @abstractmethod tag off of these methods, and left the raise NotImplementedError in the body of these optional methods.  This way, if a user tries to use a method that does not implement one of these optional methods, with another solver that requires the implementation of these optionals methods, then the mis-match will be detectable.  For the majority of methods in each class, however, we have added the decorator, as you suggested and updated our test suite accordingly.
 - You could make your xerror function future-proof by giving the arguments default values and callling it with keyword arguments, like you do with __check_optionals__ (i.e. instead of xerror(1, lb, ub, {}) you would write xerror(lb=lb, ub=ub)). This would allow you to add new checks later on.
    - Good tip!  We have done as suggested.
 - Is there a reason you do from parmoo.util import xerror within __init__? It does not seem to produce a circular import if you import it at the start of the source file, which is the typical reason to import within a function. Importing at the start of the file might profit software that analyzes your code, including your IDE.
    - Fixed.  As you know, there are several reasons to avoid unnecessary imports, so we had wanted to do so by placing imports that are not used throughout the containing class inside the individual methods that actually use them.  This was probably an oversight, however, as in the case of __init__, everything inside the __init__ function will alway be imported if the class is to be used, and as you pointed out, there is also no risk of circular import.
 - if 'name' in arg.keys() -> if 'name' in arg
    - Thanks for the tip!  We have replaced all occurrences as suggested.
 - You have some weird try: assert(_) except BaseException: raise _ checks in sum_sim_bound. Can you not just condition the raised exception on what you check within assert? Normally, assert is reserved for things that, when they go wrong, imply an error in the software that makes the assertion (as opposed to a user error).
    - Fixed.  We were doing it this way for a while to reduce the number of lines of code and avoid having a double-check (first, does the initialization succeed, then is the result of the correct type).  We thought we had corrected all occurrences, but missed a couple.
 - When checking whether something is a NumPy array, the proper exception would by TypeError instead of ValueError. Although, I cannot think of a scenario where this would make any difference, so it doesn't get any more nitpicky than this!
    - Good catch.  We made this mistake early-on in the development process, and tried to go back and fix it later, but it looks like we missed many occurrences of this mistake.  We went back and think that we have corrected all instances now (including several instances where we were using a ValueError when checking the Python intrinsic types, and when checking whether an input is callable).
 - In setup.py, exec(open("parmoo/version.py").read()) looks a bit dangerous. Have you considered from parmoo.version import __version__?
    - setup.py is used to create the package.  Importing from the package while creating the package seems dangerous, and other projects do it this way.  Open to feedback if you have further insight.
 - Thank you for making me aware of __slots__, I've literally never seen this before!
    - Great!  As you’ve probably read, it allegedly gives significant speedup in certain situations.

## Issues (not impeding acceptance; only for reference)

 - [Misleading exception when running a libE_MOOP script without mpirun/arguments parmoo/parmoo#32](https://github.com/parmoo/parmoo/issues/32)
    - This is a correct description of one of the reasons why this error could occur, but indeed, there are multiple reasons that could cause the error.  We have added additional language to the error message, describing these other causes.
