### 2022-09-25 #4468 @Viech
I've re-read the paper, which to me is much clearer now! I've also reviewed parmoo/parmoo#39 and this looks good to me. From my point of view, the submission can be accepted as is.

> > - In setup.py, exec(open("parmoo/version.py").read()) looks a bit dangerous. Have you considered from parmoo.version import version?
>
>  - setup.py is used to create the package. Importing from the package while creating the package seems dangerous, and other projects do it this way. Open to feedback if you have further insight.

Unfortunately I do not have any more insight and I suppose `exec` is not so different from `import`, anyway.

>  In our experience, this helps with maintainability, especially with complicated functions containing multiple return statements, by explicitly marking the possible return locations

Just as a side remark, if this kind of explicit/self-documenting code is important to you, I should advertise Python's new type hinting features! Type hints allow you to document function argument and return types and to have those statically checked. You can write, for instance, `def divide(n: int, d: int) -> float:` and your IDE/CI can tell you when this function is not called with integer arguments or when it does not always return a `float`. However some care must be taken with respect to which Python versions you support as these features were incrementally added since version `3.5`.
