### 2022-10-03 #4468 @thchang
Thanks @Viech

We are aware of the hinting features, and it is definitely something that we want to look into in the future, but I think it is not so easy in our case since we are simulating polymorphism by defining some function behaviors based on the input types (thereby allowing for multiple different input types).

Not sure what the recommendation/policy is in these cases:  e.g., suppose ``arg1`` could be an ``int``, ``list``, or ``dict`` type, and the output will be a ``float``, ``str``, or ``None`` depending on what the input type is.

In any case, we'll look into it in the future, at least for the simpler function definitions.  Thanks for all your feedback so far!
