### 2022-10-04 #4468 @Viech
> Not sure what the recommendation/policy is in these cases: e.g., suppose arg1 could be an int, list, or dict type, and the output will be a float, str, or None depending on what the input type is.

A simple (but not very strict) approach is via `Union`:
```python
from typing import Union

def f(arg1: Union[int, list, dict]) -> Union[float, str, None]:
    ...
```

If you want to model explicitly how return types depend on input types, you can use the [typing.overload](https://docs.python.org/3/library/typing.html#typing.overload) decorator for this.
