### 2022-10-04 #4468 @thchang
Oh, thanks for this tip.  I was not aware of ``Union``. That is probably a good approach to take, although I still need to check the Python version support before implementing something like this.
