### 2022-12-06 #4468 @kellyrowland
Hi @N-Wouda @max-little 👋 would either of you be interested in and available for reviewing this JOSS submission?
