### 2022-12-06 #4468 @N-Wouda
> Hi @N-Wouda @max-little 👋 would either of you be interested in and available for reviewing this JOSS submission?

Hi @kellyrowland, topic-wise this would be a good fit, but I'm swamped with work until at least early February. I suggest you look for someone else that can review sooner. Feel free to ping again around mid-January if you're still in need of a reviewer - I might be able to help out then. Good luck!
