### 2022-12-08 #4468 @max-little
Thanks Kelly. Can I recommend a potential reviewer?
Best
Max

On Tue, 6 Dec 2022 at 18:50, Kelly L. Rowland ***@***.***>
wrote:

> Hi @N-Wouda <https://github.com/N-Wouda> @max-little
> <https://github.com/max-little> 👋 would either of you be interested in
> and available for reviewing this JOSS submission?
>
