### 2022-12-12 #4468 @danielskatz
As the track editor, I want to note here that one of the authors (Stefan Wild) has, since this review started, moved to the same institution as the editor. The JOSS editors have discussed this, and feel that, given the state of the review and the fact that the author and editor are in different units, we will let this review proceed, and are just noting this conflict that has appeared as waived.
