### 2022-12-14 #4468 @kellyrowland
Hi @JianqiaoMao 👋 thanks for stepping in here to review this submission.

You'll want to post a comment with `@\editorialbot generate my checklist` (without the slash, I don't actually want to ping editorialbot in this example but you'll want to do so).

That will create a comment here with a list of checkbox items for you to check off as your review. You can see what the checklist looks like further up in this issue; the review involves reading through the paper submission and verifying code functionality.

If you have any questions or comments for the authors, you can open an issue on the code repo or as a comment here. You're also welcome to ask any other questions you have here in this issue too.

Once you've completed the checklist and feel that the paper is ready, please let me know.
