### 2023-01-01 3ae1402
fixed typo in setting tr boundaries
