### 2023-01-09 c401120
added print debug stmts
