### 2023-01-09 77723a8
reverted to testing python 3.6, since error caused by python actions version, not python version
