+++ .github/workflows/parmoo-ci.yml
@@ -13,7 +13,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: ["3.8", "3.11"]
+        python-version: ["3.6", "3.11"]
 
     defaults:
       run:
