### 2023-01-09 aacb360
upgrading to action v3, dropping support for python 3.6 since github has also dropped support actions/setup-python/#544
