### 2023-01-24 #4468 @JianqiaoMao
Hi Kelly,

Sorry for the late reply. Will finish the revision by tomorrow. My sincere
apology!

Best,

Jianqiao

Kelly L. Rowland ***@***.***> 於 2023年1月23日 週一 下午7:40寫道：

> Hi @JianqiaoMao <https://github.com/JianqiaoMao> - can you please let me
> know as soon as possible whether or not you're able to take on this review?
> We'd like to move forward and find another reviewer if needed.
>
