### 2023-01-24 #4468 @JianqiaoMao
I am sorry for the late comment for this submission.

I have gone through the paper and main parts of the software documents, but may need to finish some more testing by tomorrow (24th of Jan.).

Looks almost perfect, but I just have an optional suggestion for the software paper. When listing other packages, it would be clear if package names are explictly stated followed by _(author name, year of publication)_, which can also make other readers easier to compare them with ParMOO and also guarantee the consistency of citation style. (Considering there are different citation style in line 40 and 22-23.) Otherwise, your statement "the above-listed software packages" may be a little confusing, because these authors' name are not, infact, the package names.
