### 2023-01-26 #4468 @Viech
@kellyrowland can you re-generate the PDF? The style suggestion was something I also remarked and I recall that the authors already changed this but the updated proof was only referenced here via a [now-expired artifact](https://github.com/parmoo/parmoo/actions/runs/3107253388).
