### 2023-01-26 9bb4816
updated Stefan's affilitiation in JOSS paper
