+++ docs/papers/joss/paper.md
@@ -12,13 +12,15 @@ authors:
     affiliation: "1" 
   - name: Stefan M. Wild
     orcid: 0000-0002-6099-2772
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)    
+    affiliation: "2, 1, 3" # (Multiple affiliations must be quoted)    
 affiliations:
  - name: Mathematics and Computer Science Division, Argonne National Laboratory, USA
    index: 1
- - name: NAISE, Northwestern University, USA 
+ - name: Applied Mathematics and Computational Research Division, Lawrence Berkeley National Laboratory, USA
    index: 2
-date: 12 May 2022
+ - name: NAISE, Northwestern University, USA 
+   index: 3
+date: 26 Jan 2023
 bibliography: paper.bib
 
 ---
