### 2023-01-26 #4468 @thchang
Hi @JianqiaoMao and @kellyrowland ,

Thanks @Viech for pointing out that the artifact had expired.  I took this opportunity to update @wildsm 's affiliation since he has changed institutions, and rebuild the paper revision.

  https://github.com/parmoo/parmoo/actions/runs/4018534923

As pointed out, I think that in our last round, we had already fixed the paper references to refer to author names instead.  As for the in-docstring references, I agree with @JianqiaoMao that they should be in a unified style, and I'll work through and make sure that they are.

This actually raises a good point though; it would probably be nice if the references in our documentation could be managed automatically, with an automatically generated bibliography.  This is challenging though, since I am referencing these papers in the docstring for the function that implements an idea from one of them.

Does anyone know an automatic bibliography generator, reference manager, or something similar, that I could use here?  To re-iterate, it has to be able to handle references in the Python docstrings.

If not, I'll move ahead with just updating the docstrings, but it would be nice to know if such a tool exists.

Thanks again all for your reviews and suggestions!
