### 2023-01-26 #4468 @thchang
@kellyrowland I'm not sure what the generate pdf request does, but we did not update the original paper since at the time, we didn't know if the other reviewers was actively reading it.  Instead, we created a new branch and the paper revisions are located there

checking the pdf above, it looks like that did not generate the latest version.  Should we merge our updated paper into that branch so that it generates the correct version?
