### 2023-01-26 #4468 @kellyrowland
Whew, there we go. @thchang @Viech @JianqiaoMao the most up-to-date PDF should be in the post linked above, thanks for the various pings on that.

I'm not sure about the automatic reference generation/management, I'll have to do some looking around.
