### 2023-01-27 0fc5301
fixed a bug in how the gaussian processes handle the model improvement step
