### 2023-01-27 #4468 @thchang
OK, it looks like I misunderstood that @JianqiaoMao was concerned also about the references in the docstrings, when in fact, he was only concerned about the references in the paper.

I have re-checked the docstrings and don't see any inconsistencies in those citation styles.

Please let me know if anyone still has outstanding concerns, given the latest PDF generation.

Otherwise, I will assume that all the reviewers' concerns have been addressed at this time since the package names appear to be stated, as requested, in the latest paper build.
