### 2023-01-30 639aa8f
updated default des tol to scale with ub - lb and model improvement regions to scale inversely with sample std dev
