+++ parmoo/moop.py
@@ -550,7 +550,7 @@ class MOOP:
                    specifies the tolerance, i.e., the minimum spacing along
                    this dimension, before two design values are considered to
                    have equal values in this dimension. If not specified, the
-                   default value is 1.0e-8.
+                   default value is 1.0e-8 * (ub - lb).
                  * 'levels' (int or list): When des_type is 'categorical', this
                    specifies the number of levels for the variable (when int)
                    or the names of each valid category (when a list).
@@ -598,7 +598,12 @@ class MOOP:
                     else:
                         raise TypeError("args['des_tol'] must be a float")
                 else:
-                    des_tol = 1.0e-8
+                    if 'lb' in arg and 'ub' in arg and \
+                       isinstance(arg['lb'], float) and \
+                       isinstance(arg['ub'], float):
+                        des_tol = 1.0e-8 * max(arg['ub'] - arg['lb'], 1.0e-4)
+                    else:
+                        des_tol = 1.0e-8
                 if 'lb' in arg and 'ub' in arg:
                     if not (isinstance(arg['lb'], float) and
                             isinstance(arg['ub'], float)):
