+++ CHANGELOG.rst
@@ -39,3 +39,46 @@ Desired features:
  - a visualization dashboard for viewing progress interactively
  - design variable types should be a class, with embed/extract methods
    that can be called by ``MOOP.__embed__()`` and ``MOOP.__extract__()``
+
+Release 0.2.0
+-------------
+
+:Date: Feb 2, 2023
+
+Official release corresponding to accepted JOSS article.
+
+ - Added support for a wider variety of design variables (including integer
+   types), as well as support for "custom" design variables that use
+   user-provided custom embedders/extractors.
+   Documentation on design variables has been expanded accordingly.
+   Although design variables are still specified through dicts not classes,
+   this addresses and therefore closes the primary issue raised in #28
+ - Updated ``extras/libe.py`` corresponding to interface changes made in
+   libEnsemble Release 0.8.0. This also addresses the issues on MacOS,
+   referenced in #34
+ - Added a post-run visualization library and corresponding
+   documentation, closing issue #27
+ - Allow solvers to start from an initial point that is infeasible, so that
+   problems with relaxable constraints and a very small feasible set can
+   still be solved
+ - Various style changes and additional usage environments requested by
+   JOSS reviewers openjournals/joss-reviews#4468 including #32
+ - Added support for multistarting optimization solvers when solving
+   surrogate problems. This is particularly important for the global
+   ``GaussRBF`` surrogate
+ - Fixed an issue in how model improvement points are calculated, as
+   implemented in the ``surrogate.improve`` method for each GaussRBF variation
+   in ``surrogates/gaussian_proc.py``, which was created when adding support
+   for custom design variables
+ - The default design tolerance for continuous variables now depends upon
+   the value of ``ub - lb``
+
+Note: 
+
+ - Dropped support for Python 3.6, due to changes to GitHub Actions documented
+   on actions/setup-python#544
+
+Known issues:
+
+ - The visualization library uses advanced plotly/dash features, which may
+   not support the chrome browser, as described in #37
