+++ README.rst
@@ -30,7 +30,6 @@ exploit simulation-based structure in objective and constraint functions.
 
 To exploit structure, ParMOO models *simulations* separately from
 *objectives* and *constraints*. In our language:
-
  * a **design variable** is an input to the problem, which we can directly
    control;
  * a **simulation** is an expensive or time-consuming process, including
@@ -61,7 +60,6 @@ Dependencies
 ParMOO has been tested on Unix/Linux and MacOS systems.
 
 ParMOO's base has the following dependencies:
-
  * Python_ 3.6+
  * numpy_ -- for data structures and performant numerical linear algebra
  * scipy_ -- for scientific calculations needed for specific modules
@@ -70,11 +68,9 @@ ParMOO's base has the following dependencies:
 
 Additional dependencies are needed to use the additional features in
 ``parmoo.extras``:
-
  * libEnsemble_ -- for managing parallel simulation evaluations
 
 And for using the Pareto front visualization library in ``parmoo.viz``:
-
  * plotly_ -- for generating interactive plots
  * dash_ -- for hosting interactive plots in your browser
  * kaleido_ -- for exporting static plots post-interaction
@@ -252,7 +248,6 @@ Congratulations, you now know enough to get started solving MOOPs with
 ParMOO!
 
 Next steps:
-
  * Learn more about all that ParMOO has to offer (including saving and
    checkpointing, INFO-level logging, advanced problem definitions, and
    different surrogate and solver options) at our ReadTheDocs_ page.
@@ -266,18 +261,30 @@ Resources
 ---------
 
 To seek support or report issues, e-mail:
-
  * ``parmoo@mcs.anl.gov``
 
 Our full documentation is hosted on:
-
  * ReadTheDocs_
 
 Please read our LICENSE_ and CONTRIBUTING_ files.
 
-Citing ParMOO:
+Citing ParMOO
+-------------
 
- * Please use the following to cite ParMOO:
+Please use one of the following to cite ParMOO.
+
+Our JOSS paper:
+
+.. code-block:: bibtex
+
+    @article{parmoo-joss,
+        title       = {{ParMOO}: A {P}ython library for parallel multiobjective simulation optimization},
+        author      = {Chang, Tyler H. and Wild, Stefan M.},
+        year        = {2023},
+        journal     = {To appear in The Journal of Open Source Software}
+    }
+
+Our online documentation:
 
 .. code-block:: bibtex
 
@@ -285,11 +292,10 @@ Citing ParMOO:
         title       = {{ParMOO}: {P}ython library for parallel multiobjective simulation optimization},
         author      = {Chang, Tyler H. and Wild, Stefan M. and Dickinson, Hyrum},
         institution = {Argonne National Laboratory},
-        number      = {Version 0.1.0},
-        year        = {2022},
+        number      = {Version 0.2.0},
+        year        = {2023},
         url         = {https://parmoo.readthedocs.io/en/latest}
     }
-
 
 .. _Actions: https://github.com/parmoo/parmoo/actions
 .. _CONTRIBUTING: https://github.com/parmoo/parmoo/blob/main/CONTRIBUTING.rst
