+++ parmoo/__init__.py
@@ -8,6 +8,6 @@ simulation-based structure in objective and constraint functions.
 
 from .version import __version__
 __author__ = "Tyler H. Chang, Stefan M. Wild, and Hyrum Dickinson"
-__credits__ = "Argonne National Laboratory"
+__credits__ = "Argonne National Laboratory and Lawrence Berkeley National Laboratory"
 
 from .moop import MOOP
