### 2023-02-02 7139550
Update CHANGELOG.rst
simpler format change.
