### 2023-02-02 #4468 @thchang
Thanks @kellyrowland  We just released v0.2.0, publishing all of the changes requested during this review and closing relevant issues.  @wildsm has gotten us a Zenodo DOI corresponding to this release:  10.5281/zenodo.7600559
