### 2023-02-02 #4468 @JianqiaoMao
> Thanks all!
>
> @JianqiaoMao could you please check off the final checkbox on your list?
>
> @thchang to prepare for publication, could you please make a new release of the software, deposit it as an archive (on Zenodo or elsewhere) and then post the archive DOI here?

Done!
