### 2023-02-03 #4468 @thchang
Is the "Check final proof" step something to be completed by me and @wildsm?  Or by the editors?
