### 2023-02-03 #4468 @danielskatz
I've suggested some minor changes in https://github.com/parmoo/parmoo/pull/45 - please merge this, or let me know what you disagree with, then we can proceed
