### 2023-02-03 #4468 @thchang
Thanks @danielskatz.  I guess I have one small issue:

Line 124:  I think in American English, this should actually just be "nontrivial"

Everything else seems good to me.  I think @wildsm is busy until early next week, so we should also wait for him to approve this.
