### 2023-02-03 #4468 @danielskatz
I think either is ok, but I'll change this in my PR
