### 2023-02-03 #4468 @thchang
One is American English, the other is British English -- Personally, I don't feel strongly about it either, but I've had collaborators in the past who did (about proper usage of the prefix non in particular)

Anyway, I am willing to approve the changes now, but I'll wait for @wildsm to get back next week.
